### 2020-10-27 #2789 @whedon
[PRE REVIEW]: PyAutoLens: Open-Source Strong Gravitational Lensing
**Submitting author:** @Jammy2211 (<a href="http://orcid.org/0000-0002-8987-7401">James Nightingale</a>)
**Repository:** <a href="https://github.com/Jammy2211/PyAutoLens" target ="_blank">https://github.com/Jammy2211/PyAutoLens</a>
**Version:** 1.7.8
