### 2020-10-27 #2789 @arfon
:wave: @Jammy2211 - is this submission associated with a paper in a AAS publication?
