+++ paper/paper.bib
@@ -140,7 +140,9 @@ year = {2015}
   year={2011},
   volume={13},
   number={2},
-  pages={22-30},}
+  pages={22-30},
+  doi={10.1109/MCSE.2011.37}}
+
 @misc{pyautofit,
   author = {​James Nightingale and Richard Hayes},
   title = {PyAutoFit: A classy Probabilistic Programming Language for Complex Model-fitting},
