### 2020-10-29 157a03d
fix DOI for SciPy
