+++ paper/paper.bib
@@ -268,7 +268,7 @@ year = {2018}
       volume={17},
       pages={261--272},
       adsurl = {https://rdcu.be/b08Wh},
-      doi = {https://doi.org/10.1038/s41592-019-0686-2},
+      doi = {10.1038/s41592-019-0686-2},
 }
 @article{Alexander2019,
 abstract = {Dark matter substructure has the potential to discriminate between broad classes of dark matter models. With this in mind, we construct novel solutions to the equations of motion governing condensate dark matter candidates, namely axion Bose-Einstein condensates and superfluids. These solutions are highly compressed along one axis and thus have a disk-like geometry. We discuss linear stability of these solutions, consider the astrophysical implications as a large-scale dark disk or as small scale substructure, and find a characteristic signal in strong gravitational lensing. This adds to the growing body of work that indicates that the morphology of dark matter substructure is a powerful probe of the nature of dark matter.},
