### 2020-10-29 #2789 @xuanxu
@Jammy2211 A quick comment before I look for reviewers: I got some errors cloning the repo, because the .git directory is big (~260Mb), so I had to install the [LFS extension](https://git-lfs.github.com/) for git to be able to work with it. I'd recommend adding a note in the _Cloning/forking_ section of the Installation guide, and/or trying to reduce the size of the repo objects in the git history if possible.
