### 2020-10-29 #2789 @Jammy2211
Thanks. I literally added the LFS stuff last night, having not used it before. Sounds like it is cumbersome so I will remove it entirely and switch the files to some sort of a URL download or separate GIt repo :).
