### 2020-10-29 #2789 @Jammy2211
We're also aware of an issue where the imports in the `transdimensional` and `slam` runners in the autolens_workspace will give import errors, which require setting your PYTHONPATH to fix.

We did a big refactor before submitting (probably not the best idea...) changing how we handle paths / configs / output and are still having a few teething issues, but the goal is to make it so users don't have to set environment variables to run every!
