### 2020-10-30 9c2f065
fixed subhalo plot issues
