+++ howtolens/chapter_3_pipelines/scripts/tutorial_2_pipeline_runner.py
@@ -1,7 +1,3 @@
-"""
-
-"""
-
 # %%
 """
 Tutorial 2: Two Lens Galaxies
@@ -43,9 +39,8 @@ we'll use new strong lensing data, where:
 """
 
 # %%
-dataset_type = "chapter_3"
 dataset_name = "light_sersic_x2__mass_sie_x2__source_exp"
-dataset_path = f"dataset/howtolens/{dataset_type}/{dataset_name}"
+dataset_path = f"dataset/howtolens/chapter_3/{dataset_name}"
 
 imaging = al.Imaging.from_fits(
     image_path=f"{dataset_path}/image.fits",
@@ -108,38 +103,6 @@ So, with this in mind, we've written a pipeline composed of 4 phases:
 
 # %%
 """
-__Pipeline_Setup_And_Tagging__:
-
-For this pipeline the pipeline setup customizes:
-
- - If there is an `ExternalShear` in the mass model or not.
-
-The pipeline setup again `tags` the output path of a pipeline.
-
-The `path_prefix` belows specify the path the pipeline results are written 
-
- `autolens_workspace/howtolens/output/c3_t2_x2_galaxies/pipeline__light_and_source`
-
-The redshift of the lens and source galaxies are also input (see `examples/model/customize/redshift.py`) for a 
-description of what inputting redshifts into **PyAutoLens** does.
-"""
-
-# %%
-setup_light = al.SetupLightParametric()
-setup_mass = al.SetupMassTotal(with_shear=True)
-setup_source = al.SetupSourceParametric()
-
-setup = al.SetupPipeline(
-    path_prefix="howtolens/c3_t2_x2_galaxies",
-    redshift_lens=0.5,
-    redshift_source=1.0,
-    setup_light=setup_light,
-    setup_mass=setup_mass,
-    setup_source=setup_source,
-)
-
-# %%
-"""
 __Pipeline Creation__
 
 To create a `Pipeline`, we call a `make_pipeline` function, which is written in its own Python script: 
@@ -156,7 +119,10 @@ from autolens_workspace.howtolens.chapter_3_pipelines import (
 )
 
 pipeline_x2_galaxies = tutorial_2_pipeline_x2_lens_galaxies.make_pipeline(
-    setup=setup, settings=settings
+    path_prefix="howtolens/c3_t2_x2_galaxies",
+    settings=settings,
+    redshift_lens=0.5,
+    redshift_source=1.0,
 )
 
 # Uncomment to run.
