### 2020-10-30 d1684a6
fixed subhalo plot issues
