+++ test_autolens/integration/tests/imaging/phase_features/positions.py
@@ -4,14 +4,13 @@ from test_autolens.integration.tests.imaging import runner
 
 test_type = "phase_features"
 test_name = "positions"
-dataset_name = "lens_sis__source_smooth"
+dataset_name = "mass_sie__source_sersic"
 instrument = "vro"
 
 
-def make_pipeline(name, path_prefix, search=af.DynestyStatic()):
+def make_pipeline(name, path_prefix):
     phase1 = al.PhaseImaging(
         name="phase[1]",
-        path_prefix=path_prefix,
         galaxies=dict(
             lens=al.GalaxyModel(redshift=0.5, mass=al.mp.SphericalIsothermal),
             source=al.GalaxyModel(redshift=1.0, light=al.lp.EllipticalSersic),
