+++ test_autolens/integration/tests/interferometer/lens__source_inversion/adaptive_magnification/lens_mass__source.py
@@ -4,11 +4,11 @@ from test_autolens.integration.tests.interferometer import runner
 
 test_type = "lens__source_inversion"
 test_name = "lens_mass__source_adaptive_magnification"
-dataset_name = "lens_sie__source_smooth"
+dataset_name = "mass_sie__source_sersic"
 instrument = "sma"
 
 
-def make_pipeline(name, path_prefix, real_space_mask, search=af.DynestyStatic()):
+def make_pipeline(name, path_prefix, real_space_mask):
 
     mass = af.PriorModel(al.mp.EllipticalIsothermal)
 
@@ -18,7 +18,6 @@ def make_pipeline(name, path_prefix, real_space_mask, search=af.DynestyStatic())
 
     phase1 = al.PhaseInterferometer(
         name="phase[1]",
-        path_prefix=path_prefix,
         galaxies=dict(
             lens=al.GalaxyModel(redshift=0.5, mass=mass),
             source=al.GalaxyModel(
@@ -39,7 +38,6 @@ def make_pipeline(name, path_prefix, real_space_mask, search=af.DynestyStatic())
 
     phase2 = al.PhaseInterferometer(
         name="phase[2]",
-        path_prefix=path_prefix,
         galaxies=dict(
             lens=al.GalaxyModel(
                 redshift=0.5, mass=phase1.result.model.galaxies.lens.mass
