+++ test_autolens/integration/tests/result_passing/lens_light/instance_via_af_last_dont_specify_light.py
@@ -3,16 +3,15 @@ import autolens as al
 from test_autolens.integration.tests.imaging import runner
 
 test_type = "reult_passing"
-test_name = "lens_light_instance_via_af_last_dont_specify_light"
-dataset_name = "lens_sie__source_smooth"
+test_name = "light_sersic___instance_via_af_last_dont_specify_light"
+dataset_name = "mass_sie__source_sersic"
 instrument = "vro"
 
 
-def make_pipeline(name, path_prefix, search=af.DynestyStatic()):
+def make_pipeline(name, path_prefix):
 
     phase1 = al.PhaseImaging(
         name="phase[1]",
-        path_prefix=path_prefix,
         galaxies=dict(
             lens=al.GalaxyModel(
                 redshift=0.5,
@@ -38,7 +37,6 @@ def make_pipeline(name, path_prefix, search=af.DynestyStatic()):
 
     phase2 = al.PhaseImaging(
         name="phase[2]",
-        path_prefix=path_prefix,
         galaxies=dict(lens=lens, source=phase1.result.model.galaxies.source),
         sub_size=1,
         search=search,
