+++ test_autolens/plot/arrays/array_and_mask.py
@@ -8,7 +8,7 @@ from test_autolens.simulators.imaging import instrument_util
 
 # Lets quickly remind ourselves of the image, and the 3.0" circular mask we'll use to mask it.
 imaging = instrument_util.load_test_imaging(
-    dataset_name="lens_light_dev_vaucouleurs", instrument="vro"
+    dataset_name="light_dev_vaucouleurs", instrument="vro"
 )
 
 array = imaging.image

[…]
