### 2020-10-31 32db112
subhalo plot fix
