+++ test_autolens/integration/tests/interferometer/lens__source_inversion/adaptive_brightness/lens_mass__source__hyper.py
@@ -4,18 +4,17 @@ from test_autolens.integration.tests.interferometer import runner
 
 test_type = "lens__source_inversion"
 test_name = "lens_mass__source_adaptive_brightness__hyper"
-dataset_name = "lens_sie__source_smooth"
+dataset_name = "mass_sie__source_sersic"
 instrument = "sma"
 
 
-def make_pipeline(name, path_prefix, real_space_mask, search=af.DynestyStatic()):
+def make_pipeline(name, path_prefix, real_space_mask):
     class Phase1(al.PhaseInterferometer):
         def customize_priors(self, results):
             self.galaxies.source.light.sersic_index = af.UniformPrior(3.9, 4.1)
 
     phase1 = Phase1(
         name="phase[1]",
-        path_prefix=path_prefix,
         galaxies=dict(
             lens=al.GalaxyModel(
                 redshift=0.5, mass=al.mp.EllipticalIsothermal, shear=al.mp.ExternalShear
@@ -37,7 +36,6 @@ def make_pipeline(name, path_prefix, real_space_mask, search=af.DynestyStatic())
     )
     phase2 = al.PhaseInterferometer(
         name="phase_2_weighted_regularization",
-        path_prefix=path_prefix,
         galaxies=dict(
             lens=al.GalaxyModel(
                 redshift=0.5,
@@ -69,7 +67,6 @@ def make_pipeline(name, path_prefix, real_space_mask, search=af.DynestyStatic())
 
     phase3 = al.PhaseInterferometer(
         name="phase[3]",
-        path_prefix=path_prefix,
         galaxies=dict(
             lens=al.GalaxyModel(
                 redshift=0.5,
