### 2020-10-31 d5eb57a
image path now handled by autofit, passed to visualize method
