+++ test_autolens/unit/plot/test_fit_imaging_plots.py
@@ -21,7 +21,7 @@ def test__fit_quantities_are_output(
         plotter=aplt.Plotter(output=aplt.Output(path=plot_path, format="png")),
     )
 
-    assert plot_path + "image.png" in plot_patch.paths
+    assert f"{plot_path}/image.png" in plot_patch.paths
 
     aplt.FitImaging.noise_map(
         fit=masked_imaging_fit_x2_plane_7x7,
@@ -29,7 +29,7 @@ def test__fit_quantities_are_output(
         plotter=aplt.Plotter(output=aplt.Output(path=plot_path, format="png")),
     )
 
-    assert plot_path + "noise_map.png" in plot_patch.paths
+    assert f"{plot_path}/noise_map.png" in plot_patch.paths
 
     aplt.FitImaging.signal_to_noise_map(
         fit=masked_imaging_fit_x2_plane_7x7,
@@ -37,7 +37,7 @@ def test__fit_quantities_are_output(
         plotter=aplt.Plotter(output=aplt.Output(path=plot_path, format="png")),
     )
 
-    assert plot_path + "signal_to_noise_map.png" in plot_patch.paths
+    assert f"{plot_path}/signal_to_noise_map.png" in plot_patch.paths
 
     aplt.FitImaging.model_image(
         fit=masked_imaging_fit_x2_plane_7x7,
@@ -45,7 +45,7 @@ def test__fit_quantities_are_output(
         plotter=aplt.Plotter(output=aplt.Output(path=plot_path, format="png")),
     )
 
-    assert plot_path + "model_image.png" in plot_patch.paths
+    assert f"{plot_path}/model_image.png" in plot_patch.paths
 
     aplt.FitImaging.residual_map(
         fit=masked_imaging_fit_x2_plane_7x7,
@@ -53,7 +53,7 @@ def test__fit_quantities_are_output(
         plotter=aplt.Plotter(output=aplt.Output(path=plot_path, format="png")),
     )
 
-    assert plot_path + "residual_map.png" in plot_patch.paths
+    assert f"{plot_path}/residual_map.png" in plot_patch.paths
 
     aplt.FitImaging.normalized_residual_map(
         fit=masked_imaging_fit_x2_plane_7x7,
@@ -61,7 +61,7 @@ def test__fit_quantities_are_output(
         plotter=aplt.Plotter(output=aplt.Output(path=plot_path, format="png")),
     )
 
-    assert plot_path + "normalized_residual_map.png" in plot_patch.paths
+    assert f"{plot_path}/normalized_residual_map.png" in plot_patch.paths
 
     aplt.FitImaging.chi_squared_map(
         fit=masked_imaging_fit_x2_plane_7x7,
@@ -69,7 +69,7 @@ def test__fit_quantities_are_output(
         plotter=aplt.Plotter(output=aplt.Output(path=plot_path, format="png")),
     )
 
-    assert plot_path + "chi_squared_map.png" in plot_patch.paths
+    assert f"{plot_path}/chi_squared_map.png" in plot_patch.paths
 
 
 def test__fit_sub_plot(
@@ -81,7 +81,7 @@ def test__fit_sub_plot(
         sub_plotter=aplt.SubPlotter(output=aplt.Output(path=plot_path, format="png")),
     )
 
-    assert plot_path + "subplot_fit_imaging.png" in plot_patch.paths
+    assert f"{plot_path}/subplot_fit_imaging.png" in plot_patch.paths
 
 
 def test__subtracted_image_of_plane_is_output(
@@ -98,7 +98,7 @@ def test__subtracted_image_of_plane_is_output(
         plotter=aplt.Plotter(output=aplt.Output(plot_path, format="png")),
     )
 
[…]

[…]
