### 2020-10-31 b93b649
image path now handled by autofit, passed to visualize method
