+++ test_autolens/unit/plot/test_fit_interferometer_plots.py
@@ -18,35 +18,35 @@ def test__fit_quantities_are_output(
         plotter=aplt.Plotter(output=aplt.Output(path=plot_path, format="png")),
     )
 
-    assert plot_path + "visibilities.png" in plot_patch.paths
+    assert f"{plot_path}/visibilities.png" in plot_patch.paths
 
     aplt.FitInterferometer.noise_map(
         fit=masked_interferometer_fit_x2_plane_7x7,
         plotter=aplt.Plotter(output=aplt.Output(path=plot_path, format="png")),
     )
 
-    assert plot_path + "noise_map.png" in plot_patch.paths
+    assert f"{plot_path}/noise_map.png" in plot_patch.paths
 
     aplt.FitInterferometer.signal_to_noise_map(
         fit=masked_interferometer_fit_x2_plane_7x7,
         plotter=aplt.Plotter(output=aplt.Output(path=plot_path, format="png")),
     )
 
-    assert plot_path + "signal_to_noise_map.png" in plot_patch.paths
+    assert f"{plot_path}/signal_to_noise_map.png" in plot_patch.paths
 
     aplt.FitInterferometer.model_visibilities(
         fit=masked_interferometer_fit_x2_plane_7x7,
         plotter=aplt.Plotter(output=aplt.Output(path=plot_path, format="png")),
     )
 
-    assert plot_path + "model_visibilities.png" in plot_patch.paths
+    assert f"{plot_path}/model_visibilities.png" in plot_patch.paths
 
     aplt.FitInterferometer.residual_map_vs_uv_distances(
         fit=masked_interferometer_fit_x2_plane_7x7,
         plotter=aplt.Plotter(output=aplt.Output(path=plot_path, format="png")),
     )
 
-    assert plot_path + "residual_map_vs_uv_distances_real.png" in plot_patch.paths
+    assert f"{plot_path}/residual_map_vs_uv_distances_real.png" in plot_patch.paths
 
     aplt.FitInterferometer.residual_map_vs_uv_distances(
         fit=masked_interferometer_fit_x2_plane_7x7,
@@ -54,7 +54,7 @@ def test__fit_quantities_are_output(
         plotter=aplt.Plotter(output=aplt.Output(path=plot_path, format="png")),
     )
 
-    assert plot_path + "residual_map_vs_uv_distances_imag.png" in plot_patch.paths
+    assert f"{plot_path}/residual_map_vs_uv_distances_imag.png" in plot_patch.paths
 
     aplt.FitInterferometer.normalized_residual_map_vs_uv_distances(
         fit=masked_interferometer_fit_x2_plane_7x7,
@@ -62,7 +62,7 @@ def test__fit_quantities_are_output(
     )
 
     assert (
-        plot_path + "normalized_residual_map_vs_uv_distances_real.png"
+        f"{plot_path}/normalized_residual_map_vs_uv_distances_real.png"
         in plot_patch.paths
     )
 
@@ -73,7 +73,7 @@ def test__fit_quantities_are_output(
     )
 
     assert (
-        plot_path + "normalized_residual_map_vs_uv_distances_imag.png"
+        f"{plot_path}/normalized_residual_map_vs_uv_distances_imag.png"
         in plot_patch.paths
     )
 
@@ -82,7 +82,7 @@ def test__fit_quantities_are_output(
         plotter=aplt.Plotter(output=aplt.Output(path=plot_path, format="png")),
     )
 
-    assert plot_path + "chi_squared_map_vs_uv_distances_real.png" in plot_patch.paths
+    assert f"{plot_path}/chi_squared_map_vs_uv_distances_real.png" in plot_patch.paths
 
[…]
