+++ autolens/pipeline/phase/interferometer/analysis.py
@@ -10,9 +10,7 @@ from autolens.pipeline.phase.dataset import analysis as analysis_dataset
 
 
 class Analysis(ag_analysis.Analysis, analysis_dataset.Analysis):
-    def __init__(
-        self, masked_interferometer, settings, cosmology, image_path=None, results=None
-    ):
+    def __init__(self, masked_interferometer, settings, cosmology, results=None):
 
         super(Analysis, self).__init__(
             masked_dataset=masked_interferometer,
@@ -22,12 +20,7 @@ class Analysis(ag_analysis.Analysis, analysis_dataset.Analysis):
         )
 
         self.visualizer = visualizer.PhaseInterferometerVisualizer(
-            masked_dataset=masked_interferometer, image_path=image_path
-        )
-
-        self.visualizer.visualize_hyper_images(
-            hyper_galaxy_image_path_dict=self.hyper_galaxy_image_path_dict,
-            hyper_model_image=self.hyper_model_image,
+            masked_dataset=masked_interferometer
         )
 
         result = ag_analysis.last_result_with_use_as_hyper_dataset(results=results)
@@ -136,7 +129,15 @@ class Analysis(ag_analysis.Analysis, analysis_dataset.Analysis):
             settings_inversion=self.settings.settings_inversion,
         )
 
-    def visualize(self, instance, during_analysis):
+    def visualize(self, paths: af.Paths, instance, during_analysis):
+
+        self.visualizer.visualize_interferometer(paths=paths)
+
+        self.visualizer.visualize_hyper_images(
+            paths=paths,
+            hyper_galaxy_image_path_dict=self.hyper_galaxy_image_path_dict,
+            hyper_model_image=self.hyper_model_image,
+        )
 
         self.associate_hyper_images(instance=instance)
         tracer = self.tracer_for_instance(instance=instance)
@@ -166,9 +167,9 @@ class Analysis(ag_analysis.Analysis, analysis_dataset.Analysis):
             visualizer = self.visualizer
 
         visualizer.visualize_ray_tracing(
-            tracer=fit.tracer, during_analysis=during_analysis
+            paths=paths, tracer=fit.tracer, during_analysis=during_analysis
         )
-        visualizer.visualize_fit(fit=fit, during_analysis=during_analysis)
+        visualizer.visualize_fit(paths=paths, fit=fit, during_analysis=during_analysis)
 
     def make_attributes(self):
         return Attributes(

[…]
