+++ autolens/pipeline/phase/phase_galaxy.py
@@ -24,7 +24,7 @@ class AnalysisSingle(Analysis):
         fit = self.fit_for_instance(instance=instance)
         return fit.figure_of_merit
 
-    def visualize(self, instance, during_analysis):
+    def visualize(self, paths, instance, during_analysis):
         fit = self.fit_for_instance(instance=instance)
 
         self.visualizer.plot_galaxy_fit_subplot(fit)
@@ -78,7 +78,7 @@ class AnalysisDeflections(Analysis):
         fit_y, fit_x = self.fit_for_instance(instance=instance)
         return fit_y.figure_of_merit + fit_x.figure_of_merit
 
-    def visualize(self, instance, during_analysis):
+    def visualize(self, paths, instance, during_analysis):
 
         fit_y, fit_x = self.fit_for_instance(instance=instance)
 
