+++ howtolens/chapter_2_lens_modeling/scripts/tutorial_1_non_linear_search.py
@@ -218,7 +218,7 @@ print(
     " This Jupyter notebook cell with progress once Dynesty has completed - this could take some time!"
 )
 
-# result =  phase.run(dataset=imaging, mask=mask)
+result = phase.run(dataset=imaging, mask=mask)
 
 print("Dynesty has finished run - you may now continue the notebook.")
 
@@ -252,7 +252,7 @@ The best-fit solution (i.e. the maximum log likelihood) is stored in the `result
 """
 
 # %%
-# aplt.FitImaging.subplot_fit_imaging(fit=result.max_log_likelihood_fit)
+aplt.FitImaging.subplot_fit_imaging(fit=result.max_log_likelihood_fit)
 
 # %%
 """
