+++ howtolens/chapter_2_lens_modeling/scripts/tutorial_2_parameter_space_and_priors.py
@@ -206,9 +206,9 @@ print(
     " This Jupyter notebook cell with progress once Dynesty has completed - this could take some time!"
 )
 
-# result =  phase.run(dataset=imaging, mask=mask)
+result = phase.run(dataset=imaging, mask=mask)
 
-# aplt.FitImaging.subplot_fit_imaging(fit=result.max_log_likelihood_fit)
+aplt.FitImaging.subplot_fit_imaging(fit=result.max_log_likelihood_fit)
 
 print("Dynesty has finished run - you may now continue the notebook.")
 

[…]
