+++ test_autolens/unit/pipeline/test_visualizer.py
@@ -13,7 +13,7 @@ directory = path.dirname(path.realpath(__file__))
 
 @pytest.fixture(name="plot_path")
 def make_visualizer_plotter_setup():
-    return "{}/files/plot/visualizer/".format(
+    return "{}/files/plot/visualizer".format(
         os.path.dirname(os.path.realpath(__file__))
     )
 
@@ -50,7 +50,7 @@ class TestAbstractPhaseVisualizer:
             paths=af.Paths(), tracer=tracer_x2_plane_7x7, during_analysis=True
         )
 
-        assert f"{plot_path}/image//subplots/subplot_tracer.png" in plot_patch.paths
+        assert f"{plot_path}/image/subplots/subplot_tracer.png" in plot_patch.paths
 
 
 class TestPhaseDataSetVisualizer:
@@ -77,21 +77,23 @@ class TestPhaseDataSetVisualizer:
             paths=af.Paths(), tracer=tracer_x2_plane_7x7, during_analysis=False
         )
 
-        assert f"{plot_path}/image//subplots/subplot_tracer.png" in plot_patch.paths
-        assert f"{plot_path}/image//ray_tracing/image.png" in plot_patch.paths
-        assert f"{plot_path}/image//ray_tracing/source_plane.png" in plot_patch.paths
-        assert f"{plot_path}/image//ray_tracing/convergence.png" in plot_patch.paths
-        assert f"{plot_path}/image//ray_tracing/potential.png" not in plot_patch.paths
+        print(plot_patch.paths)
+
+        assert f"{plot_path}/image/subplots/subplot_tracer.png" in plot_patch.paths
+        assert f"{plot_path}/image/ray_tracing/image.png" in plot_patch.paths
+        assert f"{plot_path}/image/ray_tracing/source_plane.png" in plot_patch.paths
+        assert f"{plot_path}/image/ray_tracing/convergence.png" in plot_patch.paths
+        assert f"{plot_path}/image/ray_tracing/potential.png" not in plot_patch.paths
         assert (
-            f"{plot_path}/image//ray_tracing/deflections_y.png" not in plot_patch.paths
+            f"{plot_path}/image/ray_tracing/deflections_y.png" not in plot_patch.paths
         )
         assert (
-            f"{plot_path}/image//ray_tracing/deflections_x.png" not in plot_patch.paths
+            f"{plot_path}/image/ray_tracing/deflections_x.png" not in plot_patch.paths
         )
-        assert f"{plot_path}/image//ray_tracing/magnification.png" in plot_patch.paths
+        assert f"{plot_path}/image/ray_tracing/magnification.png" in plot_patch.paths
 
         convergence = al.util.array.numpy_array_2d_from_fits(
-            file_path=f"{plot_path}/image//ray_tracing/fits/convergence.fits", hdu=0
+            file_path=f"{plot_path}/image/ray_tracing/fits/convergence.fits", hdu=0
         )
 
         assert convergence.shape == (5, 5)
@@ -109,7 +111,7 @@ class TestPhaseDataSetVisualizer:
             during_analysis=True,
         )
         assert (
-            f"{plot_path}/image//other/stochastic_histogram.png" not in plot_patch.paths
+            f"{plot_path}/image/other/stochastic_histogram.png" not in plot_patch.paths
         )
 
         visualizer.visualize_stochastic_histogram(
@@ -118,7 +120,7 @@ class TestPhaseDataSetVisualizer:
             max_log_evidence=3.0,
             during_analysis=False,
         )
-        assert f"{plot_path}/image//other/stochastic_histogram.png" in plot_patch.paths
[…]
