+++ autolens/pipeline/phase/interferometer/analysis.py
@@ -72,7 +72,9 @@ class Analysis(ag_analysis.Analysis, analysis_dataset.Analysis):
             fit = self.masked_interferometer_fit_for_tracer(
                 tracer=tracer, hyper_background_noise=hyper_background_noise
             )
-            return fit.figure_of_merit
+            fig = fit.figure_of_merit
+            print(fig)
+            return fig
         except (
             PixelizationException,
             InversionException,
