### 2020-11-11 #2825 @zonca
@Jammy2211 I'll collect feedback on the repository here as I go through the review:

- [ ] I noticed that while there are a lot of tagged releases on PyPI, on Github there are only very old tags. Do you happen to have `git` tags available locally that you can push? otherwise, what about tagging on Github just the last 3/4 releases? It is convenient to be able to browse older versions of the code without downloading them from PyPI.
