### 2020-11-11 Jammy2211/PyAutoLens#145 @zonca
[JOSS] Installation instructions - conda
JOSS review: https://github.com/openjournals/joss-reviews/issues/2825

https://pyautolens.readthedocs.io/en/latest/general/installation.html#installation-with-conda

* I suggest to rename this section to "Installation in a conda environment", otherwise it seems there is conda package for `pyautolens`
* `anaconda` has really a lot of dependencies, I suggest to only install the largest compiled packages, for example `scipy` `astropy` `matplotlib` `numba`, and then let `pip` install the rest
