### 2020-11-11 Jammy2211/PyAutoLens#146 @zonca
[JOSS] unit tests failing on master
JOSS review: https://github.com/openjournals/joss-reviews/issues/2825

I am trying to run unit tests on the current master branch with `pytest`,
I see all the plotting tests failing plus another 2.
I tested `welcome.py` from the `autolens_workspace` and there plotting works fine.

Do I need to configure plotting specifically to run the tests?
Or is `master` currently broken and I should try another version?

I attach test `pytest` logs:
[pyautolens_pytest.log](https://github.com/Jammy2211/PyAutoLens/files/5521241/pyautolens_pytest.log)

And the packages in the conda env I'm using:
[environment.txt](https://github.com/Jammy2211/PyAutoLens/files/5521259/environment.txt)
