### 2020-11-11 Jammy2211/PyAutoLens#146 @Jammy2211
The failing tests are due to my poor handling of path construction and is actually on my to do list to fix in the next few days :). We're actually about to make autolens Windows compatbible which these tests would not play nice with at all, so need to switch everything to pathlib (should of been doing this all along but got lazy).

As you say, this shouldn't impact the actual code performance for you as the tests are failing because they're making assertions on the path string structure. I'll let you know once we've got this sorted.
