### 2020-11-11 Jammy2211/PyAutoLens#145 @Jammy2211
Good shouts, will update the docs accordingly :).

Do you think we should make the conda instructions top of the installations (and the 'recommended' method) or stick to pip at the top? Both have ups and downs and it feels like we swap them round every 6 months. Newer versions of llvmlite (for numba) have made certain pip installs a pain so going back to conda seems preferable currently...
