### 2020-11-11 Jammy2211/PyAutoLens#145 @Jammy2211
Yep, I'm expanding the installation instructions in the docs now :), getting things up and running via a git clone / fork is worth expanding on too as I'd imagine most users will want to do this for other projects (e.g. PyAutoGalaxy) where a large chunk of lensing calculations are actually performed.
