### 2020-11-11 #2825 @zonca
@Jammy2211 @xuanxu ok, my review is complete, if you address the 2 issues I opened, the repository feedback above and the suggestions on the paper below, we are done (also justifying why you don't want or can't to implement something I suggested is also valid, of course).

## Feedback on paper

### Summary

* I don't think we need to mention the names of the classes in the summary, better keep it more general, just mention they functionality
* HowToLens "lectures" suggests they are videos maybe? what about calling them "tutorial notebooks" or just specify "HowToLens lectures in Jupyter Notebook format"

### Background

* can you explain a bit how you go from left panel to center panel in Fig 1? Are you removing the foreground galaxy beforehand or you are fitting simultaneously for the foreground galaxy and the lens and then plotting only the lens?
* when you mention all the applications, maybe a bullet list would make it more readable?

### Software API and Features

* can you clarify what you mean by "abstracted objects"? I wonder if a programmer could think you are referring to abstract classes
* "By passing these objects a Grid ... are computed" what is computing this? is the `Tracer`? better specify
* Optimized by `numba`: is the code multi-threaded or not?
* As you mention performance, are there particular memory constraints? does the code use a lot of RAM for standard runs? No need to be too specific, but at least give the idea of what I can run easily on my 4 GB RAM laptop and when I need a HPC node
* can you specify what is a `uv-plane`?
* I would reserve `backticks` formatting just for names of classes or software, if you want to highlight works maybe use *corsive* or **bold**?

### Workspace and HowToLens Tutorials

* In case you decide to add binder support (see above) you can add it here
