+++ howtolens/chapter_2_lens_modeling/scripts/tutorial_4_dealing_with_failure.py
@@ -37,12 +37,12 @@ we'll use the same strong lensing data as the previous tutorial, where:
 
 # %%
 dataset_name = "light_sersic__mass_sie__source_exp"
-dataset_path = f"dataset/howtolens/chapter_2/{dataset_name}"
+dataset_path = path.join("dataset", "howtolens", "chapter_2", dataset_name)
 
 imaging = al.Imaging.from_fits(
-    image_path=f"{dataset_path}/image.fits",
-    noise_map_path=f"{dataset_path}/noise_map.fits",
-    psf_path=f"{dataset_path}/psf.fits",
+    image_path=path.join(dataset_path, "image.fits"),
+    noise_map_path=path.join(dataset_path, "noise_map.fits"),
+    psf_path=path.join(dataset_path, "psf.fits"),
     pixel_scales=0.1,
 )
 
@@ -199,7 +199,7 @@ regions of parameter space, given our improved and more informed priors.
 # %%
 phase = al.PhaseImaging(
     search=af.DynestyStatic(
-        path_prefix=f"howtolens", name="phase_t4_custom_priors", n_live_points=50
+        path_prefix="howtolens", name="phase_t4_custom_priors", n_live_points=50
     ),
     settings=settings,
     galaxies=dict(lens=lens, source=source),
@@ -288,7 +288,7 @@ Again, we create this phase and run it. The `NonLinearSearch` now has a less com
 # %%
 phase_light_traces_mass = al.PhaseImaging(
     search=af.DynestyStatic(
-        path_prefix=f"howtolens", name="phase_t4_light_traces_mass", n_live_points=40
+        path_prefix="howtolens", name="phase_t4_light_traces_mass", n_live_points=40
     ),
     settings=settings,
     galaxies=dict(lens=lens, source=source),
