+++ howtolens/simulators/chapter_3/light_sersic_x2__mass_sie_x2__source_exp.py
@@ -26,7 +26,7 @@ dataset_name = "light_sersic_x2__mass_sie_x2__source_exp"
 The path where the dataset will be output, which in this case is:
 `/autolens_workspace/howtolens/dataset/chapter_3/light_sersic_x2__mass_sie_x2__source_exp/`
 """
-dataset_path = f"dataset/howtolens/{dataset_type}/{dataset_name}"
+dataset_path = path.join("dataset", "howtolens", dataset_type, dataset_name)
 
 """
 For simulating an image of a strong lens, we recommend using a GridIterate object. This represents a grid of $(y,x)$ 
@@ -115,9 +115,9 @@ imaging = simulator.from_tracer_and_grid(tracer=tracer, grid=grid)
 
 """Output our simulated dataset to the dataset path as .fits files"""
 imaging.output_to_fits(
-    image_path=f"{dataset_path}/image.fits",
-    psf_path=f"{dataset_path}/psf.fits",
-    noise_map_path=f"{dataset_path}/noise_map.fits",
+    image_path=path.join(dataset_path, "image.fits"),
+    psf_path=path.join(dataset_path, "psf.fits"),
+    noise_map_path=path.join(dataset_path, "noise_map.fits"),
     overwrite=True,
 )
 
