+++ howtolens/chapter_3_pipelines/scripts/tutorial_1_pipeline_runner.py
@@ -53,12 +53,12 @@ we'll use strong lensing data, where:
 
 # %%
 dataset_name = "light_sersic__mass_sie__source_exp"
-dataset_path = f"dataset/howtolens/chapter_3/{dataset_name}"
+dataset_path = path.join("dataset", "howtolens", "chapter_3", dataset_name)
 
 imaging = al.Imaging.from_fits(
-    image_path=f"{dataset_path}/image.fits",
-    noise_map_path=f"{dataset_path}/noise_map.fits",
-    psf_path=f"{dataset_path}/psf.fits",
+    image_path=path.join(dataset_path, "image.fits"),
+    noise_map_path=path.join(dataset_path, "noise_map.fits"),
+    psf_path=path.join(dataset_path, "psf.fits"),
     pixel_scales=0.1,
 )
 

[…]
