### 2020-11-13 00e5396
subhalo plot paths
