+++ autolens/pipeline/phase/imaging/phase.py
@@ -1,3 +1,4 @@
+from os import path
 import autofit as af
 import autogalaxy as ag
 from astropy import cosmology as cosmo
@@ -125,7 +126,7 @@ class PhaseImaging(dataset.PhaseDataset):
 
     def output_phase_info(self):
 
-        file_phase_info = "{}/{}".format(self.search.paths.output_path, "phase.info")
+        file_phase_info = path.join(self.search.paths.output_path, "phase.info")
 
         with open(file_phase_info, "w") as phase_info:
             phase_info.write("Optimizer = {} \n".format(type(self.search).__name__))
