+++ autolens/pipeline/visualizer.py
@@ -267,7 +267,7 @@ class PhaseDatasetVisualizer(AbstractVisualizer):
     def visualize_ray_tracing_in_fits(self, paths: af.Paths, tracer):
 
         fits_plotter = self.plotter_from_paths(
-            paths=paths, subfolders="ray_tracing/fits", format="fits"
+            paths=paths, subfolders=path.join("ray_tracing", "fits"), format="fits"
         )
 
         ray_tracing_plots.individual(
@@ -542,7 +542,7 @@ class PhaseImagingVisualizer(PhaseDatasetVisualizer):
     def visualize_fit_in_fits(self, paths: af.Paths, fit):
 
         fits_plotter = self.plotter_from_paths(
-            paths=paths, subfolders="fit_imaging/fits", format="fits"
+            paths=paths, subfolders=path.join("fit_imaging", "fits"), format="fits"
         )
 
         fit_imaging_plots.individuals(
@@ -564,7 +564,7 @@ class PhaseImagingVisualizer(PhaseDatasetVisualizer):
         if fit.inversion is not None:
 
             fits_plotter = self.plotter_from_paths(
-                paths=paths, subfolders="inversion/fits", format="fits"
+                paths=paths, subfolders=path.join("inversion", "fits"), format="fits"
             )
 
             inversion_plots.individuals(
@@ -757,7 +757,7 @@ class PhaseInterferometerVisualizer(PhaseDatasetVisualizer):
             if self.plot_fit_all_at_end_fits:
 
                 fits_plotter = self.plotter_from_paths(
-                    paths=paths, subfolders="fit_interferometer/fits", format="fits"
+                    paths=paths, subfolders=path.join("fit_interferometer", "fits"), format="fits"
                 )
 
                 fit_interferometer_plots.individuals(
@@ -775,7 +775,7 @@ class PhaseInterferometerVisualizer(PhaseDatasetVisualizer):
 
                 if fit.inversion is not None:
                     fits_plotter = self.plotter_from_paths(
-                        paths=paths, subfolders="inversion/fits", format="fits"
+                        paths=paths, subfolders=path.join("inversion", "fits"), format="fits"
                     )
 
                     inversion_plots.individuals(
