+++ howtolens/simulators/chapter_2/mass_sis__source_exp_x2.py
@@ -1,6 +1,5 @@
-import autofit as af
+from os import path
 import autolens as al
-import autolens.plot as aplt
 
 """
 This script simulates `Imaging` of a strong lens where:
@@ -19,6 +18,7 @@ gives it a descriptive name. They define the folder the dataset is output to on 
  - The noise-map will be output to `/autolens_workspace/dataset/dataset_type/dataset_name/lens_name/noise_map.fits`.
  - The psf will be output to `/autolens_workspace/dataset/dataset_type/dataset_name/psf.fits`.
 """
+
 dataset_type = "chapter_2"
 dataset_name = "mass_sis__source_exp_x2"
 
@@ -37,6 +37,7 @@ sub-size of the grid is iteratively increased (in steps of 2, 4, 8, 16, 24) unti
 This ensures that the divergent and bright central regions of the source galaxy are fully resolved when determining the
 total flux emitted within a pixel.
 """
+
 grid = al.GridIterate.uniform(
     shape_2d=(100, 100),
     pixel_scales=0.1,
@@ -45,6 +46,7 @@ grid = al.GridIterate.uniform(
 )
 
 """Simulate a simple Gaussian PSF for the image."""
+
 psf = al.Kernel.from_gaussian(
     shape_2d=(11, 11), sigma=0.1, pixel_scales=grid.pixel_scales
 )
@@ -68,6 +70,7 @@ in degrees and defined counter clockwise from the positive x-axis.
 
 We can use the **PyAutoLens** `convert` module to determine the elliptical components from the axis-ratio and phi.
 """
+
 lens_galaxy = al.Galaxy(
     redshift=0.5, mass=al.mp.SphericalIsothermal(centre=(0.0, 0.0), einstein_radius=1.6)
 )
@@ -95,9 +98,11 @@ tracer = al.Tracer.from_galaxies(
 We can now pass this simulator a tracer, which creates the ray-traced image plotted above and simulates it as an
 imaging dataset.
 """
+
 imaging = simulator.from_tracer_and_grid(tracer=tracer, grid=grid)
 
 """Output our simulated dataset to the dataset path as .fits files"""
+
 imaging.output_to_fits(
     image_path=path.join(dataset_path, "image.fits"),
     psf_path=path.join(dataset_path, "psf.fits"),
@@ -111,4 +116,5 @@ check how the dataset was simulated in the future.
 
 This will also be accessible via the `Aggregator` if a model-fit is performed using the dataset.
 """
+
 tracer.save(file_path=dataset_path, filename="true_tracer")
