+++ howtolens/chapter_2_lens_modeling/scripts/tutorial_2_parameter_space_and_priors.py
@@ -97,6 +97,7 @@ workspace_path = str(here())
 #%cd $workspace_path
 print(f"Working Directory has been set to `{workspace_path}`")
 
+from os import path
 import autolens as al
 import autolens.plot as aplt
 import autofit as af
