+++ howtolens/chapter_4_inversions/scripts/tutorial_5_borders.py
@@ -15,6 +15,7 @@ workspace_path = str(here())
 #%cd $workspace_path
 print(f"Working Directory has been set to `{workspace_path}`")
 
+from os import path
 import autolens as al
 import autolens.plot as aplt
 
