+++ howtolens/chapter_2_lens_modeling/scripts/tutorial_1_non_linear_search.py
@@ -63,6 +63,7 @@ workspace_path = str(here())
 #%cd $workspace_path
 print(f"Working Directory has been set to `{workspace_path}`")
 
+from os import path
 import autofit as af  # <- This library is used for non-linear fitting.
 import autolens as al
 import autolens.plot as aplt
