+++ test_autolens/unit/test_simulate_and_fit.py
@@ -1,4 +1,5 @@
 import os
+from os import path
 import shutil
 
 import autolens as al
@@ -35,28 +36,30 @@ def test__simulate_imaging_data_and_fit__no_psf_blurring__chi_squared_is_0__nois
 
     imaging.noise_map = al.Array.ones(shape_2d=imaging.image.shape_2d, pixel_scales=0.2)
 
-    path = "{}/data_temp/simulate_and_fit".format(
-        os.path.dirname(os.path.realpath(__file__))
-    )  # Setup path so we can output the simulated image.
+    file_path = path.join(
+        "{}".format(path.dirname(path.realpath(__file__))),
+        "data_temp",
+        "simulate_and_fit",
+    )
 
     try:
-        shutil.rmtree(path)
+        shutil.rmtree(file_path)
     except FileNotFoundError:
         pass
 
-    if os.path.exists(path) is False:
-        os.makedirs(path)
+    if path.exists(file_path) is False:
+        os.makedirs(file_path)
 
     imaging.output_to_fits(
-        image_path=f"{path}/image.fits",
-        noise_map_path=f"{path}/noise_map.fits",
-        psf_path=f"{path}/psf.fits",
+        image_path=path.join(file_path, "image.fits"),
+        noise_map_path=path.join(file_path, "noise_map.fits"),
+        psf_path=path.join(file_path, "psf.fits"),
     )
 
     imaging = al.Imaging.from_fits(
-        image_path=f"{path}/image.fits",
-        noise_map_path=f"{path}/noise_map.fits",
-        psf_path=f"{path}/psf.fits",
+        image_path=path.join(file_path, "image.fits"),
+        noise_map_path=path.join(file_path, "noise_map.fits"),
+        psf_path=path.join(file_path, "psf.fits"),
         pixel_scales=0.2,
     )
 
@@ -76,12 +79,12 @@ def test__simulate_imaging_data_and_fit__no_psf_blurring__chi_squared_is_0__nois
 
     assert fit.chi_squared == 0.0
 
-    path = "{}/data_temp".format(
-        os.path.dirname(os.path.realpath(__file__))
-    )  # Setup path so we can output the simulated image.
-
-    if os.path.exists(path) == True:
-        shutil.rmtree(path)
+    file_path = path.join(
+        "{}".format(path.dirname(path.realpath(__file__))),
+        "data_temp",
+    )
+    if path.exists(file_path) == True:
+        shutil.rmtree(file_path)
 
 
 def test__simulate_imaging_data_and_fit__include_psf_blurring__chi_squared_is_0__noise_normalization_correct():
@@ -110,28 +113,30 @@ def test__simulate_imaging_data_and_fit__include_psf_blurring__chi_squared_is_0_
     imaging = simulator.from_tracer_and_grid(tracer=tracer, grid=grid)
     imaging.noise_map = al.Array.ones(shape_2d=imaging.image.shape_2d, pixel_scales=0.2)
 
-    path = "{}/data_temp/simulate_and_fit".format(
-        os.path.dirname(os.path.realpath(__file__))
-    )  # Setup path so we can output the simulated image.
+    file_path = path.join(
+        "{}".format(path.dirname(path.realpath(__file__))),
+        "data_temp",
+        "simulate_and_fit",
+    )
 
     try:
-        shutil.rmtree(path)
+        shutil.rmtree(file_path)
     except FileNotFoundError:
         pass
 
-    if os.path.exists(path) is False:
-        os.makedirs(path)
+    if path.exists(file_path) is False:
+        os.makedirs(file_path)
 
     imaging.output_to_fits(
-        image_path=f"{path}/image.fits",
-        noise_map_path=f"{path}/noise_map.fits",
[…]
