+++ howtolens/chapter_2_lens_modeling/scripts/tutorial_6_alternative_searches.py
@@ -17,6 +17,7 @@ workspace_path = str(here())
 #%cd $workspace_path
 print(f"Working Directory has been set to `{workspace_path}`")
 
+from os import path
 import autolens as al
 import autolens.plot as aplt
 import autofit as af
