+++ howtolens/chapter_4_inversions/scripts/tutorial_6_lens_modeling.py
@@ -25,6 +25,7 @@ workspace_path = str(here())
 #%cd $workspace_path
 print(f"Working Directory has been set to `{workspace_path}`")
 
+from os import path
 import autolens as al
 import autolens.plot as aplt
 
