+++ docs/installation/troubleshooting.rst
@@ -1,5 +1,8 @@
 .. _troubleshooting:
 
+Troubleshooting
+===============
+
 LLVMLite / numba
 ================
 
