+++ test_autolens/unit/pipeline/phase/imaging/test_analysis_imaging.py
@@ -22,7 +22,7 @@ class TestLogLikelihoodFunction:
         self, phase_imaging_7x7, imaging_7x7, mask_7x7
     ):
 
-        imaging_7x7.positions = al.GridCoordinates([[(1.0, 100.0), (200.0, 2.0)]])
+        imaging_7x7.positions = al.GridIrregularGrouped([[(1.0, 100.0), (200.0, 2.0)]])
 
         phase_imaging_7x7 = al.PhaseImaging(
             galaxies=dict(
