+++ test_autolens/unit/fit/test_fit_positions.py
@@ -15,35 +15,35 @@ class MockTracerPositions:
 class TestAbstractFitPositionsSourcePlane:
     def test__x1_positions__mock_position_tracer__maximum_separation_is_correct(self):
 
-        positions = al.GridCoordinates(coordinates=[[(0.0, 0.0), (0.0, 1.0)]])
+        positions = al.GridIrregularGrouped(coordinates=[[(0.0, 0.0), (0.0, 1.0)]])
         tracer = MockTracerPositions(positions=positions)
         fit = al.FitPositionsSourcePlaneMaxSeparation(
             positions=positions, tracer=tracer, noise_value=1.0
         )
         assert fit.maximum_separations[0] == 1.0
 
-        positions = al.GridCoordinates([[(0.0, 0.0), (1.0, 1.0)]])
+        positions = al.GridIrregularGrouped([[(0.0, 0.0), (1.0, 1.0)]])
         tracer = MockTracerPositions(positions=positions)
         fit = al.FitPositionsSourcePlaneMaxSeparation(
             positions=positions, tracer=tracer, noise_value=1.0
         )
         assert fit.maximum_separations[0] == np.sqrt(2)
 
-        positions = al.GridCoordinates([[(0.0, 0.0), (1.0, 3.0)]])
+        positions = al.GridIrregularGrouped([[(0.0, 0.0), (1.0, 3.0)]])
         tracer = MockTracerPositions(positions=positions)
         fit = al.FitPositionsSourcePlaneMaxSeparation(
             positions=positions, tracer=tracer, noise_value=1.0
         )
         assert fit.maximum_separations[0] == np.sqrt(np.square(1.0) + np.square(3.0))
 
-        positions = al.GridCoordinates([[(-2.0, -4.0), (1.0, 3.0)]])
+        positions = al.GridIrregularGrouped([[(-2.0, -4.0), (1.0, 3.0)]])
         tracer = MockTracerPositions(positions=positions)
         fit = al.FitPositionsSourcePlaneMaxSeparation(
             positions=positions, tracer=tracer, noise_value=1.0
         )
         assert fit.maximum_separations[0] == np.sqrt(np.square(3.0) + np.square(7.0))
 
-        positions = al.GridCoordinates([[(8.0, 4.0), (-9.0, -4.0)]])
+        positions = al.GridIrregularGrouped([[(8.0, 4.0), (-9.0, -4.0)]])
         tracer = MockTracerPositions(positions=positions)
         fit = al.FitPositionsSourcePlaneMaxSeparation(
             positions=positions, tracer=tracer, noise_value=1.0
@@ -53,28 +53,28 @@ class TestAbstractFitPositionsSourcePlane:
     def test_multiple_positions__mock_position_tracer__maximum_separation_is_correct(
         self,
     ):
-        positions = al.GridCoordinates([[(0.0, 0.0), (0.0, 1.0), (0.0, 0.5)]])
+        positions = al.GridIrregularGrouped([[(0.0, 0.0), (0.0, 1.0), (0.0, 0.5)]])
         tracer = MockTracerPositions(positions=positions)
         fit = al.FitPositionsSourcePlaneMaxSeparation(
             positions=positions, tracer=tracer, noise_value=1.0
         )
         assert fit.maximum_separations[0] == 1.0
 
-        positions = al.GridCoordinates([[(0.0, 0.0), (0.0, 0.0), (3.0, 3.0)]])
+        positions = al.GridIrregularGrouped([[(0.0, 0.0), (0.0, 0.0), (3.0, 3.0)]])
         tracer = MockTracerPositions(positions=positions)
[…]
