+++ test_autolens/unit/pipeline/phase/abstract/test_result_abstract.py
@@ -52,7 +52,9 @@ class TestTracer:
             dataset=imaging_7x7, mask=mask_7x7, results=mock.MockResults()
         )
 
-        assert result.source_plane_light_profile_centres.in_grouped_list == [[(1.0, 2.0)]]
+        assert result.source_plane_light_profile_centres.in_grouped_list == [
+            [(1.0, 2.0)]
+        ]
 
         source = al.Galaxy(
             redshift=1.0,
@@ -122,7 +124,7 @@ class TestTracer:
             np.array([0.80, 0.80, 0.80, 0.80, 0.80, 0.80, 0.80, 0.80, 0.80]), 1.0e-1
         )
 
-        assert result.source_plane_inversion_centres.in_grouped_list == [[(0.0, 0.0)]]
+        assert result.source_plane_inversion_centres.in_1d_list == [(0.0, 0.0)]
 
         lens = al.Galaxy(redshift=0.5, light=al.lp.SphericalSersic(intensity=1.0))
         source = al.Galaxy(redshift=1.0)
@@ -221,6 +223,15 @@ class TestTracer:
             source_plane_coordinate=result.source_plane_inversion_centres[0],
         )
 
-        assert multiple_images.in_grouped_list[0] == multiple_images_manual_0.in_grouped_list[0]
-        assert multiple_images.in_grouped_list[1] == multiple_images_manual_1.in_grouped_list[0]
-        assert multiple_images.in_grouped_list[2] == multiple_images_manual_2.in_grouped_list[0]
+        assert (
+            multiple_images.in_grouped_list[0]
+            == multiple_images_manual_0.in_grouped_list[0]
+        )
+        assert (
+            multiple_images.in_grouped_list[1]
+            == multiple_images_manual_1.in_grouped_list[0]
+        )
+        assert (
+            multiple_images.in_grouped_list[2]
+            == multiple_images_manual_2.in_grouped_list[0]
+        )
