+++ test_autolens/unit/lens/test_settings.py
@@ -42,7 +42,8 @@ class TestCheckPositionsTrace:
         settings = al.SettingsLens(positions_threshold=0.0)
         with pytest.raises(exc.RayTracingException):
             settings.check_positions_trace_within_threshold_via_tracer(
-                tracer=tracer, positions=al.GridIrregularGrouped([[(1.0, 1.0), (2.0, 2.0)]])
+                tracer=tracer,
+                positions=al.GridIrregularGrouped([[(1.0, 1.0), (2.0, 2.0)]]),
             )
 
         # No mass profile - doesnt raise exception
