+++ test_autolens/unit/pipeline/test_visualizer.py
@@ -40,10 +40,8 @@ class TestAbstractPhaseVisualizer:
         assert visualizer.include.preloaded_critical_curves == None
         assert visualizer.include.preloaded_caustics == None
 
-        visualizer = (
-            visualizer.new_visualizer_with_preloaded_critical_curves_and_caustics(
-                preloaded_critical_curves=1, preloaded_caustics=2
-            )
+        visualizer = visualizer.new_visualizer_with_preloaded_critical_curves_and_caustics(
+            preloaded_critical_curves=1, preloaded_caustics=2
         )
 
         assert visualizer.include.preloaded_critical_curves == 1
@@ -76,11 +74,9 @@ class TestPhaseDataSetVisualizer:
 
         visualizer = vis.PhaseDatasetVisualizer(masked_dataset=masked_imaging_7x7)
 
-        visualizer = (
-            visualizer.new_visualizer_with_preloaded_critical_curves_and_caustics(
-                preloaded_critical_curves=include_all.preloaded_critical_curves,
-                preloaded_caustics=include_all.preloaded_caustics,
-            )
+        visualizer = visualizer.new_visualizer_with_preloaded_critical_curves_and_caustics(
+            preloaded_critical_curves=include_all.preloaded_critical_curves,
+            preloaded_caustics=include_all.preloaded_caustics,
         )
 
         visualizer.visualize_ray_tracing(
@@ -168,11 +164,9 @@ class TestPhaseImagingVisualizer:
 
         visualizer = vis.PhaseImagingVisualizer(masked_dataset=masked_imaging_7x7)
 
-        visualizer = (
-            visualizer.new_visualizer_with_preloaded_critical_curves_and_caustics(
-                preloaded_critical_curves=include_all.preloaded_critical_curves,
-                preloaded_caustics=include_all.preloaded_caustics,
-            )
+        visualizer = visualizer.new_visualizer_with_preloaded_critical_curves_and_caustics(
+            preloaded_critical_curves=include_all.preloaded_critical_curves,
+            preloaded_caustics=include_all.preloaded_caustics,
         )
 
         visualizer.visualize_imaging(paths=af.Paths())
@@ -218,11 +212,9 @@ class TestPhaseImagingVisualizer:
 
         visualizer = vis.PhaseImagingVisualizer(masked_dataset=masked_imaging_7x7)
 
-        visualizer = (
-            visualizer.new_visualizer_with_preloaded_critical_curves_and_caustics(
-                preloaded_critical_curves=include_all.preloaded_critical_curves,
-                preloaded_caustics=include_all.preloaded_caustics,
-            )
+        visualizer = visualizer.new_visualizer_with_preloaded_critical_curves_and_caustics(
+            preloaded_critical_curves=include_all.preloaded_critical_curves,
+            preloaded_caustics=include_all.preloaded_caustics,
         )
 
         visualizer.visualize_fit(
@@ -369,11 +361,9 @@ class TestPhaseImagingVisualizer:
 
         visualizer = vis.PhaseImagingVisualizer(masked_dataset=masked_imaging_7x7)
 
-        visualizer = (
-            visualizer.new_visualizer_with_preloaded_critical_curves_and_caustics(
-                preloaded_critical_curves=include_all.preloaded_critical_curves,
-                preloaded_caustics=include_all.preloaded_caustics,
-            )
+        visualizer = visualizer.new_visualizer_with_preloaded_critical_curves_and_caustics(
+            preloaded_critical_curves=include_all.preloaded_critical_curves,
+            preloaded_caustics=include_all.preloaded_caustics,
         )
 
         visualizer.visualize_hyper_images(
@@ -400,11 +390,9 @@ class TestPhaseInterferometerVisualizer:
[…]
