+++ autolens/pipeline/phase/interferometer/analysis.py
@@ -113,9 +113,9 @@ class Analysis(ag_analysis.Analysis, analysis_dataset.Analysis):
             ):
                 if galaxy_path in self.hyper_galaxy_visibilities_path_dict:
                     galaxy.hyper_model_visibilities = self.hyper_model_visibilities
-                    galaxy.hyper_galaxy_visibilities = (
-                        self.hyper_galaxy_visibilities_path_dict[galaxy_path]
-                    )
+                    galaxy.hyper_galaxy_visibilities = self.hyper_galaxy_visibilities_path_dict[
+                        galaxy_path
+                    ]
 
         return instance
 
