+++ autolens/__init__.py
@@ -84,4 +84,4 @@ from autoconf import conf
 
 conf.instance.register(__file__)
 
-__version__ = '1.8.0'
+__version__ = '1.8.1'
