+++ test_autolens/simulators/imaging/instrument_util.py
@@ -1,0 +1,219 @@
+from os import path
+import autolens as al
+import autolens.plot as aplt
+
+from test_autogalaxy.simulators.imaging import instrument_util
+
+test_path = path.join("{}".format(path.dirname(path.realpath(__file__))), "..", "..")
+
+
+def pixel_scale_from_instrument(instrument):
+    """
+    Returns the pixel scale from an instrument type based on real observations.
+
+    These options are representative of VRO, Euclid, HST, over-sampled HST and Adaptive Optics image.
+
+    Parameters
+    ----------
+    instrument : str
+        A string giving the resolution of the desired instrument (VRO | Euclid | HST | HST_Up | AO).
+    """
+    if instrument in "vro":
+        return (0.2, 0.2)
+    elif instrument in "euclid":
+        return (0.1, 0.1)
+    elif instrument in "hst":
+        return (0.05, 0.05)
+    elif instrument in "hst_up":
+        return (0.03, 0.03)
+    elif instrument in "ao":
+        return (0.01, 0.01)
+    else:
+        raise ValueError("An invalid instrument was entered - ", instrument)
+
+
+def grid_from_instrument(instrument):
+    """
+    Returns the `Grid` from an instrument type based on real observations.
+
+    These options are representative of VRO, Euclid, HST, over-sampled HST and Adaptive Optics image.
+
+    Parameters
+    ----------
+    instrument : str
+        A string giving the resolution of the desired instrument (VRO | Euclid | HST | HST_Up | AO).
+    """
+    if instrument in "vro":
+        return al.GridIterate.uniform(shape_2d=(80, 80), pixel_scales=0.2)
+    elif instrument in "euclid":
+        return al.GridIterate.uniform(shape_2d=(120, 120), pixel_scales=0.1)
+    elif instrument in "hst":
+        return al.GridIterate.uniform(shape_2d=(200, 200), pixel_scales=0.05)
+    elif instrument in "hst_up":
+        return al.GridIterate.uniform(shape_2d=(300, 300), pixel_scales=0.03)
+    elif instrument in "ao":
+        return al.GridIterate.uniform(shape_2d=(800, 800), pixel_scales=0.01)
+    else:
+        raise ValueError("An invalid instrument was entered - ", instrument)
+
+
+def psf_from_instrument(instrument):
+    """
+    Returns the *PSF* from an instrument type based on real observations.
+
+    These options are representative of VRO, Euclid, HST, over-sampled HST and Adaptive Optics image.
+
+    Parameters
+    ----------
+    instrument : str
+        A string giving the resolution of the desired instrument (VRO | Euclid | HST | HST_Up | AO).
+    """
+    if instrument in "vro":
+        return al.Kernel.from_gaussian(
+            shape_2d=(31, 31), sigma=0.5, pixel_scales=0.2, renormalize=True
+        )
+
+    elif instrument in "euclid":
+        return al.Kernel.from_gaussian(
+            shape_2d=(31, 31), sigma=0.1, pixel_scales=0.1, renormalize=True
+        )
+    elif instrument in "hst":
+        return al.Kernel.from_gaussian(
+            shape_2d=(31, 31), sigma=0.05, pixel_scales=0.05, renormalize=True
+        )
+    elif instrument in "hst_up":
+        return al.Kernel.from_gaussian(
+            shape_2d=(31, 31), sigma=0.05, pixel_scales=0.03, renormalize=True
+        )
+    elif instrument in "ao":
+        return al.Kernel.from_gaussian(
+            shape_2d=(31, 31), sigma=0.025, pixel_scales=0.01, renormalize=True
+        )
+
+    else:
+        raise ValueError("An invalid instrument was entered - ", instrument)
+
+
[…]
