+++ test_autolens/integration/tests/imaging/phase_features/sub_grid_size.py
@@ -4,7 +4,7 @@ from test_autolens.integration.tests.imaging import runner
 
 test_type = "phase_features"
 test_name = "sub_size"
-dataset_name = "light_dev_vaucouleurs"
+dataset_name = "light_sersic__source_sersic"
 instrument = "vro"
 
 
