+++ test_autolens/plot/arrays/array_mask_and_zoom_elliptical_mask.py
@@ -7,7 +7,7 @@ from test_autolens.simulators.imaging import instrument_util
 
 # Lets quickly remind ourselves of the image, and the 3.0" circular mask we'll use to mask it.
 imaging = instrument_util.load_test_imaging(
-    dataset_name="light_dev_vaucouleurs", instrument="vro"
+    dataset_name="light_sersic__source_sersic", instrument="vro"
 )
 array = imaging.image
 
