### 2020-11-18 #2825 @aureliocarnero
Dear @Jammy2211 , I have some comments about the paper:
1) This phrase in not completely accurate:
"Strong gravitational lensing, where a background source galaxy appears multiple times due to its light rays being bent by the mass of one or more foreground lens galaxies"
Sometimes we do not get multiple images, but just a deformed image of the background source.
2) "When two galaxies are aligned down the line-of-sight to Earth, light rays from the background galaxy are bent by the intervening mass of one or more foreground galaxies"
Maybe change to:
 "When two galaxies are aligned down the line-of-sight to Earth, light rays from the background galaxy are bent by the intervening mass of the foreground galaxy"
3)  "Its light can be fully bent around the foreground galaxies, traversing multiple paths to the Earth, meaning that the background galaxy is observed multiple times."
Maybe change to:
"Sometimes its light can be fully bent, traversing multiple paths to the Earth, meaning that the background galaxy is observed multiple times."
4) An error here:
"Strong lensing is a competitive test of cosmological models, for example the **exapnsion** rate of the Universe."
5) I would start a new paragraph here:
The past decade has ...
6) missing an **of** here:
"In the next decade of order **of** "
7) remove the **a** here (is plural?):
"By passing these objects **a** Grid strong lens sightlines are computed"
8) it is not neccesary to indicate is python only here: "ensuring PyAutoLens’s Python-only code base runs efficiently." Also is not plural
"ensuring PyAutoLens runs efficiently."
9) and/or is extensively use, but I would not recommend it, just place and:
"which uses positions and flux-ratios"
10) Missing a coma here:
"about simpler lens models fitted in earlier phases to subsequent phases**,** which fit progressively more complex models"
11) singular phrase here (not plural):
"PyAutoFit’s aggregator tool "
12)  hard-disc should be **hard-disk**
13) Suggestion
"This uses memory-light Python generators, ensuring it is practical for fits to thousands of lenses."
to
"This uses memory-light Python generators, ensuring it is practical for thousands of lenses."

That's for now
