+++ test_autolens/config/visualize/subplots.ini
@@ -96,6 +96,19 @@ style=-
 colors=k
 pointsize=20
 
+[vector_quiverer]
+headlength=0
+pivot=middle
+linewidth=5
+units=xy
+angles=xy
+headwidth=1
+alpha=1.
+
+[patcher]
+facecolor=none
+edgecolor=cyan
+
 [array_overlayer]
 alpha=0.5
 
