+++ test_autolens/unit/fit/test_fit.py
@@ -744,8 +744,10 @@ class TestFitImaging:
                 hyper_background_noise=hyper_background_noise,
             )
 
-            hyper_noise_map_background = hyper_background_noise.hyper_noise_map_from_noise_map(
-                noise_map=masked_imaging_7x7.noise_map
+            hyper_noise_map_background = (
+                hyper_background_noise.hyper_noise_map_from_noise_map(
+                    noise_map=masked_imaging_7x7.noise_map
+                )
             )
 
             hyper_noise = tracer.hyper_noise_map_from_noise_map(
@@ -815,10 +817,12 @@ class TestFitImaging:
 
             fit = al.FitImaging(masked_imaging=masked_imaging_7x7, tracer=tracer)
 
-            blurred_images_of_planes = tracer.blurred_images_of_planes_from_grid_and_convolver(
-                grid=masked_imaging_7x7.grid,
-                convolver=masked_imaging_7x7.convolver,
-                blurring_grid=masked_imaging_7x7.blurring_grid,
+            blurred_images_of_planes = (
+                tracer.blurred_images_of_planes_from_grid_and_convolver(
+                    grid=masked_imaging_7x7.grid,
+                    convolver=masked_imaging_7x7.convolver,
+                    blurring_grid=masked_imaging_7x7.blurring_grid,
+                )
             )
 
             assert blurred_images_of_planes[0].in_2d == pytest.approx(
@@ -835,8 +839,10 @@ class TestFitImaging:
 
             assert (unmasked_blurred_image == fit.unmasked_blurred_image).all()
 
-            unmasked_blurred_image_of_planes = tracer.unmasked_blurred_image_of_planes_from_grid_and_psf(
-                grid=masked_imaging_7x7.grid, psf=masked_imaging_7x7.psf
+            unmasked_blurred_image_of_planes = (
+                tracer.unmasked_blurred_image_of_planes_from_grid_and_psf(
+                    grid=masked_imaging_7x7.grid, psf=masked_imaging_7x7.psf
+                )
             )
 
             assert (
@@ -848,8 +854,10 @@ class TestFitImaging:
                 == fit.unmasked_blurred_image_of_planes[1]
             ).all()
 
-            unmasked_blurred_image_of_galaxies = tracer.unmasked_blurred_image_of_planes_and_galaxies_from_grid_and_psf(
-                grid=masked_imaging_7x7.grid, psf=masked_imaging_7x7.psf
+            unmasked_blurred_image_of_galaxies = (
+                tracer.unmasked_blurred_image_of_planes_and_galaxies_from_grid_and_psf(
+                    grid=masked_imaging_7x7.grid, psf=masked_imaging_7x7.psf
+                )
             )
 
             assert (
@@ -879,12 +887,14 @@ class TestFitImaging:
             mapper = pix.mapper_from_grid_and_sparse_grid(
                 grid=masked_imaging_7x7.grid_inversion, sparse_grid=None
             )
-            inversion = inversions.InversionImagingMatrix.from_data_mapper_and_regularization(
-                mapper=mapper,
-                regularization=reg,
-                image=masked_imaging_7x7.image,
-                noise_map=masked_imaging_7x7.noise_map,
-                convolver=masked_imaging_7x7.convolver,
+            inversion = (
+                inversions.InversionImagingMatrix.from_data_mapper_and_regularization(
+                    mapper=mapper,
+                    regularization=reg,
+                    image=masked_imaging_7x7.image,
+                    noise_map=masked_imaging_7x7.noise_map,
+                    convolver=masked_imaging_7x7.convolver,
+                )
             )
 
             assert inversion.mapped_reconstructed_image.in_2d == pytest.approx(
[…]
