+++ test_autolens/plot/arrays/array_and_grid.py
@@ -10,16 +10,20 @@ imaging = instrument_util.load_test_imaging(
 array = imaging.image
 
 plotter = aplt.Plotter(
-    figure=aplt.Figure(figsize=(10,10)),
-    cmap=aplt.ColorMap(cmap="gray", norm="symmetric_log", norm_min=-0.13, norm_max=20, linthresh=0.02),
-    grid_scatterer=aplt.GridScatterer(marker="+", colors='cyan', size=450)
+    figure=aplt.Figure(figsize=(10, 10)),
+    cmap=aplt.ColorMap(
+        cmap="gray", norm="symmetric_log", norm_min=-0.13, norm_max=20, linthresh=0.02
+    ),
+    grid_scatterer=aplt.GridScatterer(marker="+", colors="cyan", size=450),
 )
 
 grid = al.GridIrregular(grid=[[1.0, 1.0], [2.0, 2.0], [3.0, 3.0]])
 
 print(grid)
 
-vector_field = al.VectorFieldIrregular(vectors=[(1.0, 2.0), (2.0, 1.0)], grid=[(-1.0, 0.0), (-2.0, 0.0)])
+vector_field = al.VectorFieldIrregular(
+    vectors=[(1.0, 2.0), (2.0, 1.0)], grid=[(-1.0, 0.0), (-2.0, 0.0)]
+)
 
 aplt.Array(
     array=array.in_2d,
@@ -27,4 +31,5 @@ aplt.Array(
     positions=al.GridIrregularGrouped([(0.0, 1.0), (0.0, 2.0)]),
     vector_field=vector_field,
     patches=vector_field.elliptical_patches,
-    plotter=plotter)
+    plotter=plotter,
+)
