+++ README.rst
@@ -11,7 +11,7 @@ Installation
 ------------
 
 **PyAutoLens** requires Python 3.6+ and you can install it via ``pip`` or ``conda`` (see
-`this link <https://pyautolens.readthedocs.io/en/latest/general/installation.html#installation-with-conda>`_
+`this link <https://pyautolens.readthedocs.io/en/latest/installation/conda.html>`_
 for ``conda`` instructions).
 
 .. code-block:: bash
