+++ test_autolens/config/general.ini
@@ -28,4 +28,5 @@ stochastic_histogram_samples = 2
 stochastic_histogram_bins = 2
 
 [hyper]
-hyper_minimum_percent = 0.01
+hyper_minimum_percent=0.01
+hyper_noise_limit=1.0e8
