+++ howtolens/chapter_2_lens_modeling/scripts/tutorial_7_masking_and_positions.py
@@ -126,7 +126,7 @@ always aim to use masks as big as possible which still give a reasonable run-spe
 code running fast - but it could lead you to infer an incorrect lens model!
 
 If you are fitting the foreground lens `Galaxy`'s light you pretty much have no choice but to use a large circular 
-mask anyway, as you`ll need to capture the lens`s extended emission. Chances are this will encompass the entire 
+mask anyway, as you'll need to capture the lens`s extended emission. Chances are this will encompass the entire 
 source galaxy.
 """
 
@@ -294,7 +294,7 @@ And that completes our final tutorial in this chapter! At this point, I recommen
 `autolens_workspace/preprocess section of the workspace. This includes tools which allow you create custom masks and 
 positions for a specific strong lens and output them so they can be loaded before an analysis.
 
-When we cover pipelines next, you`ll see that pipelines allow us to use a custom mask and set of positions for each 
+When we cover pipelines next, you'll see that pipelines allow us to use a custom mask and set of positions for each 
 lens we model. So, although we have to draw the masks and positions for each lens in a sample, once we've done that 
 we can fit all lenses with one standardized pipeline!
 
@@ -302,13 +302,13 @@ There are two things you should bare in mind in terms of masking and positions:
 
  1) Customizing the mask and positions for the analysis of one strong lens gets the analysis running fast and can 
  provide accurate non-linear sampling. However, for a large sample of lenses, customizing the mask and positions 
- will begin to take a lot of time. If you`re willing to put that time and effort in, great, but these solutions 
+ will begin to take a lot of time. If you're willing to put that time and effort in, great, but these solutions 
  *do not* scale-up to large samples of lenses.
     
  2) A word of warning - be *extremely* careful when using positions, especially if it is unclear if the lensed 
- source galaxy has one or multiple source`s of light. If your position threshold is small and the positions you 
+ source galaxy has one or multiple source's of light. If your position threshold is small and the positions you 
  give the analysis correspond to different parts of the source, you may remove the *correct lens model*. In my 
- experience, as long as you keep the threshold above ~0.5" you`ll be fine.
+ experience, as long as you keep the threshold above ~0.5" you'll be fine.
 
 And with that, we've completed the chapter.
 """
