+++ howtolens/chapter_4_inversions/scripts/tutorial_5_borders.py
@@ -117,7 +117,7 @@ def perform_fit_with_source_galaxy_mask_and_border(
 Okay, so lets first look at our `Mapper` without using a border using our annular `Mask2D`.
 
 First, note how we set up the border. We use a `SettingsPixelization` object, which is analogous to the 
-`SettingsMaskedImaging` and `SettingsLens` objects we used in previous tutorials. Later, you`ll see how these 
+`SettingsMaskedImaging` and `SettingsLens` objects we used in previous tutorials. Later, you'll see how these 
 settings can also be passed to a `SettingsPhaseImaging` object, to control the behaviour of the `Pixelization` during a
 model-fit.
 """
@@ -270,9 +270,9 @@ aplt.Mapper.subplot_image_and_mapper(
 This successfully addresses both of the issues above! However, you might be thinking, isn't that a bit of a hack? Its 
 not really a physical treatment of the ray-tracing, is it?
 
-Well, you`re right. However, the *only* physical way to do this would be to use a `Mask2D` so large that all demangified 
+Well, you're right. However, the *only* physical way to do this would be to use a `Mask2D` so large that all demangified 
 central pixels are surrounded by traced image-pixels. This would require a `Mask2D` so large our computer would crash, 
-That`s not a good solution, thus borders provide us with a workaround, one that I`ve extensively tested and have found 
+That`s not a good solution, thus borders provide us with a workaround, one that I've extensively tested and have found 
 that, provided your `Mask2D` isn't too small, doesn`t lead to systematic biases.
 
 Next, I'm going to quickly highlight how important borders are when modeling multiple lens galaxies. Their complex 
@@ -454,8 +454,8 @@ aplt.Inversion.reconstruction(
 
 # %%
 """
-And with that, borders are done. In truth, borders should pretty much take care of themselves when you`re using 
-__PyAutoLens__ and you probably won't think about them much. However, as I showed above, if you don't choose a large enough 
+And with that, borders are done. In truth, borders should pretty much take care of themselves when you're using 
+**PyAutoLens** and you probably won't think about them much. However, as I showed above, if you don't choose a large enough 
 mask things can go wrong - thus, its important you know what borders are, so you can look out for this potential 
 source of systematics!
 """
