+++ howtolens/chapter_1_introduction/scripts/tutorial_3_galaxies.py
@@ -226,7 +226,7 @@ aplt.Galaxy.deflections_x(galaxy=galaxy_with_many_profiles, grid=grid)
 
 # %%
 """
-And we`re done. Lets finished by just thinking about one question:
+And we're done. Lets finished by just thinking about one question:
 
  1) We've learnt we can group profiles into `Galaxy`'s to essentially sum the contribution of each `LightProfile` to 
  the galaxy image's intensity, or sum the contribution of each `MassProfile` to the convergence, potential and 
