+++ autolens/__init__.py
@@ -86,4 +86,4 @@ from autoconf import conf
 
 conf.instance.register(__file__)
 
-__version__ = '1.8.2'
+__version__ = '1.8.3'
