+++ requirements.txt
@@ -6,4 +6,4 @@ astropy>=3.0.0
 numpydoc>=1.0.0
 matplotlib>=3.0.3
 numpy>=1.17.2
-pyprojroot==0.2.0
+pyprojroot==0.2.0 
