### 2020-11-28 51b30f5
All fit.py tests passing
