+++ test_autolens/unit/conftest.py
@@ -98,19 +98,19 @@ def make_masked_imaging_7x7():
     return fixtures.make_masked_imaging_7x7()
 
 
-@pytest.fixture(name="visibilities_mask_7x2")
-def make_visibilities_mask_7x2():
-    return fixtures.make_visibilities_mask_7x2()
+@pytest.fixture(name="visibilities_mask_7")
+def make_visibilities_mask_7():
+    return fixtures.make_visibilities_mask_7()
 
 
-@pytest.fixture(name="visibilities_7x2")
+@pytest.fixture(name="visibilities_7")
 def make_visibilities_7():
     return fixtures.make_visibilities_7()
 
 
-@pytest.fixture(name="noise_map_7x2")
+@pytest.fixture(name="visibilities_noise_map_7")
 def make_noise_map_7():
-    return fixtures.make_noise_map_7()
+    return fixtures.make_visibilities_noise_map_7()
 
 
 @pytest.fixture(name="transformer_7x7_7")
