### 2020-11-28 18fd73c
Refactored for complex visiblities, all tests pass
