+++ test_autolens/unit/fit/test_fit.py
@@ -1666,7 +1666,7 @@ class TestFitInterferometer:
                 uv_wavelengths=uv_wavelengths,
             )
 
-            interferometer.visibilities[0] = 5.0+4.0j
+            interferometer.visibilities[0] = 5.0 + 4.0j
 
             visibilities_mask = np.full(fill_value=False, shape=(1,))
 
@@ -1703,12 +1703,12 @@ class TestFitInterferometer:
 
             assert (fit.visibilities_mask == np.array([False, False])).all()
 
-            assert (fit.visibilities.in_1d == np.array([5.0+4.0J])).all()
-            assert (fit.noise_map.in_1d == np.array([1.0+1.0J])).all()
-            assert (fit.model_visibilities.in_1d == np.array([2.0+0.0j])).all()
-            assert (fit.residual_map.in_1d == np.array([3.0+4.0j])).all()
-            assert (fit.normalized_residual_map.in_1d == np.array([3.0+4.0j])).all()
-            assert (fit.chi_squared_map.in_1d == np.array([9.0+16.0j])).all()
+            assert (fit.visibilities.in_1d == np.array([5.0 + 4.0j])).all()
+            assert (fit.noise_map.in_1d == np.array([1.0 + 1.0j])).all()
+            assert (fit.model_visibilities.in_1d == np.array([2.0 + 0.0j])).all()
+            assert (fit.residual_map.in_1d == np.array([3.0 + 4.0j])).all()
+            assert (fit.normalized_residual_map.in_1d == np.array([3.0 + 4.0j])).all()
+            assert (fit.chi_squared_map.in_1d == np.array([9.0 + 16.0j])).all()
 
             assert fit.chi_squared == 25.0
             assert fit.noise_normalization == (2.0 * np.log(2 * np.pi * 1.0 ** 2.0))
@@ -1773,11 +1773,10 @@ class TestFitInterferometer:
             assert (fit.visibilities_mask == np.array([False, False, False])).all()
 
             assert (
-                fit.visibilities.in_1d
-                == np.array([5.0+5.0j, 5.0+5.0j, 5.0+5.0j])
+                fit.visibilities.in_1d == np.array([5.0 + 5.0j, 5.0 + 5.0j, 5.0 + 5.0j])
             ).all()
             assert (
-                fit.noise_map.in_1d == np.array([2.0+2.0j, 2.0+2.0j, 2.0+2.0j])
+                fit.noise_map.in_1d == np.array([2.0 + 2.0j, 2.0 + 2.0j, 2.0 + 2.0j])
             ).all()
 
             assert fit.model_visibilities.in_1d == pytest.approx(
@@ -1787,15 +1786,15 @@ class TestFitInterferometer:
             # moddel visibilities are all [1.94805, 0.0]
 
             assert fit.residual_map.in_1d == pytest.approx(
-                np.array([3.0519+5.0j, 3.0519+5.0j, 3.0519+5.0j]), 1.0e-4
+                np.array([3.0519 + 5.0j, 3.0519 + 5.0j, 3.0519 + 5.0j]), 1.0e-4
             )
 
             assert fit.normalized_residual_map.in_1d == pytest.approx(
                 np.array(
                     [
-                        3.0519 / 2.0 + (5.0 / 2.0)*1.0j,
-                        3.0519 / 2.0 + (5.0 / 2.0)*1.0j,
-                        3.0519 / 2.0 + (5.0 / 2.0)*1.0j,
+                        3.0519 / 2.0 + (5.0 / 2.0) * 1.0j,
+                        3.0519 / 2.0 + (5.0 / 2.0) * 1.0j,
+                        3.0519 / 2.0 + (5.0 / 2.0) * 1.0j,
                     ]
                 ),
                 1.0e-4,
[…]
