+++ autolens/pipeline/phase/dataset/analysis.py
@@ -11,7 +11,5 @@ class Analysis:
             galaxies=instance.galaxies, cosmology=self.cosmology
         )
 
-    def stochastic_log_evidences_for_instance(
-        self, instance, samples=100,
-    ):
+    def stochastic_log_evidences_for_instance(self, instance, samples=100):
         raise NotImplementedError()
