### 2020-11-29 5b2a069
stochastic settings added
