+++ autolens/lens/settings.py
@@ -11,13 +11,15 @@ class SettingsLens:
         positions_threshold=None,
         auto_positions_factor=None,
         auto_positions_minimum_threshold=None,
-        stochastic_samples : int = 250,
+        stochastic_samples: int = 250,
+        stochastic_histogram_bins: int = 10,
     ):
 
         self.positions_threshold = positions_threshold
         self.auto_positions_factor = auto_positions_factor
         self.auto_positions_minimum_threshold = auto_positions_minimum_threshold
         self.stochastic_samples = stochastic_samples
+        self.stochastic_histogram_bins = stochastic_histogram_bins
 
     @property
     def tag(self):
