+++ autolens/lens/ray_tracing.py
@@ -666,9 +666,7 @@ class AbstractTracerData(AbstractTracerLensing, ABC):
         )
 
     def hyper_noise_map_from_noise_map(self, noise_map):
-        return sum(self.hyper_noise_maps_of_planes_from_noise_map(
-            noise_map=noise_map
-        ))
+        return sum(self.hyper_noise_maps_of_planes_from_noise_map(noise_map=noise_map))
 
     def hyper_noise_maps_of_planes_from_noise_map(self, noise_map):
         return [
