+++ requirements.txt
@@ -1,6 +1,6 @@
 autoconf==0.7.3
-autofit==0.70.3
-autoarray==0.16.0
+autofit==0.70.4
+autoarray==0.16.1
 autogalaxy==0.16.4
 astropy>=3.0.0
 numpydoc>=1.0.0
