+++ requirements.txt
@@ -1,7 +1,7 @@
 autoconf==0.7.3
 autofit==0.70.4
 autoarray==0.16.1
-autogalaxy==0.16.4
+autogalaxy==0.16.5
 astropy>=3.0.0
 numpydoc>=1.0.0
 matplotlib>=3.0.3
