### 2020-12-06 #2825 @Jammy2211
> Dear @Jammy2211 , I have some comments about the paper:
>
> 1. This phrase in not completely accurate:
>    "Strong gravitational lensing, where a background source galaxy appears multiple times due to its light rays being bent by the mass of one or more foreground lens galaxies"
>    Sometimes we do not get multiple images, but just a deformed image of the background source.

Oops. I've rephrased this to `which cam make a background source galaxy appear multiple times'.

> 2. "When two galaxies are aligned down the line-of-sight to Earth, light rays from the background galaxy are bent by the intervening mass of one or more foreground galaxies"
>    Maybe change to:
>    "When two galaxies are aligned down the line-of-sight to Earth, light rays from the background galaxy are bent by the
intervening mass of the foreground galaxy"

I don't want reads to think PyAutoLens only models single galaxy-galaxy strong lens system, as the method can be used to model systems with multiple foreground galaxies (and clusters of galaxies). I agree your suggested text is cleaner, but AutoLens has in the past been a galaxy-galaxy only code so I'm trying to shake that reputation!

> 3. "Its light can be fully bent around the foreground galaxies, traversing multiple paths to the Earth, meaning that the background galaxy is observed multiple times."
>    Maybe change to:
>    "Sometimes its light can be fully bent, traversing multiple paths to the Earth, meaning that the background galaxy is observed multiple times."

Done.

> 4. An error here:
>    "Strong lensing is a competitive test of cosmological models, for example the **exapnsion** rate of the Universe."

Corrected.

> 5. I would start a new paragraph here:
>    The past decade has ...

Good shout.

> 6. missing an **of** here:
>    "In the next decade of order **of** "

Done.

> 7. remove the **a** here (is plural?):
>    "By passing these objects **a** Grid strong lens sightlines are computed"

You only pass them a single Grid object, so it is singular (I think)

> 8. it is not neccesary to indicate is python only here: "ensuring PyAutoLens’s Python-only code base runs efficiently." Also is not plural
>    "ensuring PyAutoLens runs efficiently."

Corrected

> 9. and/or is extensively use, but I would not recommend it, just place and:
>    "which uses positions and flux-ratios"

Done.

> 10. Missing a coma here:
>     "about simpler lens models fitted in earlier phases to subsequent phases**,** which fit progressively more complex models"

Added.

> 11. singular phrase here (not plural):
>     "PyAutoFit’s aggregator tool "

updated text.

> 12. hard-disc should be **hard-disk**

Corrected.

> 13. Suggestion
>     "This uses memory-light Python generators, ensuring it is practical for fits to thousands of lenses."
>     to
>     "This uses memory-light Python generators, ensuring it is practical for thousands of lenses."
>
> That's for now

Done, cheers.

Thanks !
