### 2020-12-06 #2825 @Jammy2211
> @Jammy2211 I'll collect feedback on the repository here as I go through the review:
>
> * [ ]  I noticed that while there are a lot of tagged releases on PyPI, on Github there are only very old tags. Do you happen to have `git` tags available locally that you can push? otherwise, what about tagging on Github just the last 3/4 releases? It is convenient to be able to browse older versions of the code without downloading them from PyPI.

Oh wow, those tagged releases are _old_. I think we disabled a lot of the release tagging early in development for reasons I now don't remember, but was probably very bad practise (I was learning Python along the way).

We don't have old releases as far as I can tell. I'll speak to my codeveloper about this, I'll sort out the GItHub tagging and if its possible I could dig up some old PyPI releases and try to retro-tag them?

> * [ ]  **statement of need**: I think it would be useful to add at the very top of the docs https://pyautolens.readthedocs.io/en/latest/ a quick paragraph that explains what `pyautolens` is and what is the target audience, before digging into what is Strong lensing

Done.

> * [ ]  **automated tests**: I see you have a test suite, what about configuring Github actions to run that at every commit? https://docs.github.com/en/free-pro-team@latest/actions/guides/building-and-testing-python

We had automated testing tools in the past, but found them prohibitive. A lot of the students that contribute to the project are not familiar with unit testing and these tools often lead to more confusion then it was worth. We found simply being more careful via the PR process (and introducing them to unit tests at this point of entry) more effective.

> * [ ]  **community guidelines**: why do you ask people to open issues on the workspace repo instead of the pyautolens one?

I am hoping to keep installation issues (autolens workspace) separate from development issues (PyAutoLens). If installation issues are kept in one place it should be easier to refer back to them when necessary.

> * [ ]  **related packages**: there is a section in the paper that mentions related packages, it would be useful to had a couple of paragraphs explaining the relationship / differences / commonalities at high level between those packages and `pyautolens`.

I would prefer to avoid such a paragraph, as both `PyAutoLens` and `lenstronomy` are in active development and any paragraph I write would quickly become outdated. I think its better that a reader can simply follow the links to the webpages / GitHub pages of the packages so they can gauge for themselves the difference between the packages at the time of viewing.

I am also trying to avoid the political issue in that we are at some level competing projects, and it would be difficult for me to write such a paragraph and it not come across as biased or misrepresentative.

If you really think the paragraph is warranted despite this let me know and I'll try and write something which is timeless and agnostic, but there’s a good chance it’ll end up not really saying much at all!

> * [ ]  **binder?**: what about setting up `binder` so people can go through the lessons without installing anything locally? https://mybinder.readthedocs.io/en/latest/
>

Good shout. We've looked at Binder in the past, but always decided we didn't want another deployment system / server to manage, especially during the period a lot of the documentation was coming together. I think we're at a point in development now where this wouldn't be too prohibitive, so I'll get to work on setting it up.

We felt the same about Docker, and I think we're now at a point where Docker support is worth our time before we put this paper out there for the world!

> Ok, I've done a first pass, overall great software package, I'm really impressed by the documentation, in particular the tutorials in notebook format, just some feedback here and a couple of issues in the repository. In the next week I'll read the paper more carefully and give specific feedback on that, but it already looks very good, so it's going to be minor things.

Thank you for the kind works, a lot time and thought went into the tutorials so its means a lot to hear they came out alright in the end!
