### 2020-12-06 #2825 @Jammy2211
> @Jammy2211 @xuanxu ok, my review is complete, if you address the 2 issues I opened, the repository feedback above and the suggestions on the paper below, we are done (also justifying why you don't want or can't to implement something I suggested is also valid, of course).
>
> ## Feedback on paper
> ### Summary
> * I don't think we need to mention the names of the classes in the summary, better keep it more general, just mention they functionality

Agreed and edited.

> * HowToLens "lectures" suggests they are videos maybe? what about calling them "tutorial notebooks" or just specify "HowToLens lectures in Jupyter Notebook format"

Gone with your latter suggestion - it is more clear I agree.

>
> ### Background
> * can you explain a bit how you go from left panel to center panel in Fig 1? Are you removing the foreground galaxy beforehand or you are fitting simultaneously for the foreground galaxy and the lens and then plotting only the lens?

Ive added a sentence clarifying that the lens light is simutlaneously fitted for and subtracted.

> * when you mention all the applications, maybe a bullet list would make it more readable?
>

I prefer it as a paragraph, stylistically this is more how I'm used to reading an introduction.

> ### Software API and Features
> * can you clarify what you mean by "abstracted objects"? I wonder if a programmer could think you are referring to abstract classes

I agree the term abstracted objects isnt clear... I've gone for "A gravitational lens system can be quickly assembled from Python object which provide abstract data representations of the different components of a strong lens."

> * "By passing these objects a Grid ... are computed" what is computing this? is the `Tracer`? better specify

Any of the objects can take a grid - I've updated the sentence too "By passing any of these objects a `Grid` strong lens quantities
can be computed, including multi-plane ray-tracing sightlines [@McCully2014].".

> * Optimized by `numba`: is the code multi-threaded or not?

It is not currently, we simply use numba for the low level compilation.

> * As you mention performance, are there particular memory constraints? does the code use a lot of RAM for standard runs? No need to be too specific, but at least give the idea of what I can run easily on my 4 GB RAM laptop and when I need a HPC node

I have added a PErformance section:

"The analysis of direct imaging datasets and interferometer datasets (up to of order 1 million visibilities) are both
feasible on hardware with at least 4GB of RAM. The time it takes to perform lens modeling with `PyAutoLens` are
highly variable and depend on the size of the dataset being analysed and complexity of the model being fitted. They can
vary from minutes to thousands of CPU hours. For large jobs we recommend users install `PyAutoLens` on a HPC cluster and
documentation is included on how to set this up."

> * can you specify what is a `uv-plane`?

Its quite a specifc term in radio astronomy that describes how interfermeter data is taken. I have changed the first uv-plane with 'Fourier space' but kept the second... I don't think its really a term I can specify in anymore detail in the text without dedicating multiple setences explaning how interferometer work unfortunately!

> * I would reserve `backticks` formatting just for names of classes or software, if you want to highlight works maybe use _corsive_ or **bold**?
>

I agree, anything I did that was a typo so i'll correct them now!

> ### Workspace and HowToLens Tutorials
> * In case you decide to add binder support (see above) you can add it here

Will do!
