### 2020-12-07 #2825 @zonca
@Jammy2211 thanks for addressing the feedback, please push your changes to the repository or in a pull request.

My only followup is:

> > @Jammy2211 @xuanxu ok, my review is complete, if you address the 2 issues I opened, the repository feedback above and the suggestions on the paper below, we are done (also justifying why you don't want or can't to implement something I suggested is also valid, of course).
> > ## Feedback on paper
>
> > * Optimized by `numba`: is the code multi-threaded or not?
>
> It is not currently, we simply use numba for the low level compilation.

ok, I think it is useful to specify in the performance section if your code is threaded in some sections (even just using the MKL version of numpy that automatically threads) or if it is fully single threaded.

Also, in the performance section you can give a specific example like: "analyzing 1 million visibilities of 1 specific instrument on a recent laptop with 4 GB of ram takes ~3 hours".
