### 2020-12-07 #2825 @zonca
## Feedback on the repository

ok, I deleted the solved issues, here just the ones I am following up on:

> > @Jammy2211 I'll collect feedback on the repository here as I go through the review:
> >
> > * [ ]   I noticed that while there are a lot of tagged releases on PyPI, on Github there are only very old tags. Do you happen to have `git` tags available locally that you can push? otherwise, what about tagging on Github just the last 3/4 releases? It is convenient to be able to browse older versions of the code without downloading them from PyPI.
>
> Oh wow, those tagged releases are _old_. I think we disabled a lot of the release tagging early in development for reasons I now don't remember, but was probably very bad practise (I was learning Python along the way).
>
> We don't have old releases as far as I can tell. I'll speak to my codeveloper about this, I'll sort out the GItHub tagging and if its possible I could dig up some old PyPI releases and try to retro-tag them?

yes, it mostly useful just to tag the last 3 or 4. Or if you know most of your users always use the last version, even just tag the last one and keep tagging it as you publish newer releases.

> > * [ ]   **automated tests**: I see you have a test suite, what about configuring Github actions to run that at every commit? https://docs.github.com/en/free-pro-team@latest/actions/guides/building-and-testing-python
>
> We had automated testing tools in the past, but found them prohibitive. A lot of the students that contribute to the project are not familiar with unit testing and these tools often lead to more confusion then it was worth. We found simply being more careful via the PR process (and introducing them to unit tests at this point of entry) more effective.

It is really easy to setup a Github action to run the test suite, otherwise you find yourself in the situation that your unit tests are failing on master, see for example: https://github.com/Jammy2211/PyAutoLens/issues/146
I believe this is extremely important to increase reliability of the software.
