+++ test_autolens/integration/tests/imaging/lens__source/lens_mass__source_x2__hyper.py
@@ -33,17 +33,17 @@ def make_pipeline(name, path_prefix):
         search=search,
     )
 
-    phase2 = phase2.extend_with_multiple_hyper_phases(hyper_galaxies_search=True)
+    phase2 = phase2.extend_with_hyper_phase(hyper_galaxies_search=True)
 
     class HyperLensSourcePlanePhase(al.PhaseImaging):
         def customize_priors(self, results):
 
             self.galaxies.source_0.hyper_galaxy = (
-                results.last.hyper_combined.instance.galaxies.source_0.hyper_galaxy
+                results.last.hyper.instance.galaxies.source_0.hyper_galaxy
             )
 
             self.galaxies.source_1.hyper_galaxy = (
-                results.last.hyper_combined.instance.galaxies.source_1.hyper_galaxy
+                results.last.hyper.instance.galaxies.source_1.hyper_galaxy
             )
 
     phase3 = HyperLensSourcePlanePhase(
@@ -55,12 +55,12 @@ def make_pipeline(name, path_prefix):
             source_0=al.GalaxyModel(
                 redshift=1.0,
                 light=phase2.result.model.galaxies.source_0.setup_light,
-                hyper_galaxy=phase2.result.hyper_combined.instance.galaxies.source_0.hyper_galaxy,
+                hyper_galaxy=phase2.result.hyper.instance.galaxies.source_0.hyper_galaxy,
             ),
             source_1=al.GalaxyModel(
                 redshift=1.0,
                 light=phase2.result.model.galaxies.source_1.setup_light,
-                hyper_galaxy=phase2.result.hyper_combined.instance.galaxies.source_1.hyper_galaxy,
+                hyper_galaxy=phase2.result.hyper.instance.galaxies.source_1.hyper_galaxy,
             ),
         ),
         search=search,

[…]
