+++ test_autolens/unit/fit/test_fit.py
@@ -808,6 +808,7 @@ class TestFitImaging:
                 tracer=tracer,
                 hyper_image_sky=hyper_image_sky,
                 hyper_background_noise=hyper_background_noise,
+                use_hyper_scaling=True,
             )
 
             hyper_noise_map_background = hyper_background_noise.hyper_noise_map_from_noise_map(
@@ -862,6 +863,17 @@ class TestFitImaging:
 
             assert log_likelihood == pytest.approx(fit.log_likelihood, 1e-4)
             assert log_likelihood == fit.figure_of_merit
+
+            fit = al.FitImaging(
+                masked_imaging=masked_imaging_7x7,
+                tracer=tracer,
+                hyper_image_sky=hyper_image_sky,
+                hyper_background_noise=hyper_background_noise,
+                use_hyper_scaling=False,
+            )
+
+            assert fit.image == pytest.approx(masked_imaging_7x7.image, 1.0e-4)
+            assert fit.noise_map == pytest.approx(masked_imaging_7x7.noise_map, 1.0e-4)
 
         def test___blurred_and_model_images_of_planes_and_unmasked_blurred_image_properties(
             self, masked_imaging_7x7
@@ -2053,6 +2065,18 @@ class TestFitInterferometer:
             )
 
             assert hyper_noise_map.in_1d == pytest.approx(fit.noise_map.in_1d)
+
+            fit = al.FitInterferometer(
+                masked_interferometer=masked_interferometer_7,
+                tracer=tracer,
+                hyper_background_noise=hyper_background_noise,
+                use_hyper_scaling=False,
+            )
+
+            assert fit.noise_map == pytest.approx(
+                masked_interferometer_7.noise_map, 1.0e-4
+            )
+            assert fit.noise_map != pytest.approx(hyper_noise_map.in_1d, 1.0e-4)
 
     class TestCompareToManualInversionOnly:
         def test___all_lens_fit_quantities__no_hyper_methods(
