+++ autolens/fit/fit.py
@@ -13,6 +13,7 @@ class FitImaging(aa_fit.FitImaging):
         tracer,
         hyper_image_sky=None,
         hyper_background_noise=None,
+        use_hyper_scaling=True,
         settings_pixelization=pix.SettingsPixelization(),
         settings_inversion=inv.SettingsInversion(),
     ):
@@ -29,25 +30,32 @@ class FitImaging(aa_fit.FitImaging):
 
         self.tracer = tracer
 
-        image = hyper_image_from_image_and_hyper_image_sky(
-            image=masked_imaging.image, hyper_image_sky=hyper_image_sky
-        )
+        if use_hyper_scaling:
 
-        noise_map = hyper_noise_map_from_noise_map_tracer_and_hyper_background_noise(
-            noise_map=masked_imaging.noise_map,
-            tracer=tracer,
-            hyper_background_noise=hyper_background_noise,
-        )
-
-        if (
-            tracer.has_hyper_galaxy
-            or hyper_image_sky is not None
-            or hyper_background_noise is not None
-        ):
-
-            masked_imaging = masked_imaging.modify_image_and_noise_map(
-                image=image, noise_map=noise_map
+            image = hyper_image_from_image_and_hyper_image_sky(
+                image=masked_imaging.image, hyper_image_sky=hyper_image_sky
             )
+
+            noise_map = hyper_noise_map_from_noise_map_tracer_and_hyper_background_noise(
+                noise_map=masked_imaging.noise_map,
+                tracer=tracer,
+                hyper_background_noise=hyper_background_noise,
+            )
+
+            if (
+                tracer.has_hyper_galaxy
+                or hyper_image_sky is not None
+                or hyper_background_noise is not None
+            ):
+
+                masked_imaging = masked_imaging.modify_image_and_noise_map(
+                    image=image, noise_map=noise_map
+                )
+
+        else:
+
+            image = masked_imaging.image
+            noise_map = masked_imaging.noise_map
 
         self.blurred_image = tracer.blurred_image_from_grid_and_convolver(
             grid=masked_imaging.grid,
@@ -157,6 +165,7 @@ class FitInterferometer(aa_fit.FitInterferometer):
         masked_interferometer,
         tracer,
         hyper_background_noise=None,
+        use_hyper_scaling=True,
         settings_pixelization=pix.SettingsPixelization(),
         settings_inversion=inv.SettingsInversion(),
     ):
@@ -171,18 +180,24 @@ class FitInterferometer(aa_fit.FitInterferometer):
             A function which maps the 1D lens hyper_galaxies to its unmasked 2D arrays.
         """
 
-        if hyper_background_noise is not None:
-            noise_map = hyper_background_noise.hyper_noise_map_from_complex_noise_map(
-                noise_map=masked_interferometer.noise_map
-            )
+        if use_hyper_scaling:
+
+            if hyper_background_noise is not None:
+                noise_map = hyper_background_noise.hyper_noise_map_from_complex_noise_map(
+                    noise_map=masked_interferometer.noise_map
+                )
+            else:
+                noise_map = masked_interferometer.noise_map
+
+            if hyper_background_noise is not None:
+
+                masked_interferometer = masked_interferometer.modify_noise_map(
+                    noise_map=noise_map
+                )
+
         else:
+
             noise_map = masked_interferometer.noise_map
-
-        if hyper_background_noise is not None:
-
-            masked_interferometer = masked_interferometer.modify_noise_map(
-                noise_map=noise_map
-            )
 
         self.tracer = tracer
 

[…]
