+++ test_autolens/config/general.ini
@@ -31,4 +31,4 @@ stochastic_histogram_bins = 2
 hyper_minimum_percent=0.01
 hyper_noise_limit=1.0e8
 stochastic_outputs=False
-backwards_compatible_tag=False
+rename_hyper_combined=False
