+++ test_autolens/integration/tests/imaging/lens__source_inversion/adaptive_magnification/lens_mass__source.py
@@ -1,59 +1,0 @@
-import autofit as af
-import autolens as al
-from test_autolens.integration.tests.imaging import runner
-
-test_type = "lens__source_inversion"
-test_name = "lens_mass__source_adaptive_magnification"
-dataset_name = "mass_sie__source_sersic"
-instrument = "euclid"
-
-
-def make_pipeline(name, path_prefix):
-
-    phase1 = al.PhaseImaging(
-        name="phase[1]",
-        galaxies=dict(
-            lens=al.GalaxyModel(redshift=0.5, mass=al.mp.EllipticalIsothermal),
-            source=al.GalaxyModel(
-                redshift=1.0,
-                pixelization=al.pix.VoronoiMagnification,
-                regularization=al.reg.Constant,
-            ),
-        ),
-        search=search,
-    )
-
-    phase1.search.const_efficiency_mode = True
-    phase1.search.n_live_points = 60
-    phase1.search.facc = 0.8
-
-    phase1 = phase1.extend_with_inversion_phase()
-
-    phase2 = al.PhaseImaging(
-        name="phase[2]",
-        galaxies=dict(
-            lens=al.GalaxyModel(
-                redshift=0.5, mass=phase1.result.model.galaxies.lens.mass
-            ),
-            source=al.GalaxyModel(
-                redshift=1.0,
-                pixelization=phase1.result.instance.galaxies.source.pixelization,
-                regularization=phase1.result.instance.galaxies.source.regularization,
-            ),
-        ),
-        search=search,
-    )
-
-    phase2.search.const_efficiency_mode = True
-    phase2.search.n_live_points = 60
-    phase2.search.facc = 0.8
-
-    phase2 = phase2.extend_with_inversion_phase()
-
-    return al.PipelineDataset(name, path_prefix, phase1, phase2)
-
-
-if __name__ == "__main__":
-    import sys
-
-    runner.run(sys.modules[__name__])
