+++ requirements.txt
@@ -1,7 +1,7 @@
 autoconf==0.7.3
-autofit == 0.71.2
-autoarray==0.16.3
-autogalaxy==0.16.7
+autofit == 0.71.3
+autoarray==0.16.4
+autogalaxy==0.16.8
 astropy>=3.0.0
 numpydoc>=1.0.0
 matplotlib>=3.0.3
