+++ autolens/__init__.py
@@ -79,4 +79,4 @@ from autoconf import conf
 
 conf.instance.register(__file__)
 
-__version__ = "1.9.2"
+__version__ = '1.9.3'
