### 2020-12-23 e6c3ca2
Changed image source link in docs
