+++ docs/index.rst
@@ -20,7 +20,7 @@ imaged source galaxy whose light has been distorted into an `Einstein ring'. The
 reconstructions of the source's lensed and unlensed light distributions, which are created using a model of the lens
 galaxy's mass to trace backwards how the source's light is gravitationally lensed.
 
-.. image:: https://github.com/Jammy2211/PyAutoLens/blob/development/imageaxis.png
+.. image:: https://raw.githubusercontent.com/Jammy2211/PyAutoLens/development/imageaxis.png
   :width: 1100
   :alt: Alternative text
 
