### 2020-12-23 aaaa9a5
Fixed broken installation link
