### 2020-12-24 7bb60d1
Plotter refactor now works in autolens
