+++ autolens/plot/__init__.py
@@ -13,8 +13,6 @@ from autoarray.plot.mat_wrap.mat_base import (
     Output,
 )
 
-from autoarray.plot.mat_wrap.include import Include
-
 from autoarray.plot.mat_wrap.mat_structure import (
     ArrayOverlay,
     GridScatter,
@@ -45,14 +43,14 @@ from autogalaxy.plot.mat_wrap.lensing_mat_obj import (
     CausticsLine,
 )
 
-from autogalaxy.plot.mat_wrap.lensing_include import Include
+from autogalaxy.plot.plotter.lensing_include import Include
 
-from autogalaxy.plot.mat_wrap.lensing_plotters import Plotter, SubPlotter
+from autogalaxy.plot.plotter.lensing_plotter import Plotter, SubPlotter
 
-from autogalaxy.plot.mat_wrap.lensing_plotters import plot_array as Array
-from autogalaxy.plot.mat_wrap.lensing_plotters import plot_grid as Grid
-from autogalaxy.plot.mat_wrap.lensing_plotters import plot_line as Line
-from autogalaxy.plot.mat_wrap.lensing_plotters import plot_mapper_obj as MapperObj
+from autogalaxy.plot.plotter.lensing_plotter import plot_array as Array
+from autogalaxy.plot.plotter.lensing_plotter import plot_grid as Grid
+from autogalaxy.plot.plotter.lensing_plotter import plot_line as Line
+from autogalaxy.plot.plotter.lensing_plotter import plot_mapper_obj as MapperObj
 
 from autogalaxy.plot.plots import light_profile_plots as LightProfile
 from autogalaxy.plot.plots import mass_profile_plots as MassProfile
