+++ autolens/pipeline/visualizer.py
@@ -8,7 +8,8 @@ import autofit as af
 import autoarray as aa
 from autoarray.plot.mat_wrap import mat_base
 from autogalaxy.plot.plots import fit_galaxy_plots, hyper_plots, inversion_plots
-from autogalaxy.plot.mat_wrap import lensing_plotters, lensing_include
+from autogalaxy.plot.plotter import lensing_plotter
+from autogalaxy.plot.plotter import lensing_include
 from autolens.plot.plots import (
     fit_interferometer_plots,
     ray_tracing_plots,
@@ -52,10 +53,10 @@ class AbstractVisualizer:
     @staticmethod
     def plotter_from_paths(paths: af.Paths, subfolders=None, format="png"):
         if subfolders is None:
-            return lensing_plotters.Plotter(
+            return lensing_plotter.Plotter(
                 output=mat_base.Output(path=paths.image_path, format=format)
             )
-        return lensing_plotters.Plotter(
+        return lensing_plotter.Plotter(
             output=mat_base.Output(
                 path=path.join(paths.image_path, subfolders), format=format
             )
@@ -64,10 +65,10 @@ class AbstractVisualizer:
     @staticmethod
     def sub_plotter_from_paths(paths: af.Paths, subfolders=None):
         if subfolders is None:
-            return lensing_plotters.SubPlotter(
+            return lensing_plotter.SubPlotter(
                 output=mat_base.Output(path=paths.image_path, format="png")
             )
-        return lensing_plotters.SubPlotter(
+        return lensing_plotter.SubPlotter(
             output=mat_base.Output(
                 path=path.join(paths.image_path, subfolders), format="png"
             )
