### 2020-12-24 ab5bfe8
refactoring of visualize
