+++ autolens/plot/plots/fit_interferometer_plots.py
@@ -6,65 +6,65 @@ from autogalaxy.plot.plotter import lensing_plotter
 
 
 @lensing_include.set_include
-@lensing_plotter.set_sub_plotter
+@lensing_plotter.set_plotter_for_subplot
 @plotter.set_subplot_filename
-def subplot_fit_interferometer(fit, include=None, sub_plotter=None):
+def subplot_fit_interferometer(fit, include=None, plotter=None):
 
     number_subplots = 6
 
-    sub_plotter.open_subplot_figure(number_subplots=number_subplots)
+    plotter.open_subplot_figure(number_subplots=number_subplots)
 
-    sub_plotter.setup_subplot(number_subplots=number_subplots, subplot_index=1)
-
-    ag.plot.FitInterferometer.residual_map_vs_uv_distances(
-        fit=fit, include=include, plotter=sub_plotter
-    )
-
-    sub_plotter.setup_subplot(number_subplots=number_subplots, subplot_index=2)
-
-    ag.plot.FitInterferometer.normalized_residual_map_vs_uv_distances(
-        fit=fit, include=include, plotter=sub_plotter
-    )
-
-    sub_plotter.setup_subplot(number_subplots=number_subplots, subplot_index=3)
-
-    ag.plot.FitInterferometer.chi_squared_map_vs_uv_distances(
-        fit=fit, include=include, plotter=sub_plotter
-    )
-
-    sub_plotter.setup_subplot(number_subplots=number_subplots, subplot_index=4)
+    plotter.setup_subplot(number_subplots=number_subplots, subplot_index=1)
 
     ag.plot.FitInterferometer.residual_map_vs_uv_distances(
-        fit=fit, plot_real=False, include=include, plotter=sub_plotter
+        fit=fit, include=include, plotter=plotter
     )
 
-    sub_plotter.setup_subplot(number_subplots=number_subplots, subplot_index=5)
+    plotter.setup_subplot(number_subplots=number_subplots, subplot_index=2)
 
     ag.plot.FitInterferometer.normalized_residual_map_vs_uv_distances(
-        fit=fit, plot_real=False, include=include, plotter=sub_plotter
+        fit=fit, include=include, plotter=plotter
     )
 
-    sub_plotter.setup_subplot(number_subplots=number_subplots, subplot_index=6)
+    plotter.setup_subplot(number_subplots=number_subplots, subplot_index=3)
 
     ag.plot.FitInterferometer.chi_squared_map_vs_uv_distances(
-        fit=fit, plot_real=False, include=include, plotter=sub_plotter
+        fit=fit, include=include, plotter=plotter
     )
 
-    sub_plotter.output.subplot_to_figure()
+    plotter.setup_subplot(number_subplots=number_subplots, subplot_index=4)
 
-    sub_plotter.figure.close()
+    ag.plot.FitInterferometer.residual_map_vs_uv_distances(
+        fit=fit, plot_real=False, include=include, plotter=plotter
+    )
+
+    plotter.setup_subplot(number_subplots=number_subplots, subplot_index=5)
+
+    ag.plot.FitInterferometer.normalized_residual_map_vs_uv_distances(
+        fit=fit, plot_real=False, include=include, plotter=plotter
+    )
+
+    plotter.setup_subplot(number_subplots=number_subplots, subplot_index=6)
+
+    ag.plot.FitInterferometer.chi_squared_map_vs_uv_distances(
+        fit=fit, plot_real=False, include=include, plotter=plotter
+    )
+
+    plotter.output.subplot_to_figure()
+
+    plotter.figure.close()
 
 
 @lensing_include.set_include
-@lensing_plotter.set_sub_plotter
+@lensing_plotter.set_plotter_for_subplot
 @plotter.set_subplot_filename
[…]
