+++ test_autolens/plot/lensing/all_lensing_objects_deflections.py
@@ -2,7 +2,7 @@ import autolens as al
 import autolens.plot as aplt
 
 plotter = aplt.Plotter()
-sub_plotter = aplt.SubPlotter()
+plotter = aplt.Plotter()
 
 grid = al.Grid.uniform(shape_2d=(100, 100), pixel_scales=0.05, sub_size=2)
 

[…]
