+++ test_autolens/plot/mappers/rectangular.py
@@ -35,8 +35,8 @@ mapper = rectangular.mapper_from_grid_and_sparse_grid(grid=source_plane_grid)
 # slightly less boring!
 aplt.MapperObj(
     mapper=mapper,
-    include=aplt.Include(inversion_grid=True, inversion_pixelization_grid=True),
-    plotter=aplt.Plotter(
+    include=aplt.Include2D(inversion_grid=True, mapper_source_pixelization_grid=True),
+    plotter=aplt.MatPlot2D(
         labels=aplt.Labels(title="Slightly less Boring Grid of Rectangular Pixels")
     ),
 )
