+++ howtolens/chapter_2_lens_modeling/scripts/tutorial_5_linking_phases.py
@@ -84,7 +84,10 @@ When plotted, the lens light`s is clearly visible in the centre of the image.
 """
 
 # %%
-aplt.Imaging.subplot_imaging(imaging=imaging, mask=mask)
+imaging_plotter = aplt.ImagingPlotter(
+    imaging=imaging, visuals_2d=aplt.Visuals2D(mask=mask)
+)
+imaging_plotter.subplot_imaging()
 
 # %%
 """
