+++ howtolens/chapter_4_inversions/scripts/tutorial_5_borders.py
@@ -54,7 +54,7 @@ mask_circular = al.Mask2D.circular(
 )
 
 aplt.Imaging.subplot_imaging(
-    imaging=imaging, mask=mask_circular, include=aplt.Include(border=True)
+    imaging=imaging, mask=mask_circular, include=aplt.Include2D(border=True)
 )
 
 # %%
@@ -74,7 +74,7 @@ mask_annular = al.Mask2D.circular_annular(
 )
 
 aplt.Imaging.subplot_imaging(
-    imaging=imaging, mask=mask_annular, include=aplt.Include(border=True)
+    imaging=imaging, mask=mask_annular, include=aplt.Include2D(border=True)
 )
 
 # %%
@@ -136,8 +136,8 @@ fit = perform_fit_with_source_galaxy_mask_and_border(
     settings_pixelization=al.SettingsPixelization(use_border=False),
 )
 
-aplt.Inversion.reconstruction(
-    inversion=fit.inversion, include=aplt.Include(inversion_grid=True)
+aplt.Inversion.figure_reconstruction(
+    inversion=fit.inversion, include=aplt.Include2D(inversion_grid=True)
 )
 
 # %%
@@ -154,8 +154,8 @@ fit = perform_fit_with_source_galaxy_mask_and_border(
     settings_pixelization=al.SettingsPixelization(use_border=False),
 )
 
-aplt.Inversion.reconstruction(
-    inversion=fit.inversion, include=aplt.Include(inversion_grid=True)
+aplt.Inversion.figure_reconstruction(
+    inversion=fit.inversion, include=aplt.Include2D(inversion_grid=True)
 )
 
 # %%
@@ -171,8 +171,8 @@ Lets quickly check this using a `Mapper` `Plotter`.
 aplt.Mapper.subplot_image_and_mapper(
     image=fit.masked_imaging.image,
     mapper=fit.inversion.mapper,
-    include=aplt.Include(inversion_grid=True),
-    image_pixel_indexes=[
+    include=aplt.Include2D(inversion_grid=True),
+    full_indexes=[
         [3578, 3579, 3580, 3581, 3582],
         [3678, 3679, 3680, 3681, 3682],
         [3778, 3779, 3780, 3781, 3782],
@@ -191,8 +191,8 @@ a quick look.
 """
 
 # %%
-aplt.Tracer.deflections_y(tracer=fit.tracer, grid=fit.grid)
-aplt.Tracer.deflections_x(tracer=fit.tracer, grid=fit.grid)
+aplt.Tracer.figure_deflections_y(tracer=fit.tracer, grid=fit.grid)
+aplt.Tracer.figure_deflections_x(tracer=fit.tracer, grid=fit.grid)
 
 # %%
 """
@@ -231,8 +231,8 @@ fit = perform_fit_with_source_galaxy_mask_and_border(
     settings_pixelization=al.SettingsPixelization(use_border=False),
 )
 
-aplt.Inversion.reconstruction(
-    inversion=fit.inversion, include=aplt.Include(inversion_grid=True)
+aplt.Inversion.figure_reconstruction(
+    inversion=fit.inversion, include=aplt.Include2D(inversion_grid=True)
 )
 
 # %%
@@ -254,15 +254,15 @@ fit = perform_fit_with_source_galaxy_mask_and_border(
     settings_pixelization=al.SettingsPixelization(use_border=True),
 )
 
-aplt.Inversion.reconstruction(
-    inversion=fit.inversion, include=aplt.Include(inversion_grid=True)
+aplt.Inversion.figure_reconstruction(
+    inversion=fit.inversion, include=aplt.Include2D(inversion_grid=True)
 )
 
 aplt.Mapper.subplot_image_and_mapper(
[…]
