+++ howtolens/chapter_2_lens_modeling/scripts/tutorial_3_realism_and_complexity.py
@@ -130,7 +130,8 @@ And lets look at the fit to the `Imaging` data, which as we are used to fits the
 """
 
 # %%
-aplt.FitImaging.subplot_fit_imaging(fit=result.max_log_likelihood_fit)
+fit_imaging_plotter = aplt.FitImagingPlotter(fit=result.max_log_likelihood_fit)
+fit_imaging_plotter.subplot_fit_imaging()
 
 # %%
 """
