+++ howtolens/chapter_2_lens_modeling/scripts/tutorial_5_linking_phases.py
@@ -360,8 +360,8 @@ the centres of the light and mass profiles and the alignment between their ellip
 want to pass priors *and* change the model parameterization.
 
 To do this, we have to use the ``.riorModel__ object in AutoFit, which allows us to turn light and mass profiles into 
-`model components` whose parameters have priors that can be manipulated in an analogous fashion to to ``.alaxyModel__.
-In fact, the individual components of the ``.alaxyModel__ class have been ``.riorModel__`s all along! 
+`model components` whose parameters have priors that can be manipulated in an analogous fashion to to `GalaxyModel`.
+In fact, the individual components of the `GalaxyModel` class have been ``.riorModel__`s all along! 
 """
 
 # %%
@@ -371,7 +371,7 @@ print(source.bulge)
 
 # %%
 """
-We can thus set up the ``.alaxyModel__ we desire, by first creating the individual ``.riorModel__`s of each
+We can thus set up the `GalaxyModel` we desire, by first creating the individual ``.riorModel__`s of each
 component and then passing the priors of each individual parameter. 
 """
 
@@ -399,7 +399,7 @@ lens = al.GalaxyModel(redshift=0.5, bulge=sersic, mass=mass)
 
 # %%
 """
-We now create and run the phase, using the lens ``.alaxyModel__ we created above.
+We now create and run the phase, using the lens `GalaxyModel` we created above.
 """
 
 # %%
