+++ howtolens/chapter_1_introduction/scripts/tutorial_1_grids.py
@@ -38,7 +38,7 @@ First, lets plot this `Grid`, which shows that it is a fairly bland uniform `Gri
 """
 
 # %%
-plotter = aplt.MatPlot2D(labels=aplt.Labels(title="Fairly Bland Uniform Grid Of Dots"))
+plotter = aplt.MatPlot2D(title=aplt.Title(label="Fairly Bland Uniform Grid Of Dots"))
 
 aplt.Grid(grid=grid, plotter=plotter)
 
