+++ autolens/pipeline/phase/imaging/analysis.py
@@ -4,7 +4,7 @@ from autoarray.inversion import pixelizations as pix
 from autoarray.exc import PixelizationException, InversionException, GridException
 from autofit.exc import FitException
 from autogalaxy.pipeline.phase.dataset import analysis as ag_analysis
-from autogalaxy.plot.mat_wrap import lensing_visuals, lensing_include
+from autogalaxy.plot.mat_wrap import lensing_include
 from autolens.fit import fit
 from autolens.pipeline import visualizer as vis
 from autolens.pipeline.phase.dataset import analysis as analysis_dataset
@@ -184,8 +184,11 @@ class Analysis(ag_analysis.Analysis, analysis_dataset.Analysis):
             hyper_background_noise=hyper_background_noise,
         )
 
-        tracer._preload_critical_curves = tracer.critical_curves
-        tracer._preload_caustics = tracer.caustics
+        include_2d = lensing_include.Include2D()
+        tracer._preload_critical_curves = (
+            tracer.critical_curves if include_2d.critical_curves else None
+        )
+        tracer._preload_caustics = tracer.caustics if include_2d.caustics else None
 
         visualizer = vis.Visualizer(visualize_path=paths.image_path)
 
