+++ test_autolens/unit/config/visualize/plots.ini
@@ -39,6 +39,7 @@ model_images_of_planes=False
 plane_images_of_planes=False
 
 [inversion]
+all_at_end_png=False
 subplot_inversion=False
 reconstructed_image=False
 reconstruction=False
