+++ test_autolens/unit/pipeline/config/visualize/plots.ini
@@ -37,6 +37,7 @@ model_images_of_planes=False
 plane_images_of_planes=True
 
 [inversion]
+all_at_end_png=False
 subplot_inversion=True
 reconstructed_image=True
 reconstruction=True
