### 2021-01-09 d408414
fixed stochastic outputs
