### 2021-01-09 #2825 @Jammy2211
Getting there, I've discovered how mind-blowingly-amazing binder is thanks to the review and am restructing some of the docs / turorials around it!

I should have things wrapped up next week :).
