+++ howtolens/chapter_1_introduction/scripts/tutorial_5_ray_tracing.py
@@ -6,7 +6,7 @@ Tutorial 5: Ray Tracing
 In the last tutorial, our use of `Plane`'s was a bit clunky. We manually had to input `Grid`'s to trace them, and keep
 track of which `Grid`'s were the image-plane`s and which were the source planes. It was easy to make mistakes!
 
-Fotunately, in **PyAutoLens**, you won't actually spend much hands-on time with the `Plane` objects. Instead, you'll
+Fortunately, in **PyAutoLens**, you won't actually spend much hands-on time with the `Plane` objects. Instead, you'll
 primarily use the `ray-tracing` module, which we'll cover in this example. Lets look at how easy it is to setup the
 same lens-plane + source-plane strong lens configuration as the previous tutorial, but with a lot less lines of code!
 """
@@ -116,7 +116,8 @@ This image appears as the Einstein ring we saw in the previous tutorial.
 """
 
 # %%
-aplt.Tracer.figure_image(tracer=tracer, grid=image_plane_grid)
+tracer_plotter = aplt.TracerPlotter(tracer=tracer, grid=image_plane_grid)
+tracer_plotter.figure_image()
 
 # %%
 """
@@ -141,42 +142,31 @@ print(traced_grids[1].in_2d[0, 2])
 
 # %%
 """
-We can use the plane_plotter to plot these grids, like we did before.
+We can use the TracerPlotter to plot these planes and grids.
 """
 
 # %%
-plotter = aplt.MatPlot2D(title=aplt.Title(label="Image-plane Grid"))
+include_2d = aplt.Include2D(grid=True)
 
-aplt.plane.plane_grid(plane=tracer.image_plane, grid=traced_grids[0], plotter=plotter)
-
-plotter = aplt.MatPlot2D(title=aplt.Title(label="Source-plane Grid"))
-
-aplt.plane.plane_grid(plane=tracer.source_plane, grid=traced_grids[1], plotter=plotter)
-
-aplt.plane.plane_grid(
-    plane=tracer.source_plane,
-    grid=traced_grids[1],
-    axis_limits=[-0.1, 0.1, -0.1, 0.1],
-    plotter=plotter,
+tracer_plotter = aplt.TracerPlotter(
+    tracer=tracer, grid=image_plane_grid, include_2d=include_2d
 )
+tracer_plotter.figure_plane_image_of_plane(plane_index=0)
+tracer_plotter.figure_plane_image_of_plane(plane_index=1)
 
 # %%
 """
 **PyAutoLens** has tools for plotting a `Tracer`. A ray-tracing subplot plots the following:
 
 1) The image, computed by tracing the source-`Galaxy`'s light `forwards` through the `Tracer`.
-
 2) The source-plane image, showing the source-`Galaxy`'s true appearance (i.e. if it were not lensed).
-
 3) The image-plane convergence, computed using the lens `Galaxy`'s total mass distribution.
-
 4) The image-plane gravitational potential, computed using the lens `Galaxy`'s total mass distribution.
-
 5) The image-plane deflection angles, computed using the lens `Galaxy`'s total mass distribution.
 """
 
 # %%
-aplt.Tracer.subplot_tracer(tracer=tracer, grid=image_plane_grid)
+tracer_plotter.subplot_tracer()
 
 # %%
 """
@@ -239,12 +229,61 @@ commented out again for convenience)
 """
 
 # %%
-aplt.Tracer.figure_convergence(tracer=tracer, grid=image_plane_grid)
+tracer_plotter = aplt.TracerPlotter(tracer=tracer, grid=image_plane_grid)
+tracer_plotter.figure_convergence()
 
[…]
