### 2021-01-10 e70a634
mid magnification edits
