+++ test_autolens/positions_solver/uncentred/positions_maker.py
@@ -1,4 +1,3 @@
-# %%
 """
 __Positions Maker__
 
@@ -11,45 +10,40 @@ upscaling. This means:
 These results are used to test whether more efficient position solvers implementations lose multiple images.
 """
 
-# %%
 import autofit as af
 import autolens as al
 import autolens.plot as aplt
 import os
+from os import path
 
-# %%
 """The pickle path is where the `Tracer` and `Positions` are output, so they can be loaded by other scripts."""
 
-# %%
-path = "{}".format(os.path.dirname(os.path.realpath(__file__)))
-pickle_path = f"{path}/pickles"
+pickle_path = path.join("{}".format(os.path.dirname(os.path.realpath(__file__))), "pickles")
+os.makedirs(pickle_path, exist_ok=True)
 
-# %%
 """A high resolution grid is used to ensure positions are computed to a given accuracy."""
 
-# %%
 grid = al.Grid.uniform(
     shape_2d=(200, 200),
     pixel_scales=0.05,  # <- The pixel-scale describes the conversion from pixel units to arc-seconds.
 )
 
-# %%
 """
 The mass-profile and source light profile in this example have fixed centre (0.0, 0.0), restricting the range of 
 lensing geometries.
 """
 
-# %%
 mass_profile_model = af.PriorModel(al.mp.EllipticalIsothermal)
+
+mass_profile_model.centre.centre_0= af.UniformPrior(lower_limit=-1.0, upper_limit=1.0)
 mass_profile_model.centre.centre_0 = af.UniformPrior(lower_limit=-1.0, upper_limit=1.0)
-mass_profile_model.centre.centre_1 = af.UniformPrior(lower_limit=-1.0, upper_limit=1.0)
-mass_profile_model.elliptical_comps.ellipitical_comps_0 = af.UniformPrior(
+mass_profile_model.elliptical_comps.elliptical_comps_0 = af.UniformPrior(
     lower_limit=-1.0, upper_limit=1.0
 )
-mass_profile_model.elliptical_comps.ellipitical_comps_1 = af.UniformPrior(
+mass_profile_model.elliptical_comps.elliptical_comps_1 = af.UniformPrior(
     lower_limit=-1.0, upper_limit=1.0
 )
-mass_profile_model.centre.einstein_radius = af.UniformPrior(
+mass_profile_model.einstein_radius = af.UniformPrior(
     lower_limit=0.3, upper_limit=2.0
 )
 
@@ -65,7 +59,7 @@ iters = 50
 
 """Use a `PositionsSolver` which does not use grid upscaling."""
 
-solver = al.PositionsFinder(
+solver = al.PositionsSolver(
     grid=grid, use_upscaling=True, pixel_scale_precision=0.0001, upscale_factor=2
 )
 
@@ -97,12 +91,10 @@ for i in range(iters):
 
     """Visually inspect the positions (comment this out if you are confident the code is behaving as expected)."""
 
-    aplt.Tracer.figure_image(
-        tracer=tracer,
-        grid=grid,
-        positions=positions,
-        include=aplt.Include2D(origin=False, critical_curves=False, caustics=False),
-    )
+    visuals_2d = aplt.Visuals2D(positions=positions)
+
+    tracer_plotter = aplt.TracerPlotter(tracer=tracer, grid=grid, visuals_2d=visuals_2d)
+    tracer_plotter.figure_image()
 
     """Save the `Tracer` and `Positions` so they can be used for testing other `PositionsSolver` settings."""
 
