+++ test_autolens/unit/lens/test_positions_solver.py
@@ -1,3 +1,4 @@
+from autoarray.structures import grids
 import autolens as al
 from autolens.lens import positions_solver as pos
 
@@ -77,20 +78,47 @@ class TestAbstractPositionsSolver:
 
         assert (grid == grid_util).all()
 
+    def test__grid_with_points_below_magnification_threshold_removed(self):
 
-class TestPositionFinder:
+        sis = al.mp.SphericalIsothermal(centre=(0.0, 0.0), einstein_radius=1.0)
+
+        grid = grids.GridIrregularGroupedUniform(grid=[(1.0, 0.0), (0.1, 0.0)], pixel_scales=2.0)
+
+        magnification = sis.magnification_irregular_from_grid(grid=grid)
+
+        assert magnification[0] > 0.0
+        assert magnification[1] < 0.0
+
+        solver = pos.AbstractPositionsSolver(magnification_threshold=0.0)
+
+        positions = solver.grid_with_points_below_magnification_threshold_removed(lensing_obj=sis, grid=grid)
+
+        assert positions.in_grouped_list == [[(1.0, 0.0)]]
+        assert positions.pixel_scales == (2.0, 2.0)
+
+        solver = pos.AbstractPositionsSolver(magnification_threshold=-100000.0)
+
+        positions = solver.grid_with_points_below_magnification_threshold_removed(lensing_obj=sis, grid=grid)
+
+        assert positions.in_grouped_list == [[(1.0, 0.0), (0.1, 0.0)]]
+        assert positions.pixel_scales == (2.0, 2.0)
+
+
+class TestPositionSolver:
     def test__positions_found_for_simple_mass_profiles(self):
 
         grid = al.Grid.uniform(shape_2d=(100, 100), pixel_scales=0.05)
 
         sis = al.mp.SphericalIsothermal(centre=(0.0, 0.0), einstein_radius=1.0)
 
-        solver = al.PositionsFinder(grid=grid, pixel_scale_precision=0.01)
+        solver = al.PositionsSolver(grid=grid, pixel_scale_precision=0.01)
 
         positions = solver.solve(lensing_obj=sis, source_plane_coordinate=(0.0, 0.11))
 
-        assert positions[0] == pytest.approx(np.array([0.003125, -0.89062]), 1.0e-4)
-        assert positions[1] == pytest.approx(np.array([-0.003125, -0.89062]), 1.0e-4)
+        assert positions[0] == pytest.approx(np.array([0.003125, 1.109375]), 1.0e-4)
+        assert positions[1] == pytest.approx(np.array([-0.003125, 1.109375]), 1.0e-4)
+
+        stop
 
         grid = al.Grid.uniform(shape_2d=(100, 100), pixel_scales=0.05, sub_size=1)
 
@@ -107,11 +135,13 @@ class TestPositionFinder:
 
         tracer = al.Tracer.from_galaxies(galaxies=[g0, g1])
 
-        solver = pos.PositionsFinder(grid=grid, pixel_scale_precision=0.01)
+        solver = pos.PositionsSolver(grid=grid, pixel_scale_precision=0.01)
 
         coordinates = solver.solve(
             lensing_obj=tracer, source_plane_coordinate=(0.0, 0.0)
         )
+
+        print(coordinates)
 
         assert coordinates.in_grouped_list[0][0] == pytest.approx(
             (1.028125, -0.003125), 1.0e-4
@@ -145,7 +175,7 @@ class TestPositionFinder:
 
         tracer = al.Tracer.from_galaxies(galaxies=[g0, g1])
 
-        solver = pos.PositionsFinder(grid=grid, pixel_scale_precision=0.01)
[…]
