### 2021-01-10 faa7097
More updates to PositionSolver
