### 2021-01-11 829e0ca
slam updated for smnh tag
