+++ autolens/pipeline/phase/point_source/analysis.py
@@ -1,3 +1,4 @@
+from autofit.exc import FitException
 from autogalaxy.pipeline.phase.abstract import analysis as ag_analysis
 from autolens.fit import fit_point_source
 from autolens.pipeline import visualizer as vis
@@ -6,13 +7,15 @@ from autolens.lens import ray_tracing
 
 class Analysis(ag_analysis.Analysis):
     def __init__(
-        self, positions, noise_map, solver, imaging, settings, cosmology, results
+        self, positions, noise_map, fluxes, fluxes_noise_map, solver, imaging, settings, cosmology, results
     ):
 
         super().__init__(settings=settings, cosmology=cosmology)
 
         self.positions = positions
         self.noise_map = noise_map
+        self.fluxes = fluxes
+        self.fluxes_noise_map = fluxes_noise_map
         self.solver = solver
         self.imaging = imaging
         self.results = results
@@ -39,15 +42,37 @@ class Analysis(ag_analysis.Analysis):
         """
 
         tracer = self.tracer_for_instance(instance=instance)
-        fit = self.positions_fit_for_tracer(tracer=tracer)
-        return fit.log_likelihood
 
-    def positions_fit_for_tracer(self, tracer):
+        try:
+            fit_positions = self.fit_positions_for_tracer(tracer=tracer)
+        except AttributeError as e:
+            raise FitException from e
+
+        log_likelihood_positions = fit_positions.log_likelihood
+
+        if self.fluxes is not None:
+            fit_fluxes = self.fit_fluxes_for_tracer(tracer=tracer)
+            log_likelihood_fluxes = fit_fluxes.log_likelihood
+        else:
+            log_likelihood_fluxes = 0.0
+
+        return log_likelihood_positions + log_likelihood_fluxes
+
+    def fit_positions_for_tracer(self, tracer):
 
         return fit_point_source.FitPositionsImage(
             positions=self.positions,
             noise_map=self.noise_map,
             positions_solver=self.solver,
+            tracer=tracer,
+        )
+
+    def fit_fluxes_for_tracer(self, tracer):
+
+        return fit_point_source.FitFluxes(
+            fluxes=self.fluxes,
+            noise_map=self.noise_map,
+            positions=self.positions,
             tracer=tracer,
         )
 
