+++ autolens/lens/ray_tracing.py
@@ -748,6 +748,12 @@ class AbstractTracerData(AbstractTracerLensing, ABC):
 
 
 class Tracer(AbstractTracerData):
+
+    @property
+    def flux_hack(self):
+        """This is a placeholder to get flux modeling working for Nan Li before I do this proeprly. with dictionaries."""
+        return self.planes[1].galaxies[0].light_profiles[0].flux
+
     @classmethod
     def from_galaxies(cls, galaxies, cosmology=cosmo.Planck15):
 
