### 2021-01-14 df5d89b
More simplification of pltoter APi
