+++ test_autolens/simulators/imaging/instrument_util.py
@@ -183,23 +183,23 @@ def simulate_imaging_from_instrument(instrument, dataset_name, galaxies):
 
     aplt.imaging.individual(
         imaging=imaging,
-        plot_image=True,
-        plot_noise_map=True,
-        plot_psf=True,
-        plot_signal_to_noise_map=True,
+        image=True,
+        noise_map=True,
+        psf=True,
+        signal_to_noise_map=True,
         plotter=plotter,
     )
 
     aplt.Tracer.subplot_tracer(tracer=tracer, grid=grid, plotter=plotter)
 
-    aplt.Tracer.figure_individuals(
+    aplt.Tracer.figures(
         tracer=tracer,
         grid=grid,
-        plot_image=True,
-        plot_source_plane=True,
-        plot_convergence=True,
-        plot_potential=True,
-        plot_deflections=True,
+        image=True,
+        source_plane=True,
+        convergence=True,
+        potential=True,
+        deflections=True,
         plotter=plotter,
     )
 
