+++ test_autolens/unit/pipeline/phase/dataset/test_phase_dataset.py
@@ -79,6 +79,68 @@ class TestMakeAnalysis:
         #         dataset=imaging_7x7, results=mock.MockResults()
         #     )
 
+    def test__auto_einstein_radius_is_used__einstein_radius_used_in_analysis(
+        self, imaging_7x7, mask_7x7
+    ):
+        # If position threshold is input (not None) and positions are input, make the positions part of the lens dataset.
+
+        class MockTracer:
+            def __init__(self, einstein_radius_via_tangential_critical_curve):
+
+                self.einstein_radius_via_tangential_critical_curve = (
+                    einstein_radius_via_tangential_critical_curve
+                )
+
+            def einstein_radius_from_grid(self, grid):
+                return self.einstein_radius_via_tangential_critical_curve
+
+            @property
+            def has_mass_profile(self):
+                return True
+
+        phase_imaging_7x7 = al.PhaseImaging(
+            search=mock.MockSearch("test_phase"),
+            settings=al.SettingsPhaseImaging(
+                settings_lens=al.SettingsLens(auto_einstein_radius_factor=None)
+            ),
+        )
+        tracer = MockTracer(einstein_radius_via_tangential_critical_curve=2.0)
+
+        phase_imaging_7x7.modify_settings(
+            dataset=imaging_7x7,
+            results=mock.MockResults(max_log_likelihood_tracer=tracer),
+        )
+
+        analysis = phase_imaging_7x7.make_analysis(
+            dataset=imaging_7x7, mask=mask_7x7, results=mock.MockResults()
+        )
+
+        assert analysis.settings.settings_lens.einstein_radius_estimate == None
+
+        phase_imaging_7x7 = al.PhaseImaging(
+            search=mock.MockSearch("test_phase"),
+            settings=al.SettingsPhaseImaging(
+                settings_lens=al.SettingsLens(auto_einstein_radius_factor=1.0)
+            ),
+        )
+
+        tracer = MockTracer(einstein_radius_via_tangential_critical_curve=2.0)
+
+        phase_imaging_7x7.modify_settings(
+            dataset=imaging_7x7,
+            results=mock.MockResults(max_log_likelihood_tracer=tracer),
+        )
+
+        analysis = phase_imaging_7x7.make_analysis(
+            dataset=imaging_7x7,
+            mask=mask_7x7,
+            results=mock.MockResults(max_log_likelihood_tracer=tracer),
+        )
+
+        print(analysis.settings.settings_lens.einstein_radius_estimate)
+
+        assert analysis.settings.settings_lens.einstein_radius_estimate == 2.0
+
     def test__use_border__determines_if_border_pixel_relocation_is_used(
         self, imaging_7x7, mask_7x7
     ):
