### 2021-01-18 d46bb82
slam source methods now use a result
