+++ howtolens/chapter_2_lens_modeling/scripts/tutorial_5_linking_phases.py
@@ -116,7 +116,7 @@ source = al.GalaxyModel(redshift=1.0, bulge=al.lp.EllipticalExponential)
 
 # %%
 """
-You haven`t actually seen a line like this one before. By setting a parameter to a number (and not a prior) it is be 
+You haven't actually seen a line like this one before. By setting a parameter to a number (and not a prior) it is be 
 removed from non-linear parameter space and always fixed to that value. Pretty neat, huh?
 """
 
@@ -278,7 +278,7 @@ source.bulge.effective_radius = af.GaussianPrior(
 # %%
 """
 Lets setup and run the phase. As expected, it gives us the correct lens model. However, it does so significantly faster 
-than we're used to - I didn`t have to edit the config files to get this phase to run fast!
+than we're used to - I didn't have to edit the config files to get this phase to run fast!
 """
 
 # %%
@@ -361,9 +361,9 @@ The priors passed above retained the model parameterization of phase 1, includin
 the centres of the light and mass profiles and the alignment between their elliptical components. However, we often 
 want to pass priors *and* change the model parameterization.
 
-To do this, we have to use the ``.riorModel__ object in AutoFit, which allows us to turn light and mass profiles into 
+To do this, we have to use the `PriorModel` object in AutoFit, which allows us to turn light and mass profiles into 
 `model components` whose parameters have priors that can be manipulated in an analogous fashion to to `GalaxyModel`.
-In fact, the individual components of the `GalaxyModel` class have been ``.riorModel__`s all along! 
+In fact, the individual components of the `GalaxyModel` class have been `PriorModel`'s all along! 
 """
 
 # %%
@@ -373,7 +373,7 @@ print(source.bulge)
 
 # %%
 """
-We can thus set up the `GalaxyModel` we desire, by first creating the individual ``.riorModel__`s of each
+We can thus set up the `GalaxyModel`  we desire, by first creating the individual `PriorModel`'s of each
 component and then passing the priors of each individual parameter. 
 """
 
@@ -401,7 +401,11 @@ lens = al.GalaxyModel(redshift=0.5, bulge=sersic, mass=mass)
 
 # %%
 """
+<<<<<<< HEAD
 We now create and run the phase, using the lens `GalaxyModel` we created above.
+=======
+We now create and run the phase, using the lens `GalaxyModel`  we created above.
+>>>>>>> 56afab9d8197a33dd3c50ce716ede9bb89276eac
 """
 
 # %%
@@ -431,7 +435,9 @@ these priors?
 
 Lets say I link two parameters as follows:
  
+ ```
  mass.einstein_radius = phase1_result.model.galaxies.lens.mass.einstein_radius
+```
 
 By invoking the `model` attribute, the prioris passed following 3 rules:
 
@@ -440,14 +446,15 @@ By invoking the `model` attribute, the prioris passed following 3 rules:
 
  2) The mean of the GaussianPrior is the median PDF value of the parameter estimated in phase 1.
     
- This ensures that the initial sampling of the new phase`s non-linear starts by searching the region of non-linear 
+ This ensures that the initial sampling of the new phase's non-linear starts by searching the region of non-linear 
  parameter space that correspond to highest log likelihood solutions in the previous phase. Thus, we're setting 
  our priors to look in the `correct` regions of parameter space.
 
  3) The sigma of the Gaussian will use the maximum of two values: 
    
  (i) the 1D error of the parameter computed at an input sigma value (default sigma=3.0).
- (ii) The value specified for the profile in the `config/priors/*.json` config file`s `width_modifer` 
+ 
+ (ii) The value specified for the profile in the `config/priors/*.json` config file's `width_modifer` 
[…]
