+++ howtolens/chapter_2_lens_modeling/scripts/tutorial_2_parameter_space_and_priors.py
@@ -36,8 +36,8 @@ The point is, like we did for the simple functions above, we again have a parame
 down analytically and its undoubtedly very complex and non-linear. Fortunately, we've already learnt how to search
 it, and find the solutions which maximize our log_likelihood function!
 
-Lets inspect the results of the last tutorial's non-linear search. we're going to look at what are called `probably
-density functions` or PDF`s for short. These represent where the highest log likelihood regions of parameter space were
+Lets inspect the results of the last tutorial's non-linear search. we're going to look at what are called `Probability
+density functions` or `PDF`s for short. These represent where the highest log likelihood regions of parameter space were
 found for each parameter.
 
 Navigate to the image folder in `autolens_workspace/output/howtolens/phase_t1_non_linear_search`
