### 2021-01-19 e8b1cf9
simplified slam module
