### 2021-01-21 f87983c
einstein radius constraining methods added
