+++ autolens/__init__.py
@@ -85,4 +85,4 @@ from autoconf import conf
 
 conf.instance.register(__file__)
 
-__version__ = '1.10.0'
+__version__ = "1.10.0"
