### 2021-01-21 42304c6
normalization from mass angular
