### 2021-01-23 5e74764
Mvoed stellar mass profile method
