+++ howtolens/chapter_4_inversions/scripts/tutorial_3_inversions.py
@@ -9,12 +9,11 @@ can use these `Mapper`'s (which map things) to reconstruct the source galaxy - I
 
 # %%
 #%matplotlib inline
-
-from pyprojroot import here
-
-workspace_path = str(here())
-#%cd $workspace_path
-print(f"Working Directory has been set to `{workspace_path}`")
+# %matplotlib inline
+# from pyprojroot import here
+# workspace_path = str(here())
+# %cd $workspace_path
+# print(f"Working Directory has been set to `{workspace_path}`")
 
 from os import path
 import autolens as al
