+++ howtolens/chapter_4_inversions/scripts/tutorial_5_borders.py
@@ -8,12 +8,11 @@ In the previous tutorial, we told our `Inversion` to use a border. Here, we'll d
 
 # %%
 #%matplotlib inline
-
-from pyprojroot import here
-
-workspace_path = str(here())
-#%cd $workspace_path
-print(f"Working Directory has been set to `{workspace_path}`")
+# %matplotlib inline
+# from pyprojroot import here
+# workspace_path = str(here())
+# %cd $workspace_path
+# print(f"Working Directory has been set to `{workspace_path}`")
 
 from os import path
 import autolens as al
