+++ howtolens/chapter_1_introduction/scripts/tutorial_9_summary.py
@@ -21,12 +21,11 @@ system. Lets get a `fit` to a strong lens, by setting up an image, mask, tracer,
 
 # %%
 #%matplotlib inline
-
-from pyprojroot import here
-
-workspace_path = str(here())
-#%cd $workspace_path
-print(f"Working Directory has been set to `{workspace_path}`")
+# %matplotlib inline
+# from pyprojroot import here
+# workspace_path = str(here())
+# %cd $workspace_path
+# print(f"Working Directory has been set to `{workspace_path}`")
 
 from os import path
 import autolens as al
