+++ howtolens/chapter_2_lens_modeling/scripts/tutorial_7_masking_and_positions.py
@@ -9,12 +9,11 @@ choose your mask. I'll also show you another neat trick to improve the speed and
 
 # %%
 #%matplotlib inline
-
-from pyprojroot import here
-
-workspace_path = str(here())
-#%cd $workspace_path
-print(f"Working Directory has been set to `{workspace_path}`")
+# %matplotlib inline
+# from pyprojroot import here
+# workspace_path = str(here())
+# %cd $workspace_path
+# print(f"Working Directory has been set to `{workspace_path}`")
 
 from os import path
 import autolens as al
