+++ howtolens/chapter_3_pipelines/scripts/tutorial_2_pipeline_runner.py
@@ -19,12 +19,11 @@ pipeline that we can generalize to many lenses isn't currently possible with **P
 
 # %%
 #%matplotlib inline
-
-from pyprojroot import here
-
-workspace_path = str(here())
-#%cd $workspace_path
-print(f"Working Directory has been set to `{workspace_path}`")
+# %matplotlib inline
+# from pyprojroot import here
+# workspace_path = str(here())
+# %cd $workspace_path
+# print(f"Working Directory has been set to `{workspace_path}`")
 
 from os import path
 import autolens as al
