+++ howtolens/chapter_4_inversions/scripts/tutorial_2_mappers.py
@@ -9,12 +9,11 @@ was actually mapping. Infact, it didn`t do much mapping at all! Therefore, in th
 
 # %%
 #%matplotlib inline
-
-from pyprojroot import here
-
-workspace_path = str(here())
-#%cd $workspace_path
-print(f"Working Directory has been set to `{workspace_path}`")
+# %matplotlib inline
+# from pyprojroot import here
+# workspace_path = str(here())
+# %cd $workspace_path
+# print(f"Working Directory has been set to `{workspace_path}`")
 
 from os import path
 import autolens as al
