+++ howtolens/chapter_3_pipelines/scripts/tutorial_4_setup_and_slam.py
@@ -22,12 +22,11 @@ Lets take a look.
 
 # %%
 #%matplotlib inline
-
-from pyprojroot import here
-
-workspace_path = str(here())
-#%cd $workspace_path
-print(f"Working Directory has been set to `{workspace_path}`")
+# %matplotlib inline
+# from pyprojroot import here
+# workspace_path = str(here())
+# %cd $workspace_path
+# print(f"Working Directory has been set to `{workspace_path}`")
 
 import autolens as al
 
