+++ howtolens/chapter_1_introduction/scripts/tutorial_4_planes.py
@@ -41,12 +41,11 @@ For simplicity, we'll assume 1 lens galaxy and 1 source galaxy.
 
 # %%
 #%matplotlib inline
-
-from pyprojroot import here
-
-workspace_path = str(here())
-#%cd $workspace_path
-print(f"Working Directory has been set to `{workspace_path}`")
+# %matplotlib inline
+# from pyprojroot import here
+# workspace_path = str(here())
+# %cd $workspace_path
+# print(f"Working Directory has been set to `{workspace_path}`")
 
 import autolens as al
 import autolens.plot as aplt
