### 2021-01-24 2c0a86f
slim / native autoarray API
