+++ howtolens/simulators/chapter_3/light_sersic__mass_sie__source_exp.py
@@ -30,8 +30,8 @@ The path where the dataset will be output, which in this case is:
 dataset_path = path.join("dataset", "howtolens", dataset_type, dataset_name)
 
 """
-For simulating an image of a strong lens, we recommend using a GridIterate object. This represents a grid of $(y,x)$ 
-coordinates like an ordinary Grid, but when the light-profile`s image is evaluated below (using the Tracer) the 
+For simulating an image of a strong lens, we recommend using a Grid2DIterate object. This represents a grid of $(y,x)$ 
+coordinates like an ordinary Grid2D, but when the light-profile`s image is evaluated below (using the Tracer) the 
 sub-size of the grid is iteratively increased (in steps of 2, 4, 8, 16, 24) until the input fractional accuracy of 
 99.99% is met.
 
@@ -39,8 +39,8 @@ This ensures that the divergent and bright central regions of the source galaxy 
 total flux emitted within a pixel.
 """
 
-grid = al.GridIterate.uniform(
-    shape_2d=(100, 100),
+grid = al.Grid2DIterate.uniform(
+    shape_native=(100, 100),
     pixel_scales=0.1,
     fractional_accuracy=0.9999,
     sub_steps=[2, 4, 8, 16, 24],
@@ -48,8 +48,8 @@ grid = al.GridIterate.uniform(
 
 """Simulate a simple Gaussian PSF for the image."""
 
-psf = al.Kernel.from_gaussian(
-    shape_2d=(11, 11), sigma=0.1, pixel_scales=grid.pixel_scales
+psf = al.Kernel2D.from_gaussian(
+    shape_native=(11, 11), sigma=0.1, pixel_scales=grid.pixel_scales
 )
 
 """
