+++ howtolens/chapter_1_introduction/scripts/tutorial_1_grids.py
@@ -3,7 +3,7 @@
 Tutorial 1: Grids
 =================
 
-In this example, we'll create `Grid`'s of Cartesian $(y,x)$ coordinates, representing the arc-second coordinate `Grid`
+In this example, we'll create `Grid2D`'s of Cartesian $(y,x)$ coordinates, representing the arc-second coordinate `Grid2D`
 of an observed data-set (e.g. imaging).
 """
 
@@ -20,101 +20,101 @@ import autolens.plot as aplt
 
 # %%
 """
-In **PyAutoLens**, a `Grid` is a set of two-dimensional $(y,x)$ coordinates (in arc-seconds) that are deflected and traced by 
+In **PyAutoLens**, a `Grid2D` is a set of two-dimensional $(y,x)$ coordinates (in arc-seconds) that are deflected and traced by 
 a strong lensing system.
 
-This coordinate `Grid` is aligned with the image we analyze, such that each coordinate on a  `Grid` maps to the centre 
-of each image-pixel. Lets make a `Grid` using 100 x 100 pixels, with a pixel scale (arcsecond-to-pixel conversion 
+This coordinate `Grid2D` is aligned with the image we analyze, such that each coordinate on a  `Grid2D` maps to the centre 
+of each image-pixel. Lets make a `Grid2D` using 100 x 100 pixels, with a pixel scale (arcsecond-to-pixel conversion 
 factor) of 0.05", giving us a 5" x 5" grid.
 """
 
 # %%
-grid = al.Grid.uniform(shape_2d=(100, 100), pixel_scales=0.05)
+grid = al.Grid2D.uniform(shape_native=(100, 100), pixel_scales=0.05)
 
 # %%
 """
-First, lets plot this `Grid`, which shows that it is a fairly bland uniform `Grid` of dots.
+First, lets plot this `Grid2D`, which shows that it is a fairly bland uniform `Grid2D` of dots.
 """
 
 # %%
 mat_plot_2d = aplt.MatPlot2D(
-    title=aplt.Title(label="Fairly Bland Uniform Grid Of Dots")
+    title=aplt.Title(label="Fairly Bland Uniform Grid2D Of Dots")
 )
 
-grid_plotter = aplt.GridPlotter(grid=grid, mat_plot_2d=mat_plot_2d)
+grid_plotter = aplt.Grid2DPlotter(grid=grid, mat_plot_2d=mat_plot_2d)
 grid_plotter.figure()
 
 # %%
 """
-We can print each coordinate of this `Grid`, revealing that it consists of a set of arc-second coordinates (where the 
+We can print each coordinate of this `Grid2D`, revealing that it consists of a set of arc-second coordinates (where the 
 spacing between each coordinate corresponds to the `pixel_scales` of 0.05" defined above)
 """
 
 # %%
 print("(y,x) pixel 0:")
-print(grid.in_2d[0, 0])
+print(grid.native[0, 0])
 print("(y,x) pixel 1:")
-print(grid.in_2d[0, 1])
+print(grid.native[0, 1])
 print("(y,x) pixel 2:")
-print(grid.in_2d[0, 2])
+print(grid.native[0, 2])
 print("(y,x) pixel 100:")
-print(grid.in_2d[1, 0])
+print(grid.native[1, 0])
 print("etc.")
 
 # %%
 """
-`Grid``s in **PyAutoLens** are stored as both 1D and 2D ndarrays, because different calculations benefit from us using 
-the array in different formats. We can access both the 1D and 2D `Array`'s automatically by specifying the input as a 1D 
+`Grid2D``s in **PyAutoLens** are stored as both 1D and 2D ndarrays, because different calculations benefit from us using 
[…]
