+++ howtolens/chapter_3_pipelines/scripts/tutorial_1_pipeline_runner.py
@@ -70,7 +70,7 @@ of its light in the central regions of the image, so lets use a circular mask.
 
 # %%
 mask = al.Mask2D.circular(
-    shape_2d=imaging.shape_2d, pixel_scales=imaging.pixel_scales, radius=3.0
+    shape_native=imaging.shape_native, pixel_scales=imaging.pixel_scales, radius=3.0
 )
 
 imaging_plotter = aplt.ImagingPlotter(
@@ -91,7 +91,7 @@ The settings chosen here are applied to all phases in the pipeline.
 """
 
 # %%
-settings_masked_imaging = al.SettingsMaskedImaging(grid_class=al.Grid, sub_size=2)
+settings_masked_imaging = al.SettingsMaskedImaging(grid_class=al.Grid2D, sub_size=2)
 
 settings = al.SettingsPhaseImaging(settings_masked_imaging=settings_masked_imaging)
 

[…]
