+++ autolens/pipeline/phase/imaging/phase.py
@@ -104,7 +104,7 @@ class PhaseImaging(dataset.PhaseDataset):
             )
             phase_info.write(
                 "PSF shape = {} \n".format(
-                    self.settings.settings_masked_imaging.psf_shape_native
+                    self.settings.settings_masked_imaging.psf_shape_2d
                 )
             )
             phase_info.write(
