+++ test_autolens/plot/criticals_and_caustics/plotting_via_different_plotters.py
@@ -1,23 +1,0 @@
-import autolens as al
-import autolens.plot as aplt
-
-grid = al.Grid2D.uniform(shape_native=(100, 100), pixel_scales=0.05, sub_size=4)
-
-sis_mass_profile = al.mp.EllipticalIsothermal(
-    centre=(1.0, 1.0), einstein_radius=1.6, axis_ratio=0.7
-)
-
-aplt.MassProfile.figure_convergence(
-    mass_profile=sis_mass_profile,
-    grid=grid,
-    include=aplt.Include2D(critical_curves=True, caustics=True),
-)
-
-# galaxy = al.Galaxy(mass=sis_mass_profile, redshift=1)
-#
-# aplt.Galaxy.convergence(
-#     galaxy=galaxy,
-#     grid=grid,
-#     include_critical_curves=False,
-#     include_caustics=True,
-# )
