+++ howtolens/chapter_2_lens_modeling/scripts/tutorial_8_results.py
@@ -66,7 +66,7 @@ a fast way to visualize the result.
 (Uncomment the line below to plot the tracer).
 """
 tracer_plotter = aplt.TracerPlotter(
-    tracer=result.max_log_likelihood_tracer, grid=mask.unmasked_grid
+    tracer=result.max_log_likelihood_tracer, grid=mask.unmasked_grid_sub_1
 )
 tracer_plotter.subplot_tracer()
 
