+++ howtolens/simulators/chapter_4/mass_sie__source_sersic_x5.py
@@ -1,177 +1,0 @@
-from os import path
-import autolens as al
-
-"""
-This script simulates `Imaging` of a strong lens where:
-
- - The lens `Galaxy`'s total mass distribution is a *SphericalIsothermal*.
- - The source `Galaxy`'s `LightProfile` is a *SphericalExponential*.
-
-This dataset is used in chapter 2, tutorials 1-3.
-"""
-
-"""
-The `dataset_type` describes the type of data being simulated (in this case, `Imaging` data) and `dataset_name` 
-gives it a descriptive name. They define the folder the dataset is output to on your hard-disk:
-
- - The image will be output to `/autolens_workspace/dataset/dataset_type/dataset_name/image.fits`.
- - The noise-map will be output to `/autolens_workspace/dataset/dataset_type/dataset_name/lens_name/noise_map.fits`.
- - The psf will be output to `/autolens_workspace/dataset/dataset_type/dataset_name/psf.fits`.
-"""
-dataset_type = "chapter_4"
-dataset_name = "mass_sie__source_sersic_x5"
-
-"""
-The path where the dataset will be output, which in this case is:
-`/autolens_workspace/howtolens/dataset/chapter_2/lens_sis__source_sersic_x5/`
-"""
-dataset_path = path.join("dataset", "howtolens", dataset_type, dataset_name)
-
-"""
-For simulating an image of a strong lens, we recommend using a Grid2DIterate object. This represents a grid of $(y,x)$ 
-coordinates like an ordinary Grid2D, but when the light-profile`s image is evaluated below (using the Tracer) the 
-sub-size of the grid is iteratively increased (in steps of 2, 4, 8, 16, 24) until the input fractional accuracy of 
-99.99% is met.
-
-This ensures that the divergent and bright central regions of the source galaxy are fully resolved when determining the
-total flux emitted within a pixel.
-"""
-grid = al.Grid2DIterate.uniform(
-    shape_native=(100, 100),
-    pixel_scales=0.1,
-    fractional_accuracy=0.9999,
-    sub_steps=[2, 4, 8, 16, 24],
-)
-
-"""Simulate a simple Gaussian PSF for the image."""
-psf = al.Kernel2D.from_gaussian(
-    shape_native=(11, 11), sigma=0.1, pixel_scales=grid.pixel_scales
-)
-
-"""
-To simulate the `Imaging` dataset we first create a simulator, which defines the exposure time, background sky,
-noise levels and psf of the dataset that is simulated.
-"""
-simulator = al.SimulatorImaging(
-    exposure_time=300.0, psf=psf, background_sky_level=0.1, add_poisson_noise=True
-)
-
-"""
-Setup the lens `Galaxy`'s mass (SIE+Shear) and source galaxy light (elliptical Sersic) for this simulated lens.
-
-For lens modeling, defining ellipticity in terms of the `elliptical_comps` improves the model-fitting procedure.
-
-However, for simulating a strong lens you may find it more intuitive to define the elliptical geometry using the 
-axis-ratio of the profile (axis_ratio = semi-major axis / semi-minor axis = b/a) and position angle phi, where phi is
-in degrees and defined counter clockwise from the positive x-axis.
-
-We can use the **PyAutoLens** `convert` module to determine the elliptical components from the axis-ratio and phi.
-"""
-
-lens_galaxy = al.Galaxy(
-    redshift=0.5,
-    mass=al.mp.EllipticalIsothermal(
[…]
