+++ howtolens/chapter_4_inversions/scripts/tutorial_7_adaptive_pixelization.py
@@ -1,200 +1,0 @@
-# %%
-"""
-Tutorial 7: Adaptive Pixelization
-=================================
-
-In this tutorial we'll use a new `Pixelization`, called the VoronoiMagnification `Pixelization`. This pixelization
-doesn`t use a uniform `Grid2D` of rectangular pixels, but instead uses an irregular `Voronoi` pixels. So, why do we
-want to do that? Lets take another quick look at the rectangular grid..
-"""
-
-# %%
-#%matplotlib inline
-# %matplotlib inline
-# from pyprojroot import here
-# workspace_path = str(here())
-# %cd $workspace_path
-# print(f"Working Directory has been set to `{workspace_path}`")
-
-from os import path
-import autolens as al
-import autolens.plot as aplt
-
-
-# %%
-"""
-we'll use the same strong lensing data as the previous tutorial, where:
-
- - The lens `Galaxy`'s light is omitted.
- - The lens `Galaxy`'s total mass distribution is an `EllipticalIsothermal`.
- - The source `Galaxy`'s `LightProfile` is an `EllipticalSersic`.
-"""
-
-# %%
-dataset_name = "mass_sie__source_sersic"
-dataset_path = path.join("dataset", "howtolens", "chapter_4", dataset_name)
-
-imaging = al.Imaging.from_fits(
-    image_path=path.join(dataset_path, "image.fits"),
-    noise_map_path=path.join(dataset_path, "noise_map.fits"),
-    psf_path=path.join(dataset_path, "psf.fits"),
-    pixel_scales=0.1,
-)
-
-mask = al.Mask2D.circular(
-    shape_native=imaging.shape_native,
-    pixel_scales=imaging.pixel_scales,
-    sub_size=2,
-    radius=3.0,
-)
-
-imaging_plotter = aplt.ImagingPlotter(
-    imaging=imaging, visuals_2d=aplt.Visuals2D(mask=mask)
-)
-imaging_plotter.subplot_imaging()
-
-# %%
-"""
-The lines of code below do everything we're used to, that is, setup an image, mask it, trace it via a tracer, 
-setup the rectangular mapper, etc.
-"""
-
-# %%
-lens_galaxy = al.Galaxy(
-    redshift=0.5,
-    mass=al.mp.EllipticalIsothermal(
-        centre=(0.0, 0.0), elliptical_comps=(0.111111, 0.0), einstein_radius=1.6
-    ),
-)
-
-masked_imaging = al.MaskedImaging(
-    imaging=imaging, mask=mask, settings=al.SettingsMaskedImaging(sub_size=2)
-)
-
-source_galaxy = al.Galaxy(
-    redshift=1.0,
-    pixelization=al.pix.Rectangular(shape=(40, 40)),
-    regularization=al.reg.Constant(coefficient=0.5),
-)
-
-tracer = al.Tracer.from_galaxies(galaxies=[lens_galaxy, source_galaxy])
-
-fit = al.FitImaging(masked_imaging=masked_imaging, tracer=tracer)
-
-include_2d = aplt.Include2D(mask=True)
-
-fit_imaging_plotter = aplt.FitImagingPlotter(fit=fit, include_2d=include_2d)
-fit_imaging_plotter.subplot_fit_imaging()
-fit_imaging_plotter.subplot_of_planes(plane_index=1)
-
-# %%
-"""
-Okay, so lets think about the `Rectangular` `Pixelization`. Is this the optimal way to reconstruct our source? Are 
-there features in the source-plane that arn`t ideal? How do you think we could do a better job?
-
-Well, given we're doing a whole tutorial on using a different `Pixelization` to this grid, you`ve probably guessed that
[…]

[…]
