### 2021-02-08 d3f4ed2
fixed inversion bug + removed grouped grids
