+++ autolens/dataset/point_source.py
@@ -1,0 +1,8 @@
+from autoarray.structures import arrays, grids
+
+from typing import Optional
+
+#
+# class PointSourceDataset:
+#
+#     def __init__(self, name, positions : grids.Grid2DIrregular = None, fluxes : arrays.ValuesIrregular):
