+++ test_autolens/unit/lens/test_ray_tracing.py
@@ -7,7 +7,7 @@ import shutil
 from astropy import cosmology as cosmo
 from skimage import measure
 from autoarray.mock import mock as mock_inv
-
+from autolens.mock import mock
 
 test_path = path.join(
     "{}".format(path.dirname(path.realpath(__file__))), "files", "tracer"
@@ -390,6 +390,48 @@ class TestAbstractTracer:
                 2,
             ]
 
+        def test__point_source_dicts(self, ps_0, ps_1):
+
+            tracer = al.Tracer.from_galaxies(galaxies=[al.Galaxy(redshift=0.5)])
+
+            assert tracer.point_source_dict == {}
+            assert tracer.point_source_plane_index_dict == {}
+
+            tracer = al.Tracer.from_galaxies(
+                galaxies=[al.Galaxy(redshift=0.5, point_0=ps_0)]
+            )
+
+            assert tracer.point_source_dict == {"point_0": ps_0}
+            assert tracer.point_source_plane_index_dict == {"point_0": 0}
+
+            tracer = al.Tracer.from_galaxies(
+                galaxies=[al.Galaxy(redshift=0.5, point_0=ps_0, point_1=ps_1)]
+            )
+
+            assert tracer.point_source_dict == {"point_0": ps_0, "point_1": ps_1}
+            assert tracer.point_source_plane_index_dict == {"point_0": 0, "point_1": 0}
+
+            tracer = al.Tracer.from_galaxies(
+                galaxies=[
+                    al.Galaxy(redshift=0.5, point_0=ps_0, point_1=ps_1),
+                    al.Galaxy(redshift=1.0, point_2=ps_0),
+                    al.Galaxy(redshift=2.0, point_3=ps_0),
+                ]
+            )
+
+            assert tracer.point_source_dict == {
+                "point_0": ps_0,
+                "point_1": ps_1,
+                "point_2": ps_0,
+                "point_3": ps_0,
+            }
+            assert tracer.point_source_plane_index_dict == {
+                "point_0": 0,
+                "point_1": 0,
+                "point_2": 1,
+                "point_3": 2,
+            }
+
     class TestPixelizations:
         def test__no_galaxy_has_regularization__returns_list_of_ones(
             self, sub_grid_7x7
@@ -501,74 +543,6 @@ class TestAbstractTracer:
 
             assert tracer.galaxies == [g0, g1, g4, g2, g3, g5]
 
-        # def test__galaxy_in_planes_lists__comes_in_grouped_lists_of_planes_in_redshift_order(self, sub_grid_7x7):
-        #     g0 = al.Galaxy(redshift=0.5)
-        #     g1 = al.Galaxy(redshift=0.5)
-        #
-        #     tracer = al.Tracer.from_galaxies(galaxies=[g0, g1])
-        #
-        #     assert tracer.galaxies_in_planes == [[g0, g1]]
-        #
-        #     g2 = al.Galaxy(redshift=1.0)
-        #     g3 = al.Galaxy(redshift=1.0)
-        #
-        #     tracer = al.Tracer.from_galaxies(galaxies=[g0, g1], galaxies=[g2, g3],
-        #                                                  )
-        #
-        #     assert tracer.galaxies_in_planes == [[g0, g1], [g2, g3]]
-        #
-        #     g4 = al.Galaxy(redshift=0.75)
-        #     g5 = al.Galaxy(redshift=1.5)
-        #
-        #     tracer = al.Tracer.from_galaxies(galaxies=[g0, g1, g2, g3, g4, g5],
-        #                                            )
-        #
-        #     assert tracer.galaxies_in_planes == [[g0, g1], [g4], [g2, g3], [g5]]
-
-    class TestLightProfileQuantities:
[…]
