+++ docs/howtolens/chapter_1_introduction.rst
@@ -30,10 +30,10 @@ The chapter contains the following tutorials:
 - Advanced strong lens ray-tracing.
 
 `Tutorial 7: Data <https://github.com/Jammy2211/autolens_workspace/blob/master/notebooks/howtolens/chapter_1_introduction/tutorial_7_data.ipynb>`_
-- Loading and inspecting telescope imaging ``data`` of a strong lens.
+- Loading and inspecting telescope imaging data of a strong lens.
 
 `Tutorial 8: Fitting <https://github.com/Jammy2211/autolens_workspace/blob/master/notebooks/howtolens/chapter_1_introduction/tutorial_8_fittingn.ipynb>`_
-- Fitting ``data`` with a strong lens model.
+- Fitting data with a strong lens model.
 
 `Tutorial 9: Summary <https://github.com/Jammy2211/autolens_workspace/blob/master/notebooks/howtolens/chapter_1_introduction/tutorial_9_summary.ipynb>`_
 - A summary of the chapter.
