+++ docs/general/credits.rst
@@ -10,13 +10,15 @@ Credits
 
 `Richard Hayes <https://github.com/rhayes777>`_ - Lead developer & `PyAutoFit <https://github.com/rhayes777/PyAutoFit>`_ guru.
 
-`Ashley Kelly <https://github.com/AshKelly>`_ - Developer of `pyquad <https://github.com/AshKelly/pyquad>`_ for fast deflections computations.
+`Aristeidis Amvrosiadis <https://github.com/Sketos>`_ - Interferometer Analysis.
 
 `Amy Etherington <https://github.com/amyetherington>`_ - Magnification, Critical Curves and Caustic Calculations.
 
-`Xiaoyue Cao <https://github.com/caoxiaoyue>`_ - Analytic Ellipitcal Power-Law Deflection Angle Calculations.
+`Ashley Kelly <https://github.com/AshKelly>`_ - Developer of `pyquad <https://github.com/AshKelly/pyquad>`_ for fast deflections computations.
 
 Qiuhan He  - NFW Profile Lensing Calculations.
+
+`Xiaoyue Cao <https://github.com/caoxiaoyue>`_ - Analytic Ellipitcal Power-Law Deflection Angle Calculations.
 
 `Nan Li <https://github.com/linan7788626>`_ - Docker integration & support.
 
@@ -27,7 +29,5 @@ Qiuhan He  - NFW Profile Lensing Calculations.
 Mattia Negrello - Visibility models in the uv-plane via direct Fourier transforms.
 
 `Andrea Enia <https://github.com/AndreaEnia>`_ - Voronoi source-plane plotting tools.
-
-`Aristeidis Amvrosiadis <https://github.com/Sketos>`_ - ALMA imaging ``data`` loading.
 
 Conor O'Riordan  - Broken Power-Law mass profile.
