+++ docs/howtolens/chapter_5_hyper_mode.rst
@@ -8,21 +8,21 @@ and navigating to the folder ``notebooks/howtolens/chapter_5_hyper_mode``. They 
 
 The chapter contains the following tutorials:
 
-`Tutorial 1: Fit Problems <https://github.com/Jammy2211/autolens_workspace/blob/master/howtolens/chapter_5_hyper_mode/tutorial_1_fit_problems.ipynb>`_
+`Tutorial 1: Fit Problems <https://github.com/Jammy2211/autolens_workspace/blob/master/notebooks/howtolens/chapter_5_hyper_mode/tutorial_1_fit_problems.ipynb>`_
 - The short comings of our lens models and inversions.
 
-`Tutorial 2: Brightness Adaption <https://github.com/Jammy2211/autolens_workspace/blob/master/howtolens/chapter_5_hyper_mode/tutorial_2_brightness_adaption.ipynb>`_
+`Tutorial 2: Brightness Adaption <https://github.com/Jammy2211/autolens_workspace/blob/master/notebooks/howtolens/chapter_5_hyper_mode/tutorial_2_brightness_adaption.ipynb>`_
 - Adapting the pixelization to the source's morphology.
 
-`Tutorial 3: Adaptive Regularization <https://github.com/Jammy2211/autolens_workspace/blob/master/howtolens/chapter_5_hyper_mode/tutorial_3_adaptive_regularization.ipynb>`_
+`Tutorial 3: Adaptive Regularization <https://github.com/Jammy2211/autolens_workspace/blob/master/notebooks/howtolens/chapter_5_hyper_mode/tutorial_3_adaptive_regularization.ipynb>`_
 - Adapting the regularization to the source's morphology.
 
-`Tutorial 4: Noise Map Scaling 1 <https://github.com/Jammy2211/autolens_workspace/blob/master/howtolens/chapter_5_hyper_mode/tutorial_4_noise_map_scaling_1.ipynb>`_
+`Tutorial 4: Noise Map Scaling 1 <https://github.com/Jammy2211/autolens_workspace/blob/master/notebooks/howtolens/chapter_5_hyper_mode/tutorial_4_noise_map_scaling_1.ipynb>`_
 - Increasing the noise to prevent over-fitting compact sources. 
 
-`Tutorial 5: Noise Map Scaling 2 <https://github.com/Jammy2211/autolens_workspace/blob/master/howtolens/chapter_5_hyper_mode/tutorial_5_noise_map_scaling_2.ipynb>`_
+`Tutorial 5: Noise Map Scaling 2 <https://github.com/Jammy2211/autolens_workspace/blob/master/notebooks/howtolens/chapter_5_hyper_mode/tutorial_5_noise_map_scaling_2.ipynb>`_
 - Increasing the noise to improve fitting irregular lens light morphologies.
 
-`Tutorial 6: Hyper Pipeline  <https://github.com/Jammy2211/autolens_workspace/blob/master/howtolens/chapter_5_hyper_mode/tutorial_6_hyper_pipeline_runner.ipynb>`_
+`Tutorial 6: Hyper Pipeline  <https://github.com/Jammy2211/autolens_workspace/blob/master/notebooks/howtolens/chapter_5_hyper_mode/tutorial_6_hyper_pipeline_runner.ipynb>`_
 - An example pipeline use hyper-mode.
 
