+++ citations.bib
@@ -142,13 +142,17 @@ year = {2015}
   volume={13},
   number={2},
   pages={22-30},}
-@misc{pyautofit,
-  author = {James Nightingale and Richard Hayes},
-  title = {PyAutoFit: A classy Probabilistic Programming Language for Complex Model-fitting},
-  year = {2020},
-  publisher = {​GitHub},
-  journal = {​GitHub repository},
-  url = {https://github.com/rhayes777/PyAutoFit}
+@article{pyautofit,
+  doi = {10.21105/joss.02550},
+  url = {https://doi.org/10.21105/joss.02550},
+  year = {2021},
+  publisher = {The Open Journal},
+  volume = {6},
+  number = {58},
+  pages = {2550},
+  author = {James. W. Nightingale and Richard G. Hayes and Matthew Griffiths},
+  title = {`PyAutoFit`: A Classy Probabilistic Programming Language for Model Composition and Fitting},
+  journal = {Journal of Open Source Software}
 }
 @misc{pyautolens,
   author = {James Nightingale and Richard Hayes},
