### 2021-02-09 #2825 @Jammy2211
So, it turns out that switching all our docs to binder was quite an undertaking! But very much worth it in the end :).

Our team has begun setting up GitHub actions and has a meeting on Wednesday to fully sort this, which should sort most of the remaining issues (automated tests, tags). So, I should have finished addressing the remaining issues by the end of the week :).
