+++ README.rst
@@ -29,13 +29,18 @@ like this one:
 Getting Started
 ---------------
 
-You can try **PyAutoLens** now by following the `introduction Jupyter Notebook on
+The following links are useful for new starters:
+
+- Try **PyAutoLens** in a web browser (without installation) by following the `introduction Jupyter Notebook on
 Binder <https://mybinder.org/v2/gh/Jammy2211/autolens_workspace/3b48dbc1b0ee85e68a24394895702df78e465323?filepath=introduction.ipynb>`_.
 
-On `readthedocs <https://pyautolens.readthedocs.io/>`_ you'll find the installation guide, a complete overview
-of **PyAutoLens**'s features, examples scripts, and
-the `HowToLens Jupyter notebook tutorials <https://pyautolens.readthedocs.io/en/latest/howtolens/howtolens.html>`_ which
-introduces new users to **PyAutoLens**.
+- The **PyAutoLens** `documentation on readthedocs <https://pyautolens.readthedocs.io/en/latest>`_, which includes
+the `PyAutoLens installation guide <https://pyautolens.readthedocs.io/en/latest/installation/overview.html>`_ and
+an overview of **PyAutoLens**'s core features.
+
+- The `autolens_workspace GitHub repository <https://github.com/Jammy2211/autolens_workspace>`_, which includes example
+scripts and the `HowToLens Jupyter notebook tutorials <https://github.com/Jammy2211/autolens_workspace/tree/master/notebooks/howtolens>`_
+which give new users a step-by-step introduction to **PyAutoLens**.
 
 API Overview
 ------------
@@ -55,22 +60,23 @@ lens ``Galaxy`` with an ``EllipticalIsothermal`` ``MassProfile`` lenses a backgr
     To describe the deflection of light by mass, two-dimensional grids of (y,x) Cartesian
     coordinates are used.
     """
-
     grid = al.Grid2D.uniform(
         shape_native=(50, 50),
         pixel_scales=0.05,  # <- Conversion from pixel units to arc-seconds.
     )
 
-    """The lens galaxy has an EllipticalIsothermal MassProfile and is at redshift 0.5."""
-
+    """
+    The lens galaxy has an EllipticalIsothermal MassProfile and is at redshift 0.5.
+    """
     mass = al.mp.EllipticalIsothermal(
         centre=(0.0, 0.0), elliptical_comps=(0.1, 0.05), einstein_radius=1.6
     )
 
     lens_galaxy = al.Galaxy(redshift=0.5, mass=mass)
 
-    """The source galaxy has an EllipticalExponential LightProfile and is at redshift 1.0."""
-
+    """
+    The source galaxy has an EllipticalExponential LightProfile and is at redshift 1.0.
+    """
     disk = al.lp.EllipticalExponential(
         centre=(0.3, 0.2),
         elliptical_comps=(0.05, 0.25),
@@ -84,7 +90,6 @@ lens ``Galaxy`` with an ``EllipticalIsothermal`` ``MassProfile`` lenses a backgr
     We create the strong lens using a Tracer, which uses the galaxies, their redshifts
     and an input cosmology to determine how light is deflected on its path to Earth.
     """
-
     tracer = al.Tracer.from_galaxies(
         galaxies=[lens_galaxy, source_galaxy], cosmology=cosmo.Planck15
     )
@@ -93,7 +98,6 @@ lens ``Galaxy`` with an ``EllipticalIsothermal`` ``MassProfile`` lenses a backgr
     We can use the Grid2D and Tracer to perform many lensing calculations, for example
     plotting the image of the lensed source.
     """
-
     tracer_plotter = aplt.TracerPlotter(tracer=tracer, grid=grid)
     tracer_plotter.figures(image=True)
 
@@ -107,8 +111,9 @@ with an ``EllipticalSersic``.
[…]
