+++ docs/requirements.txt
@@ -7,7 +7,7 @@ nbsphinx>=0.7.0
 nbsphinx-link>=1.3.0
 numpydoc>=1.0.0
 matplotlib>=3.0.3
-numpy>=1.17.2
+numpy>=1.18.0
 ipython>=7.0.1
 ipykernel>=5.1.0
 jupyter_client>=5.2.3
