+++ requirements.txt
@@ -5,5 +5,5 @@ autogalaxy==0.18.4
 astropy>=3.0.0
 numpydoc>=1.0.0
 matplotlib>=3.0.3
-numpy>=1.17.2
+numpy>=1.18.0
 pyprojroot==0.2.0
