### 2021-02-11 6fbb224
Prototype of new requirements layout
 - Have only one requirements file with 3rd party deps
 - 1st party deps are in setup.py
