+++ requirements_build.txt
@@ -1,6 +1,0 @@
-autoconf==0.7.7
-astropy>=3.0.0
-numpydoc>=1.0.0
-matplotlib>=3.0.3
-numpy>=1.17.2
-pyprojroot==0.2.0
