+++ citations.tex
@@ -55,7 +55,7 @@ This work uses the following software packages:
 \citep{pylops}
 
 \item
-\href{https://github.com/JohannesBuchner/PyMultiNest}{\textt{PyMulitNest}}
+\href{https://github.com/JohannesBuchner/PyMultiNest}{\textt{PyMultiNest}}
 \citep{multinest, pymultinest}
 
 \item
