+++ citations.md
@@ -19,7 +19,7 @@ This work uses the following software packages:
 - `PyAutoFit` https://github.com/rhayes777/PyAutoFit [@pyautofit]
 - `PyAutoLens` https://github.com/Jammy2211/PyAutoLens [@Nightingale2015] [@Nightingale2018] [@pyautolens]
 - `PyLops` https://github.com/equinor/pylops [@pylops]
-- `PyMulitNest` https://github.com/JohannesBuchner/PyMultiNest [@multinest] [@pymultinest]
+- `PyMultiNest` https://github.com/JohannesBuchner/PyMultiNest [@multinest] [@pymultinest]
 - `PyNUFFT` https://github.com/jyhmiinlin/pynufft [@pynufft]
 - `pyquad` https://github.com/AshKelly/pyquad [@pyquad]
 - `PySwarms` https://github.com/ljvmiranda921/pyswarms [@pyswarms]
