+++ paper/paper.bib
@@ -625,21 +625,19 @@ volume = {444},
 year = {2014}
 }
 @article{Powell2020,
-abstract = {We present a new gravitational lens modelling technique designed to model high-resolution interferometric observations with large numbers of visibilities without the need to pre-average the data in time or frequency. We demonstrate the accuracy of the method using validation tests on mock observations. Using small data sets with ∼103 visibilities, we first compare our approach with the more traditional direct Fourier transform (DFT) implementation and direct linear solver. Our tests indicate that our source inversion is indistinguishable from that of the DFT. Our method also infers lens parameters to within 1 to 2 per cent of both the ground truth and DFT, given sufficiently high signal-to-noise ratio (SNR). When the SNR is as low as 5, both approaches lead to errors of several tens of per cent in the lens parameters and a severely disrupted source structure, indicating that this is related to the SNR and choice of priors rather than the modelling technique itself. We then analyze a large data set with ∼108 visibilities and a SNR matching real global Very Long Baseline Interferometry observations of the gravitational lens system MG J0751+2716. The size of the data is such that it cannot be modelled with traditional implementations. Using our novel technique, we find that we can infer the lens parameters and the source brightness distribution, respectively, with an RMS error of 0.25 and 0.97 per cent relative to the ground truth.},
-archivePrefix = {arXiv},
-arxivId = {2005.03609},
-author = {Powell, Devon and Vegetti, Simona and McKean, John P and Spingola, Cristiana and Rizzo, Francesca and Stacey, Hannah R},
-doi = {10.1093/mnras/staa2740},
-eprint = {2005.03609},
-file = {:home/jammy/Documents/Papers/Strong{\_}Lens/Powell2020interfermeter.pdf:pdf},
-issn = {0035-8711},
-journal = {Monthly Notices of the Royal Astronomical Society},
-keywords = {data analysis,gravitational lensing,high angular resolution,image,methods,processing,strong,techniques},
-number = {October},
-pages = {1--16},
-title = {{A novel approach to visibility-space modelling of interferometric gravitational lens observations at high angular resolution}},
-volume = {16},
-year = {2020}
+    author = {Powell, Devon and Vegetti, Simona and McKean, John P and Spingola, Cristiana and Rizzo, Francesca and Stacey, Hannah R},
+    title = "{A novel approach to visibility-space modelling of interferometric gravitational lens observations at high angular resolution}",
+    journal = {Monthly Notices of the Royal Astronomical Society},
+    volume = {501},
+    number = {1},
+    pages = {515-530},
+    year = {2020},
+    month = {09},
[…]
