+++ paper/paper.md
@@ -15,6 +15,9 @@ authors:
   - name: Ashley Kelly
     orcid: 0000-0003-3850-4469
     affiliation: 1
+  - name: Aristeidis Amvrosiadis
+    orcid: 0000-0002-4465-1564
+    affiliation: 1    
   - name: Amy Etherington
     affiliation: 1
   - name: Qiuhan He
@@ -25,8 +28,8 @@ authors:
     affiliation: 2
   - name: XiaoYue Cao
     affiliation: 3
-  - name: Aristeidis Amvrosiadis
-    orcid: 0000-0002-4465-1564
+  - name: Jonathan Frawley
+    orcid: 0000-0002-9437-7399
     affiliation: 1
   - name: Shaun Cole
     orcid: 0000-0002-5954-7903
@@ -77,15 +80,14 @@ bibliography: paper.bib
 Strong gravitational lensing, which can make a background source galaxy appears multiple times due to its light rays being 
 deflected by the mass of one or more foreground lens galaxies, provides astronomers with a powerful tool to study dark 
 matter, cosmology and the most distant Universe. `PyAutoLens` is an open-source Python 3.6+ package for strong 
-gravitational  lensing, with core features including fully automated strong lens modeling of galaxies and galaxy 
+gravitational lensing, with core features including fully automated strong lens modeling of galaxies and galaxy 
 clusters, support for direct imaging and interferometer datasets and comprehensive tools for simulating samples of 
 strong lenses. The API allows users to perform ray-tracing by using analytic light and mass profiles to build strong 
-lens systems. Accompanying `PyAutoLens` is the `autolens workspace`, which 
-includes example scripts, lens datasets and 
-the [HowToLens](https://pyautolens.readthedocs.io/en/latest/howtolens/howtolens.html) lectures in Jupyter notebook format 
-which introduce non-experts to strong lensing using `PyAutoLens`. To get started readers should go to 
-our [readthedocs](https://pyautolens.readthedocs.io/en/latest/) and contact us to join 
-the [PyAutoLens Slack channel](https://pyautolens.slack.com/) where we are building our online community.
+lens systems. Accompanying `PyAutoLens` is the [autolens workspace](https://github.com/Jammy2211/autolens_workspace), which 
+includes example scripts, lens datasets and the [HowToLens](https://pyautolens.readthedocs.io/en/latest/howtolens/howtolens.html) 
+lectures in Jupyter notebook format which introduce non-experts to strong lensing using `PyAutoLens`. Readers can 
+try `PyAutoLens` right now by going to [the introduction Jupyter notebook on Binder](https://mybinder.org/v2/gh/Jammy2211/autolens_workspace/master) 
+or checkout the [readthedocs](https://pyautolens.readthedocs.io/en/latest/) for a complete overview of `PyAutoLens`'s features.
 
 # Background
 
@@ -97,8 +99,8 @@ below. The massive elliptical lens galaxy can be seen in the centre of the left 
 imaged source galaxy whose light has been distorted into an Einstein ring. The central panel shows a `PyAutoLens`
 reconstruction of the lensed source's light, where the foreground lens's light was simultaneously fitted for and 
 subtracted to reveal the source. The right panel shows a pixelized reconstruction of the source's unlensed light 
-distribution, which is created using a model of the lens galaxy's mass to trace backwards how the source's light is 
-gravitationally lensed.
+distribution performed by `PyAutoLens`, which is created using a model of the lens galaxy's mass to trace backwards 
+how the source's light is gravitationally lensed.
 
[…]
