### 2021-02-13 #2825 @Jammy2211
> I noticed that while there are a lot of tagged releases on PyPI, on Github there are only very old tags. Do you happen to have git tags available locally that you can push? otherwise, what about tagging on Github just the last 3/4 releases? It is convenient to be able to browse older versions of the code without downloading them from PyPI.

Looking back, we made a bit of a hash of our tagging. I pushed tags for some recent releases, but don't have a backlog to recover old versions.

We are now in the process of doing a major overhaul of versioning / builds, including an automated nightly release pipeline which will link tagging to GitHub commits (https://github.com/jonathanfrawley/PyAutoLens/tree/feature/release-process). This should begin to go live next week, and I think is the best I can do to address this issue.

You were right to bring it to my attention that this was a part of development I got into pretty bad habits with :).

> It is really easy to setup a Github action to run the test suite, otherwise you find yourself in the situation that your unit tests are failing on master, see for example: Jammy2211/PyAutoLens#146 I believe this is extremely important to increase reliability of the software.

Yep, another really valid point that become immediately apparent when we had someone new join the project a few weeks ago. The proper build pipelines we are writing will include automated unit tests on the master pipeline (we scrapped the development branch and went to a master-only model this week). Currently we've just got it set to send an email when unit tests on master fail, but we'll be turning that to a reject next week once the build pipelines are fully implemented.

>  binder?: what about setting up binder so people can go through the lessons without installing anything locally? https://mybinder.readthedocs.io/en/latest/

We've adopted binder pretty heavily throughout the project docs now - https://mybinder.org/v2/gh/Jammy2211/autofit_workspace/HEAD

Thanks again for this amazing suggestion :).

> ok, I think it is useful to specify in the performance section if your code is threaded in some sections (even just using the MKL version of numpy that automatically threads) or if it is fully single threaded.

> Also, in the performance section you can give a specific example like: "analyzing 1 million visibilities of 1 specific instrument on a recent laptop with 4 GB of ram takes ~3 hours".

I am trying to avoid being too specific about run times / optimization / implementation. We had a software engineer start on the project a few weeks ago who is tasked with optimizing the code-base, including multi-threading and algorithmic development.

I am going to add a run-time section to readthedocs which will give a live profile of the latest release of the code (via the automated build server) and a run-time calculator. I have mentioned this in the paper if people want more specifics, I will send a link to the page when it goes live (which should be in the next day or two).

I have also add a statement of need section, which was just a rejig of an already existing paragraph.

Let me know if there is anything else I need to address!
