### 2021-02-13 83a5aaf
removed binning util and feature
