+++ test_autolens/unit/pipeline/phase/test_settings.py
@@ -9,7 +9,6 @@ def test__tag__mixture_of_values():
             grid_inversion_class=al.Grid2D,
             sub_size=2,
             signal_to_noise_limit=2,
-            bin_up_factor=None,
             psf_shape_2d=None,
         ),
         settings_lens=al.SettingsLens(positions_threshold=2.0),
@@ -34,7 +33,6 @@ def test__tag__mixture_of_values():
             grid_inversion_class=al.Grid2DInterpolate,
             fractional_accuracy=0.5,
             pixel_scales_interp=0.3,
-            bin_up_factor=3,
             psf_shape_2d=(2, 2),
         ),
         settings_lens=al.SettingsLens(
@@ -48,13 +46,13 @@ def test__tag__mixture_of_values():
 
     assert (
         settings.phase_tag_no_inversion == "settings__"
-        "imaging[grid_facc_0.5__bin_3__psf_2x2]__"
+        "imaging[grid_facc_0.5__psf_2x2]__"
         "lens[pos_on]__"
         "lh_cap_200.0"
     )
     assert (
         settings.phase_tag_with_inversion == "settings__"
-        "imaging[grid_facc_0.5_inv_interp_0.300__bin_3__psf_2x2]__"
+        "imaging[grid_facc_0.5_inv_interp_0.300__psf_2x2]__"
         "lens[pos_on]__"
         "pix[no_border__stochastic]__"
         "inv[mat]__"
