+++ test_autolens/unit/config/notation/settings_tags.ini
@@ -18,7 +18,6 @@ signal_to_noise_limit=snr
 
 [imaging]
 imaging=imaging
-bin_up_factor=bin
 psf_shape=psf
 
 [interferometer]
