+++ test_autolens/unit/pipeline/phase/abstract/test_result_abstract.py
@@ -16,6 +16,7 @@ directory = os.path.dirname(os.path.realpath(__file__))
 
 
 class TestTracer:
+
     def test__max_log_likelihood_tracer_available_as_result(
         self, imaging_7x7, mask_7x7, samples_with_result
     ):
@@ -53,8 +54,6 @@ class TestTracer:
         result = phase_dataset_7x7.run(
             dataset=imaging_7x7, mask=mask_7x7, results=mock.MockResults()
         )
-
-        print(result.source_plane_light_profile_centre)
 
         assert result.source_plane_light_profile_centre.in_list == [(1.0, 2.0)]
 
