### 2021-02-14 cce858a
removed uncenecssary dunder methods from Array2D
