+++ test_autolens/unit/lens/test_ray_tracing.py
@@ -1923,6 +1923,7 @@ class TestAbstractTracerLensing:
             assert (grid_at_redshift == sub_grid_7x7.mask.unmasked_grid_sub_1).all()
 
     class TestContributionMap:
+
         def test__contribution_maps_are_same_as_hyper_galaxy_calculation(self):
 
             hyper_model_image = al.Array2D.manual_native(
@@ -2023,6 +2024,7 @@ class TestAbstractTracerLensing:
 
 
 class TestAbstractTracerData:
+
     class TestBlurredProfileImages:
         def test__blurred_image_from_grid_and_psf(
             self, sub_grid_7x7, blurring_grid_7x7, psf_3x3
@@ -2830,6 +2832,7 @@ class TestAbstractTracerData:
             # )
 
     class TestHyperNoiseMap:
+
         def test__hyper_noise_maps_of_planes(self, sub_grid_7x7):
 
             noise_map_1d = al.Array2D.manual_native(
@@ -2883,7 +2886,7 @@ class TestAbstractTracerData:
 
             assert (hyper_noise_maps[0].slim == hyper_noise_map_0).all()
             assert (hyper_noise_maps[1].slim == hyper_noise_map_1).all()
-            assert hyper_noise_maps[2].slim == np.zeros(shape=(3, 1))
+            assert (hyper_noise_maps[2].slim == np.zeros(shape=(3, 1))).all()
 
             hyper_noise_map = tracer.hyper_noise_map_from_noise_map(
                 noise_map=noise_map_1d
