+++ test_autolens/dataset/test_interferometer.py
@@ -1,0 +1,253 @@
+import autolens as al
+import numpy as np
+import pytest
+
+
+class TestMaskedInterferometer:
+    def test__masked_dataset_via_autoarray(
+        self,
+        interferometer_7,
+        sub_mask_7x7,
+        visibilities_mask_7,
+        visibilities_7,
+        visibilities_noise_map_7,
+    ):
+
+        masked_interferometer_7 = al.MaskedInterferometer(
+            interferometer=interferometer_7,
+            visibilities_mask=visibilities_mask_7,
+            real_space_mask=sub_mask_7x7,
+            settings=al.SettingsMaskedInterferometer(
+                transformer_class=al.TransformerDFT
+            ),
+        )
+
+        assert (
+            masked_interferometer_7.visibilities == interferometer_7.visibilities
+        ).all()
+        assert (masked_interferometer_7.visibilities == visibilities_7).all()
+
+        assert (masked_interferometer_7.noise_map == visibilities_noise_map_7).all()
+
+        assert (
+            masked_interferometer_7.visibilities_mask
+            == np.full(fill_value=False, shape=(7,))
+        ).all()
+
+        assert (
+            masked_interferometer_7.interferometer.uv_wavelengths
+            == interferometer_7.uv_wavelengths
+        ).all()
+        assert (
+            masked_interferometer_7.interferometer.uv_wavelengths[0, 0]
+            == -55636.4609375
+        )
+
+        assert type(masked_interferometer_7.transformer) == al.TransformerDFT
+
+    def test__inheritance_via_autoarray(
+        self,
+        interferometer_7,
+        sub_mask_7x7,
+        visibilities_mask_7,
+        grid_7x7,
+        sub_grid_7x7,
+    ):
+
+        masked_interferometer_7 = al.MaskedInterferometer(
+            interferometer=interferometer_7,
+            visibilities_mask=visibilities_mask_7,
+            real_space_mask=sub_mask_7x7,
+            settings=al.SettingsMaskedInterferometer(grid_class=al.Grid2D),
+        )
+
+        assert (masked_interferometer_7.grid.slim_binned == grid_7x7).all()
+        assert (masked_interferometer_7.grid == sub_grid_7x7).all()
+
+        grid = al.Grid2D.from_mask(mask=sub_mask_7x7)
+
+        assert (masked_interferometer_7.grid == grid).all()
+
+    def test__different_interferometer_without_mock_objects__customize_constructor_inputs(
+        self,
+    ):
+        interferometer = al.Interferometer(
+            visibilities=al.Visibilities.ones(shape_slim=(19,)),
+            noise_map=al.Visibilities.full(fill_value=2.0, shape_slim=(19,)),
+            uv_wavelengths=3.0 * np.ones((19, 2)),
+        )
+
+        visibilities_mask = np.full(fill_value=False, shape=(19,))
+
+        real_space_mask = al.Mask2D.unmasked(
+            shape_native=(19, 19), pixel_scales=1.0, invert=True, sub_size=8
+        )
+        real_space_mask[9, 9] = False
+
+        masked_interferometer = al.MaskedInterferometer(
+            interferometer=interferometer,
+            visibilities_mask=visibilities_mask,
+            real_space_mask=real_space_mask,
+        )
+
+        assert (
+            masked_interferometer.visibilities.slim == (1.0 + 1.0j) * np.ones((19,))
+        ).all()
+        assert (
+            masked_interferometer.noise_map.slim == (2.0 + 2.0j) * np.ones((19,))
+        ).all()
+        assert (
+            masked_interferometer.interferometer.uv_wavelengths
[…]
