+++ autolens/dataset/imaging.py
@@ -1,8 +1,6 @@
-import copy
-
 from autoarray.dataset import imaging
-from autoarray.structures import grids
-from autoarray.structures import kernel
+from autoarray.structures.grids.two_d import grid_2d
+from autoarray.structures import kernel_2d
 from autogalaxy.dataset import imaging as im
 from autolens.lens import ray_tracing
 
@@ -49,7 +47,7 @@ class SimulatorImaging(imaging.SimulatorImaging):
         self,
         exposure_time: float,
         background_sky_level: float = 0.0,
-        psf: kernel.Kernel2D = None,
+        psf: kernel_2d.Kernel2D = None,
         renormalize_psf: bool = True,
         read_noise: float = None,
         add_poisson_noise: bool = True,
@@ -148,7 +146,7 @@ class SimulatorImaging(imaging.SimulatorImaging):
 
     def from_deflections_and_galaxies(self, deflections, galaxies, name=None):
 
-        grid = grids.Grid2D.uniform(
+        grid = grid_2d.Grid2D.uniform(
             shape_native=deflections.shape_native,
             pixel_scales=deflections.pixel_scales,
             sub_size=1,
