+++ autolens/dataset/point_source.py
@@ -1,8 +1,6 @@
-from autoarray.structures import arrays, grids
-
 from typing import Optional
 
 #
 # class PointSourceDataset:
 #
-#     def __init__(self, name, positions : grids.Grid2DIrregular = None, fluxes : arrays.ValuesIrregular):
+#     def __init__(self, name, positions : grid_2d_irregular.Grid2DIrregular = None, fluxes : values.ValuesIrregular):
