+++ autolens/plot/plotters/fit_interferometer_plotters.py
@@ -1,8 +1,6 @@
-from autoarray.plot.mat_wrap import mat_plot as mp
 from autoarray.plot.plotters import fit_interferometer_plotters
 from autoarray.plot.plotters import inversion_plotters
 from autogalaxy.plot.mat_wrap import lensing_mat_plot, lensing_include, lensing_visuals
-from autogalaxy.plot.plotters import plane_plotters
 from autolens.plot.plotters import ray_tracing_plotters
 from autolens.fit import fit as f
 
