+++ autolens/lens/positions_solver.py
@@ -1,6 +1,6 @@
 from autoarray import decorator_util
 import numpy as np
-from autoarray.structures import grids
+from autoarray.structures.grids.two_d import grid_2d_irregular
 from autogalaxy.profiles import mass_profiles as mp
 
 
@@ -34,7 +34,7 @@ class AbstractPositionsSolver:
             if magnification > self.magnification_threshold:
                 grid_mag.append(grid[index, :])
 
-        return grids.Grid2DIrregularUniform(
+        return grid_2d_irregular.Grid2DIrregularUniform(
             grid=grid_mag, pixel_scales=grid.pixel_scales
         )
 
@@ -77,7 +77,9 @@ class AbstractPositionsSolver:
                     outside_distance=self.distance_from_mass_profile_centre,
                 )
 
-            return grids.Grid2DIrregularUniform(grid=grid, pixel_scales=pixel_scales)
+            return grid_2d_irregular.Grid2DIrregularUniform(
+                grid=grid, pixel_scales=pixel_scales
+            )
 
         return grid
 
@@ -127,7 +129,7 @@ class AbstractPositionsSolver:
             upscale_factor=upscale_factor,
         )
 
-        return grids.Grid2DIrregularUniform(
+        return grid_2d_irregular.Grid2DIrregularUniform(
             grid=grid_buffed,
             pixel_scales=(
                 pixel_scales[0] / upscale_factor,
@@ -178,7 +180,7 @@ class AbstractPositionsSolver:
             has_neighbors=has_neighbors,
         )
 
-        return grids.Grid2DIrregularUniform(
+        return grid_2d_irregular.Grid2DIrregularUniform(
             grid=grid_peaks, pixel_scales=grid.pixel_scales
         )
 
@@ -227,7 +229,7 @@ class AbstractPositionsSolver:
             within_distance=distance,
         )
 
-        return grids.Grid2DIrregularUniform(
+        return grid_2d_irregular.Grid2DIrregularUniform(
             grid=grid_within_distance_of_centre, pixel_scales=grid.pixel_scales
         )
 
@@ -338,7 +340,7 @@ class PositionsSolver(AbstractPositionsSolver):
 
         if not self.use_upscaling:
 
-            return grids.Grid2DIrregular(grid=coordinates_list)
+            return grid_2d_irregular.Grid2DIrregular(grid=coordinates_list)
 
         pixel_scale = self.grid.pixel_scale
 
@@ -368,7 +370,7 @@ class PositionsSolver(AbstractPositionsSolver):
 
         coordinates_list = self.grid_within_distance_of_source_plane_centre(
             lensing_obj=lensing_obj,
-            grid=grids.Grid2DIrregularUniform(
+            grid=grid_2d_irregular.Grid2DIrregularUniform(
                 grid=coordinates_list, pixel_scales=(pixel_scale, pixel_scale)
             ),
             source_plane_coordinate=source_plane_coordinate,
@@ -379,7 +381,7 @@ class PositionsSolver(AbstractPositionsSolver):
             lensing_obj=lensing_obj, grid=coordinates_list
         )
 
-        return grids.Grid2DIrregular(grid=coordinates_list)
+        return grid_2d_irregular.Grid2DIrregular(grid=coordinates_list)
 
 
 @decorator_util.jit()
