+++ autolens/text/__init__.py
@@ -1,1 +1,0 @@
-from autofit.text import samples_text as Samples

[…]
