### 2021-02-14 7dcbcf2
setup.py details added
