+++ files/citations.md
@@ -1,0 +1,29 @@
+**Insert in the main body of the paper:**
+
+We use the lens modeling software `PyAutoLens` https://github.com/Jammy2211/PyAutoLens) [@pyautolens] [@Nightingale2015] [@Nightingale2018] to...
+
+**At the end of the paper (delete as appropriate, see https://pyautofit.readthedocs.io/en/latest/general/citations.html):**
+
+# Software Citations
+
+This work uses the following software packages:
+
+- `Astropy` https://github.com/astropy/astropy [@astropy1] [@astropy2]
+- `Colossus` https://bitbucket.org/bdiemer/colossus/src/master/ [@colossus]
+- `corner.py` https://github.com/dfm/corner.py [@corner]
+- `dynesty` https://github.com/joshspeagle/dynesty [@dynesty]
+- `emcee` https://github.com/dfm/emcee [@emcee]
+- `matplotlib` https://github.com/matplotlib/matplotlib [@matplotlib]
+- `numba` https://github.com/numba/numba [@numba]
+- `NumPy` https://github.com/numpy/numpy [@numpy]
+- `PyAutoFit` https://github.com/rhayes777/PyAutoFit [@pyautofit]
+- `PyAutoLens` https://github.com/Jammy2211/PyAutoLens [@Nightingale2015] [@Nightingale2018] [@pyautolens]
+- `PyLops` https://github.com/equinor/pylops [@pylops]
+- `PyMultiNest` https://github.com/JohannesBuchner/PyMultiNest [@multinest] [@pymultinest]
+- `PyNUFFT` https://github.com/jyhmiinlin/pynufft [@pynufft]
+- `pyquad` https://github.com/AshKelly/pyquad [@pyquad]
+- `PySwarms` https://github.com/ljvmiranda921/pyswarms [@pyswarms]
+- `Python` https://www.python.org/ [@python]
+- `scikit-image` https://github.com/scikit-image/scikit-image [@scikit-image]
+- `scikit-learn` https://github.com/scikit-learn/scikit-learn [@scikit-learn]
+- `Scipy` https://github.com/scipy/scipy [@scipy]
