+++ files/citations.tex
@@ -1,0 +1,89 @@
+**Insert in the main body of the paper:**
+
+We use the lens modeling software \textt{PyAutoLens} \footnote{https://github.com/Jammy2211/PyAutoLens} \citep{pyautolens, Nightingale2015, Nightingale2018} to...
+
+**At the end of the paper (delete as appropriate, see https://pyautolens.readthedocs.io/en/latest/general/citations.html):**
+
+\section*{Software Citations}
+
+This work uses the following software packages:
+
+\begin{itemize}
+
+\item
+\href{https://github.com/astropy/astropy}{\textt{Astropy}}
+\citep{astropy1, astropy2}
+
+\item
+\href{https://bitbucket.org/bdiemer/colossus/src/master/}{\textt{Colossus}}
+\citep{colossus}
+
+\item
+\href{https://github.com/dfm/corner.py}{\textt{corner.py}}
+\citep{corner}
+
+\item
+\href{https://github.com/joshspeagle/dynesty}{\textt{dynesty}}
+\citep{dynesty}
+
+\item
+\href{https://github.com/dfm/emcee}{\textt{emcee}}
+\citep{emcee}
+
+\item
+\href{https://github.com/matplotlib/matplotlib}{\textt{matplotlib}}
+\citep{matplotlib}
+
+\item
+\href{numba` https://github.com/numba/numba}{\textt{numba}}
+\citep{numba}
+
+\item
+\href{https://github.com/numpy/numpy}{\textt{NumPy}}
+\citep{numpy}
+
+\item
+\href{https://github.com/rhayes777/PyAutoFit}{\textt{PyAutoFit}}
+\citep{pyautofit}
+
+\item
+\href{https://github.com/Jammy2211/PyAutoLens}{\textt{PyAutoLens}}
+\citep{Nightingale2015, Nightingale2018, pyautolens}
+
+\item
+\href{https://github.com/equinor/pylops}{\textt{PyLops}}
+\citep{pylops}
+
+\item
+\href{https://github.com/JohannesBuchner/PyMultiNest}{\textt{PyMultiNest}}
+\citep{multinest, pymultinest}
+
+\item
+\href{https://github.com/jyhmiinlin/pynufft}{\textt{PyNUFFT}}
+\citep{pynufft}
+
+\item
+\href{https://github.com/AshKelly/pyquad}{\textt{pyquad}}
+\citep{pyquad}
+
+\item
+\href{https://github.com/ljvmiranda921/pyswarms}{\textt{PySwarms}}
+\citep{pyswarms}
+
+\item
+\href{https://www.python.org/}{\textt{Python}}
+\citep{python}
+
+\item
+\href{https://github.com/scikit-image/scikit-image}{\textt{scikit-image}}
+\citep{scikit-image}
+
+\item
+\href{https://github.com/scikit-learn/scikit-learn}{\textt{scikit-learn}}
+\citep{scikit-learn}
+
+\item
+\href{https://github.com/scipy/scipy}{\textt{Scipy}}
+\citep{scipy}
+
+\end{itemize}
