+++ test_autolens/fit/test_fit.py
@@ -58,7 +58,6 @@ class TestFitImaging:
             assert fit.total_inversions == 3
 
     class TestLikelihood:
-
         def test__1x2_image__no_psf_blurring__tracing_fits_data_with_chi_sq_5(self):
             # The image plane image generated by the galaxy is [1.0, 1.0]
 
@@ -155,7 +154,9 @@ class TestFitImaging:
 
             assert fit.chi_squared == 25.0
             assert fit.reduced_chi_squared == 25.0 / 2.0
-            assert fit.noise_normalization == pytest.approx(2.0 * np.log(2 * np.pi * 1.0 ** 2.0), 1.0e-4)
+            assert fit.noise_normalization == pytest.approx(
+                2.0 * np.log(2 * np.pi * 1.0 ** 2.0), 1.0e-4
+            )
             assert fit.log_likelihood == -0.5 * (
                 25.0 + 2.0 * np.log(2 * np.pi * 1.0 ** 2.0)
             )
@@ -259,7 +260,9 @@ class TestFitImaging:
 
             assert fit.chi_squared == 16.0
             assert fit.reduced_chi_squared == 16.0 / 2.0
-            assert fit.noise_normalization == pytest.approx(2.0 * np.log(2 * np.pi * 1.0 ** 2.0), 1.0e-4)
+            assert fit.noise_normalization == pytest.approx(
+                2.0 * np.log(2 * np.pi * 1.0 ** 2.0), 1.0e-4
+            )
             assert fit.log_likelihood == -0.5 * (
                 16.0 + 2.0 * np.log(2 * np.pi * 1.0 ** 2.0)
             )
@@ -378,7 +381,9 @@ class TestFitImaging:
 
             assert fit.chi_squared == 4.0
             assert fit.reduced_chi_squared == 4.0 / 2.0
-            assert fit.noise_normalization == pytest.approx(2.0 * np.log(2 * np.pi * 2.0 ** 2.0), 1.0e-4)
+            assert fit.noise_normalization == pytest.approx(
+                2.0 * np.log(2 * np.pi * 2.0 ** 2.0), 1.0e-4
+            )
             assert fit.log_likelihood == -0.5 * (
                 4.0 + 2.0 * np.log(2 * np.pi * 2.0 ** 2.0)
             )
@@ -484,7 +489,9 @@ class TestFitImaging:
 
             assert fit.chi_squared == 41.0
             assert fit.reduced_chi_squared == 41.0 / 2.0
-            assert fit.noise_normalization == pytest.approx(2.0 * np.log(2 * np.pi * 1.0 ** 2.0), 1.0e-4)
+            assert fit.noise_normalization == pytest.approx(
+                2.0 * np.log(2 * np.pi * 1.0 ** 2.0), 1.0e-4
+            )
             assert fit.log_likelihood == -0.5 * (
                 41.0 + 2.0 * np.log(2 * np.pi * 1.0 ** 2.0)
             )
@@ -592,7 +599,9 @@ class TestFitImaging:
 
             assert fit.chi_squared == 6.25
             assert fit.reduced_chi_squared == 6.25 / 2.0
-            assert fit.noise_normalization ==pytest.approx(2.0 * np.log(2 * np.pi * 2.0 ** 2.0), 1.0e-4)
+            assert fit.noise_normalization == pytest.approx(
+                2.0 * np.log(2 * np.pi * 2.0 ** 2.0), 1.0e-4
+            )
             assert fit.log_likelihood == -0.5 * (
                 6.25 + 2.0 * np.log(2 * np.pi * 2.0 ** 2.0)
             )
