+++ test_autolens/config/notation/prior_model_tags.ini
@@ -46,6 +46,8 @@ EllipticalSersicRadialGradient=sersic_grad
 SphericalSersicRadialGradient=sersic_grad_sph
 EllipticalExponentialRadialGradient=exp_grad
 SphericalExponentialRadialGradient=exp_grad_sph
+EllipticalCoreSersic=core_sersic
+SphericalCoreSersic=core_sersic_sph
 EllipticalChameleon=cham
 SphericalChameleon=cham_sph
 
