+++ autolens/config/notation/setup_tags.ini
@@ -14,6 +14,6 @@ mass_is_model=mass_is_model
 mass_is_instance=mass_is_instance
 source_is_model=source_is_model
 source_is_instance=source_is_instance
-grid_size=grid
+number_of_steps=grid
 subhalo_centre=centre
 mass_at_200=mass
