### 2021-02-17 #2825 @arfon
@xuanxu - this PR should fix the paper compilation. Why this works is 🤯 😡  🤷 https://github.com/Jammy2211/PyAutoLens/pull/156
