### 2021-02-17 #2825 @xuanxu
@Jammy2211, everything looks good, you can take a final [proof](https://github.com/Jammy2211/PyAutoLens/pull/157) read at the [pdf draft](https://github.com/openjournals/joss-papers/blob/joss.02825/joss.02825/10.21105.joss.02825.pdf) and if you find it OK, here are the next steps:

- Please release a new tagged version from the current master so it includes all the changes made during the review process
- Then archive that latest release in Zenodo (or a similar archival service)
- Check the Zenodo deposit has the correct metadata: title and author names should match the paper; you may also add their ORCID info.

Once you do that please report here the version number and archive DOI
