+++ test_autolens/config/notation/prior_model_tags.ini
@@ -34,6 +34,7 @@ SphericalGeneralizedNFW=gnfw_sph
 SphericalNFWMCRDuffy=nfw_sph_duffy
 EllipticalNFWMCRLudlow=nfw_ludlow
 SphericalNFWMCRLudlow=nfw_sph_ludlow
+EllipticalGeneralizedNFWMCRLudlow=gnfw_ludlow
 EllipticalSersic=sersic
 SphericalSersic=sersic_sph
 EllipticalDevVaucouleurs=dev
