+++ paper/paper.md
@@ -30,19 +30,19 @@ authors:
     affiliation: 3
   - name: Jonathan Frawley
     orcid: 0000-0002-9437-7399
-    affiliation: 1
+    affiliation: 4
   - name: Shaun Cole
     orcid: 0000-0002-5954-7903
     affiliation: 1
   - name: Andrea Enia
     orcid: 0000-0002-0200-2857
-    affiliation: 4
+    affiliation: 5
   - name: Carlos S. Frenk
     orcid: 0000-0002-2338-716X
     affiliation: 1
   - name: David R. Harvey
     orcid: 0000-0002-6066-6707
-    affiliation: 5 
+    affiliation: 6
   - name: Ran Li
     orcid: 0000-0003-3899-0612
     affiliation: 3
@@ -51,7 +51,7 @@ authors:
     affiliation: 1
   - name: Mattia Negrello
     orcid: 0000-0002-7925-7663
-    affiliation: 6
+    affiliation: 7
   - name: Andrew Robertson
     orcid: 0000-0002-0086-0524
     affiliation: 1
@@ -62,12 +62,14 @@ affiliations:
     index: 2
   - name: National Astronomical Observatories, Chinese Academy of Sciences, 20A Datun Road, Chaoyang District, Beijing 100012, China
     index: 3
-  - name: Dipartimento di Fisica e Astronomia, Università degli Studi di Bologna, Via Berti Pichat 6/2, I-40127 Bologna, Italy
+  - name: Advanced Research Computing, Durham University, Durham DH1 3LE
     index: 4
-  - name: Lorentz Institute, Leiden University, Niels Bohrweg 2, Leiden, NL-2333 CA, The Netherlands
+  - name: Dipartimento di Fisica e Astronomia, Università degli Studi di Bologna, Via Berti Pichat 6/2, I-40127 Bologna, Italy
     index: 5
-  - name: School of Physics and Astronomy, Cardiff University, The Parade, Cardiff CF24 3AA, UK
+  - name: Lorentz Institute, Leiden University, Niels Bohrweg 2, Leiden, NL-2333 CA, The Netherlands
     index: 6
+  - name: School of Physics and Astronomy, Cardiff University, The Parade, Cardiff CF24 3AA, UK
+    index: 7
 
 date: 26 October 2020
 codeRepository: https://github.com/Jammy2211/PyAutoLens
@@ -84,7 +86,7 @@ gravitational lensing, with core features including fully automated strong lens 
 clusters, support for direct imaging and interferometer datasets and comprehensive tools for simulating samples of 
 strong lenses. The API allows users to perform ray-tracing by using analytic light and mass profiles to build strong 
 lens systems. Accompanying `PyAutoLens` is the [autolens workspace](https://github.com/Jammy2211/autolens_workspace), which 
-includes example scripts, lens datasets and the [HowToLens](https://pyautolens.readthedocs.io/en/latest/howtolens/howtolens.html) 
+includes example scripts, lens datasets and the `HowToLens`
 lectures in Jupyter notebook format which introduce non-experts to strong lensing using `PyAutoLens`. Readers can 
 try `PyAutoLens` right now by going to [the introduction Jupyter notebook on Binder](https://mybinder.org/v2/gh/Jammy2211/autolens_workspace/master) 
 or checkout the [readthedocs](https://pyautolens.readthedocs.io/en/latest/) for a complete overview of `PyAutoLens`'s features.
@@ -137,12 +139,11 @@ optimized using the packages `NumPy` [@numpy], `numba` [@numba] and `pyquad` [@p
 To perform lens modeling, `PyAutoLens` adopts the probabilistic programming 
[…]
