+++ paper/paper.bib
@@ -143,13 +143,17 @@ year = {2015}
   pages={22-30},
   doi={10.1109/MCSE.2011.37}}
 
-@misc{pyautofit,
-  author = {​James Nightingale and Richard Hayes},
-  title = {PyAutoFit: A classy Probabilistic Programming Language for Complex Model-fitting},
-  year = {2020},
-  publisher = {​GitHub},
-  journal = {​GitHub repository},
-  url = {https://github.com/rhayes777/PyAutoFit}
+@article{pyautofit,
+  doi = {10.21105/joss.02550},
+  url = {https://doi.org/10.21105/joss.02550},
+  year = {2021},
+  publisher = {The Open Journal},
+  volume = {6},
+  number = {58},
+  pages = {2550},
+  author = {James. W. Nightingale and Richard G. Hayes and Matthew Griffiths},
+  title = {`PyAutoFit`: A Classy Probabilistic Programming Language for Model Composition and Fitting},
+  journal = {Journal of Open Source Software}
 }
 @article{multinest,
 abstract = {We present further development and the first public release of our multimodal nested sampling algorithm, called MultiNest. This Bayesian inference tool calculates the evidence, with an associated error estimate, and produces posterior samples from distributions that may contain multiple modes and pronounced (curving) degeneracies in high dimensions. The developments presented here lead to further substantial improvements in sampling efficiency and robustness, as compared to the original algorithm presented in Feroz and Hobson, which itself significantly outperformed existing Markov chain Monte Carlo techniques in a wide range of astrophysical inference problems. The accuracy and economy of the MultiNest algorithm are demonstrated by application to two toy problems and to a cosmological inference problem focusing on the extension of the vanilla $\Lambda$ cold dark matter model to include spatial curvature and a varying equation of state for dark energy. The MultiNest software, which is fully parallelized using MPI and includes an interface to CosmoMC, is available at http://www.mrao.cam.ac.uk/software/multinest/. It will also be released as part of the SuperBayeS package, for the analysis of supersymmetric theories of particle physics, at http://www.superbayes.org. {\textcopyright} 2009 RAS.},
