+++ files/citations.bib
@@ -150,7 +150,7 @@ year = {2015}
   volume = {6},
   number = {58},
   pages = {2550},
-  author = {James. W. Nightingale and Richard G. Hayes and Matthew Griffiths},
+  author = {Nightingale, J. W. and Hayes, R. G. and Griffiths, M.},
   title = {`PyAutoFit`: A Classy Probabilistic Programming Language for Model Composition and Fitting},
   journal = {Journal of Open Source Software}
 }
