+++ autolens/__init__.py
@@ -89,4 +89,4 @@ from autoconf import conf
 
 conf.instance.register(__file__)
 
-__version__ = "1.12.1"
+__version__ = "1.12.2"
