+++ test_autolens/test_config.py
@@ -32,6 +32,12 @@ class TestLabel:
         assert label_config["label"]["gamma"] == r"\gamma"
         assert label_config["label"]["contribution_factor"] == r"\omega0"
 
+    def test_subscript(self, label_config):
+        assert label_config["subscript"].family(al.lp.EllipticalLightProfile) == "l"
+
+    def test_inheritance(self, label_config):
+        assert label_config["subscript"].family(al.lp.EllipticalGaussian) == "l"
+
     def test_exception(self, label_config):
         with pytest.raises(KeyError):
             label_config["subscript"].family(MockClass)
