### 2021-02-18 #2825 @Jammy2211
Sorry for the noise, had some citation issues I couldn't reproduce locally.

The zenodo archieved release is as follows, using v1.13.0 of PyAutoLens:

https://doi.org/10.5281/zenodo.4548744
10.5281/zenodo.4548744

The metadata has been updated to match the paper and author list!
