+++ test_autolens/test_config.py
@@ -16,8 +16,11 @@ class MockClass:
 @pytest.fixture(name="label_config")
 def make_label_config():
 
+    print(directory, "config")
+
     config = conf.Config(
-        path.join(directory, "config"), path.join(directory, "default")
+        path.join(directory, "config"),
+        output_path=path.join(directory, "output")
     )
 
     return config["notation"]["label"]
@@ -32,11 +35,11 @@ class TestLabel:
         assert label_config["label"]["gamma"] == r"\gamma"
         assert label_config["label"]["contribution_factor"] == r"\omega0"
 
-    def test_subscript(self, label_config):
-        assert label_config["subscript"].family(al.lp.EllipticalLightProfile) == "l"
-
-    def test_inheritance(self, label_config):
-        assert label_config["subscript"].family(al.lp.EllipticalGaussian) == "l"
+    # def test_subscript(self, label_config):
+    #     assert label_config["subscript"].family(al.lp.EllipticalLightProfile) == "l"
+    #
+    # def test_inheritance(self, label_config):
+    #     assert label_config["subscript"].family(al.lp.EllipticalGaussian) == "l"
 
     def test_exception(self, label_config):
         with pytest.raises(KeyError):
