+++ test_autolens/config/notation/prior_model_tags.ini
@@ -5,8 +5,8 @@ EllipticalDevVaucouleurs=dev
 SphericalDevVaucouleurs=dev_sph
 EllipticalExponential=exp
 SphericalExponential=exp_sph
-EllipticalGassian=gauss
-SphericalGassian=gauss_sph
+EllipticalGaussian=gauss
+SphericalGaussian=gauss_sph
 EllipticalCoreSersic=core_sersic
 SphericalCoreSersic=core_sersic_sph
 EllipticalChameleon=cham
@@ -41,8 +41,8 @@ EllipticalDevVaucouleurs=dev
 SphericalDevVaucouleurs=dev_sph
 EllipticalExponential=exp
 SphericalExponential=exp_sph
-EllipticalGassian=gauss
-SphericalGassian=gauss_sph
+EllipticalGaussian=gauss
+SphericalGaussian=gauss_sph
 EllipticalSersicRadialGradient=sersic_grad
 SphericalSersicRadialGradient=sersic_grad_sph
 EllipticalExponentialRadialGradient=exp_grad
