+++ test_autolens/config/notation/label.ini
@@ -45,7 +45,6 @@ weight_power=Wp
 [subscript]
 EllipticalLightProfile=l
 EllipticalMassProfile=m
-EllipticalGaussian=g
 EllipticalNFW=d
 ExternalShear=sh
 HyperBackgroundNoise=hnb
