### 2021-02-18 b7f224a
fixed fit tests for build
