+++ setup.py
@@ -13,8 +13,8 @@ with open(join(this_dir, "requirements.txt")) as f:
 
 requirements.extend([
     'autofit==0.73.1',
-    'autoarray==0.18.3',
-    'autogalaxy==0.18.4'
+    'autoarray==0.19.0',
+    'autogalaxy==0.18.5'
 ])
 
 setup(
