+++ requirements.txt
@@ -1,4 +1,3 @@
-autoconf==0.7.7
 astropy>=3.0.0
 numpydoc>=1.0.0
 matplotlib>=3.0.3
