### 2021-02-20 #2825 @arfon
@aureliocarnero, @zonca - many thanks for your reviews here and to @xuanxu for editing this submission.

JOSS relies upon the volunteer efforts of folks like yourselves - we simply wouldn't be able to do this without you! ✨
